Nested particle filter state

Nested particle filter internal state. This object is not ordinarily constructed directly by users, but via the $run_begin method to particle_filter. It provides an advanced interface to the particle filter that allows partially running the particle filter over part of the time trajectory.

This state object has a number of public fields that you can read but must not write (they are not read-only so you could write them, but don't).

Public fields

model

The dust model generator being simulated (cannot be re-bound)

history

The particle history, if created with save_history = TRUE. This is an internal format subject to

restart_state

Full model state at a series of points in time, if the model was created with non-NULL save_restart. This is a 4d array as described in particle_filter

log_likelihood

The log-likelihood so far. This starts at 0 when initialised and accumulates value for each step taken.

log_likelihood_step

The log-likelihood attributable to the last step (i.e., the contribution to log_likelihood made on the last call to $step().

Methods

Public methods

particle_filter_state_nested$new()

particle_filter_state_nested$run()

particle_filter_state_nested$step()

particle_filter_state_nested$fork()

Method new()

Initialise the particle filter state. Ordinarily this should not be called by users, and so arguments are not documented.

Usage

particle_filter_state_nested$new(
  pars,
  generator,
  model,
  data,
  data_split,
  steps,
  n_particles,
  n_threads,
  initial,
  index,
  compare,
  seed,
  save_history,
  save_restart
)

Method run()

Run the particle filter to the end of the data. This is a convenience function around $step() which provides the correct value of step_index

Usage

particle_filter_state_nested$run()

Method step()

Take a step with the particle filter. This moves the particle filter forward one step within the data (which may correspond to more than one step with your model) and returns the likelihood so far.

Usage

particle_filter_state_nested$step(step_index)

Arguments

step_index

The step index to move to. This is not the same as the model step, nor time, so be careful (it's the index within the data provided to the filter). It is an error to provide a value here that is lower than the current step index, or past the end of the data.

Method fork()

Create a new particle_filter_state_nested object based on this one (same model, position in time within the data) but with new parameters. To do this, we create a new particle_filter_state_nested with new parameters at the beginning of the simulation (corresponding to the start of your data or the initial argument to particle_filter) with your new pars, and then run the filter foward in time until it reaches the same step as the parent model.

Usage

particle_filter_state_nested$fork(pars)

Arguments

pars

New model parameters