Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4, 0)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 12 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.8397606 0.2622543
#>  [2,] 0.7990637 0.2683571
#>  [3,] 0.8397606 0.2622543
#>  [4,] 0.8305435 0.3009885
#>  [5,] 0.8880000 0.3181306
#>  [6,] 0.9065794 0.3150475
#>  [7,] 0.9100279 0.3379586
#>  [8,] 0.8233554 0.2572554
#>  [9,] 0.8233554 0.2572554
#> [10,] 0.8302572 0.2514221
#> [11,] 0.8157140 0.2686066
#> [12,] 0.8157140 0.2686066
#> [13,] 0.8157140 0.2686066
#> [14,] 0.8157140 0.2686066
#> [15,] 0.7981114 0.2763425
#> [16,] 0.7874227 0.2661152
#> [17,] 0.8039881 0.2588960
#> [18,] 0.8039881 0.2588960
#> [19,] 0.8039881 0.2588960
#> [20,] 0.7099792 0.2511521
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior        weight
#>  [1,]         0     -1020.2596    -1020.2596 2.097485e-268
#>  [2,]         0      -982.7040     -982.7040 4.284325e-252
#>  [3,]         0     -1009.5438    -1009.5438 9.451857e-264
#>  [4,]         0      -417.9758     -417.9758  7.766946e-07
#>  [5,]         0      -751.1889     -751.1889 1.505336e-151
#>  [6,]         0      -545.8461     -545.8461  2.274336e-62
#>  [7,]         0      -502.4815     -502.4815  1.548344e-43
#>  [8,]         0      -960.5132     -960.5132 1.858813e-242
#>  [9,]         0      -950.8877     -950.8877 2.815311e-238
#> [10,]         0      -947.8806     -947.8806 5.694976e-237
#> [11,]         0      -664.7460     -664.7460 5.239762e-114
#> [12,]         0      -665.5046     -665.5046 2.453883e-114
#> [13,]         0      -665.5046     -665.5046 2.453877e-114
#> [14,]         0      -602.7105     -602.7105  4.581124e-87
#> [15,]         0      -601.9387     -601.9387  9.912477e-87
#> [16,]         0      -472.7481     -472.7481  1.267413e-30
#> [17,]         0      -916.4214     -916.4214 2.618663e-223
#> [18,]         0      -927.5368     -927.5368 3.896622e-228
#> [19,]         0      -916.2672     -916.2672 3.055100e-223
#> [20,]         0      -403.9075     -403.9075  9.999992e-01
#> 
#> $statistics
#> $statistics$ess
#>   [1] 13.679560  6.856071 11.718031  9.273286 12.243224 11.188323  8.641640
#>   [8] 10.802114  8.186009 11.240647  9.871584  2.727233  5.145093  5.823557
#>  [15]  3.637038  2.408667  1.954163  1.827817  1.786348  3.283282  2.454184
#>  [22]  6.247533  5.067847  5.832295  6.180112  6.073261  4.025133  2.301444
#>  [29]  1.717569  2.559120  5.903281  8.093628  8.380567  4.768013 12.466129
#>  [36] 10.356718  8.712411 13.560651 10.859220  9.512575 14.437958 11.141152
#>  [43] 10.599029 10.069741  9.334780 16.309359 11.917471 10.596299  9.942742
#>  [50]  8.943500 14.412632 10.715120  9.798627  6.039729 15.060428 11.870587
#>  [57]  9.768820  8.309935 14.314956 12.323846  9.139689  6.304845  8.787492
#>  [64]  7.398115 10.374749  4.592377  5.104921 18.162290 12.279780  7.958612
#>  [71]  6.785633 12.291107  9.705732 18.862821 11.358879  6.888663  4.648658
#>  [78]  9.872321  7.362507 11.267782 12.249741 13.187064  5.828882  7.418625
#>  [85]  8.768993 19.821743 19.559857  8.574784  8.745862  8.039832  2.232106
#>  [92]  4.348408 19.907047  8.329430 19.945017  8.789758  8.822387 19.950863
#>  [99]  7.445026 11.356745
#> 
#> $statistics$acceptance_rate
#>   [1]   NA 0.85   NA 0.70   NA   NA 0.75   NA 0.60   NA 0.60 0.40 0.25 0.45 0.45
#>  [16] 0.40 0.20 0.25 0.25 0.10 0.25 0.35 0.40 0.30 0.25 0.30 0.25 0.40 0.45 0.85
#>  [31] 0.50 0.40 0.65 0.55   NA   NA 0.40   NA   NA 0.50   NA   NA   NA   NA 0.40
#>  [46]   NA   NA   NA 0.45 0.65   NA   NA 0.55 0.60   NA   NA 0.50 0.65   NA   NA
#>  [61] 0.50 0.80 0.75 0.45   NA 0.65 0.65   NA   NA 0.40 0.30   NA 0.55   NA   NA
#>  [76] 0.45 0.50 0.50 0.55   NA   NA   NA 0.45 0.25 0.50   NA   NA 0.25 0.40 0.35
#>  [91] 0.30 0.30   NA 0.25   NA 0.30 0.25   NA 0.30   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 620
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.7099792 0.2511521
#>  [2,] 0.7099792 0.2511521
#>  [3,] 0.7099792 0.2511521
#>  [4,] 0.7099792 0.2511521
#>  [5,] 0.7099792 0.2511521
#>  [6,] 0.7099792 0.2511521
#>  [7,] 0.7099792 0.2511521
#>  [8,] 0.7099792 0.2511521
#>  [9,] 0.7099792 0.2511521
#> [10,] 0.7099792 0.2511521
#> [11,] 0.7099792 0.2511521
#> [12,] 0.7099792 0.2511521
#> [13,] 0.7099792 0.2511521
#> [14,] 0.7099792 0.2511521
#> [15,] 0.7099792 0.2511521
#> [16,] 0.7099792 0.2511521
#> [17,] 0.7099792 0.2511521
#> [18,] 0.7099792 0.2511521
#> [19,] 0.7099792 0.2511521
#> [20,] 0.7099792 0.2511521