Deterministic particle likelihood

Create a deterministic version of the particle_filter object, which runs a single particle deterministically.

Public fields

model

The dust model generator being simulated (cannot be re-bound)

Methods

Public methods

particle_deterministic$new()

particle_deterministic$run()

particle_deterministic$run_many()

particle_deterministic$state()

particle_deterministic$history()

particle_deterministic$restart_state()

particle_deterministic$inputs()

particle_deterministic$set_n_threads()

Method new()

Create the particle filter

Usage

particle_deterministic$new(
  data,
  model,
  compare,
  index = NULL,
  initial = NULL,
  n_threads = 1L
)

Arguments

data

The data set to be used for the particle filter, created by particle_filter_data(). This is essentially a data.frame() with at least columns step_start and step_end, along with any additional data used in the compare function, and additional information about how your steps relate to time.

model

A stochastic model to use. Must be a dust_generator object.

compare

A comparison function. Must take arguments state, observed and pars as arguments (though the arguments may have different names). state is the simulated model state (a matrix with as many rows as there are state variables and as many columns as there are particles, data is a list of observed data corresponding to the current time's row in the data object provided here in the constructor. pars is any additional parameters passed through to the comparison function (via the pars argument to $run). Alternatively, compare can be NULL if your model provides a built-in compile compare function (if model$public_methods$has_compare() is TRUE), which may be faster.

index

An index function. This is used to compute the "interesting" indexes of your model. It must be a function of one argument, which will be the result of calling the $info() method on your model. It should return a list with elements run (indices to return at the end of each run, passed through to your compare function) and state (indices to return if saving state). These indices can overlap but do not have to. This argument is optional but using it will likely speed up your simulation if you have more than a few states as it will reduce the amount of memory copied back and forth.

initial

A function to generate initial conditions. If given, then this function must accept 3 arguments: info (the result of calling $info() as for index), n_particles (the number of particles that the particle filter is using) and pars (parameters passed in in the $run method via the pars argument). It must return a list, which can have the elements state (initial model state, passed to the particle filter - either a vector or a matrix, and overriding the initial conditions provided by your model) and step (the initial step, overriding the first step of your data - this must occur within your first epoch in your data provided to the constructor, i.e., not less than the first element of step_start and not more than step_end). Your function can also return a vector or matrix of state and not alter the starting step, which is equivalent to returning list(state = state, step = NULL).

n_threads

Number of threads to use when running the simulation. Defaults to 1, and should not be set higher than the number of cores available to the machine. This currently has no effect as the simulation will be run in serial on a single particle for now.

Method run()

Run the deterministic particle filter

Usage

particle_deterministic$run(
  pars = list(),
  save_history = FALSE,
  save_restart = NULL,
  min_log_likelihood = -Inf
)

Arguments

pars

A list representing parameters. This will be passed as the pars argument to your model, to your compare function, and (if using) to your initial function. It must be an R list (not vector or NULL) because that is what a dust model currently requires on initialisation or $reset - we may relax this later. You may want to put your observation and initial parameters under their own keys (e.g., pars$initial$whatever), but this is up to you. Extra keys are silently ignored by dust models.

save_history

Logical, indicating if the history of all particles should be saved. If saving history, then it can be queried later with the $history method on the object.

save_restart

An integer vector of time points to save restart infomation for. Not currently supported.

min_log_likelihood

Not currently supported, exists to match the inteface with particle_filter. Providing a value larger than -Inf will cause an error.

Returns

A single numeric value representing the log-likelihood (-Inf if the model is impossible)

Method run_many()

Run the deterministic particle filter on several parameter sets simultaneously. This acts as a wrapper around $run(), though runs will be in parallel if the object was created with n_threads greater than 1.

Usage

particle_deterministic$run_many(
  pars,
  save_history = FALSE,
  save_restart = NULL,
  min_log_likelihood = -Inf
)

Arguments

pars

A list of parameter sets, each element of which would be suitable to pass to $run()

save_history

Logical, indicating if the history of all particles should be saved.

save_restart

An integer vector of time points to save restart infomation for. Not currently supported.

min_log_likelihood

Not currently supported, exists to match the inteface with particle_filter. Providing a value larger than -Inf will cause an error.

Returns

A numeric vector of values representing the log-likelihood (-Inf if the model is impossible), one per parameter set

Method state()

Extract the current model state, optionally filtering. If the model has not yet been run, then this method will throw an error. Returns a matrix with the number of rows being the number of model states, and the number of columns being the number of particles.

Usage

particle_deterministic$state(index_state = NULL)

Arguments

index_state

Optional vector of states to extract

Method history()

Extract the particle trajectories. Requires that the model was run with save_history = TRUE, which does incur a performance cost. This method will throw an error if the model has not run, or was run without save_history = TRUE. Returns a 3d array with dimensions corresponding to (1) model state, filtered by index$run if provided, (2) particle (following index_particle if provided), (3) time point.

Usage

particle_deterministic$history(index_particle = NULL)

Arguments

index_particle

Optional vector of particle indices to return. If NULL we return all particles' histories.

Method restart_state()

Return the full particle filter state at points back in time that were saved with the save_restart argument to $run(). If available, this will return a 3d array, with dimensions representing (1) particle state, (2) particle index, (3) time point. If nested parameters are used then returns a 4d array, with dimensions representing (1) particle state, (2) particle index, (3) population, (4) time point. This could be quite large, especially if you are using the index argument to create the particle filter and return a subset of all state generally. In the stochastic version, this is different the saved trajectories returned by $history() because earlier saved state is not filtered by later filtering, but in the deterministic model we run with a single particle so it is the same.

Usage

particle_deterministic$restart_state(index_particle = NULL)

Arguments

index_particle

Optional vector of particle indices to return. If NULL we return all particles' states. Practically because the only valid value of index_particle is "1", this has no effect and it is included primarily for compatibility with the stochastic interface.

Method inputs()

Return a list of inputs used to configure the deterministic particle filter. These correspond directly to the argument names for the constructor and are the same as the input arguments.

Usage

particle_deterministic$inputs()

Method set_n_threads()

Set the number of threads used by the particle filter (and dust model) after creation. This can be used to allocate additional (or subtract excess) computing power from the deterministic filter Returns (invisibly) the previous value.

Usage

particle_deterministic$set_n_threads(n_threads)

Arguments

n_threads

The new number of threads to use. You may want to wrap this argument in dust::dust_openmp_threads() in order to verify that you can actually use the number of threads requested (based on environment variables and OpenMP support).