pmcmc_parameters_nested

Construct nested parameters for use with pmcmc(). This creates a utility object that is used internally to work with parameters that may be fixed and the same for all given populations, or varied and possibly-different between populations. Most users only need to construct this object, but see the examples for how it can be used.

Methods

Public methods

pmcmc_parameters_nested$new()

pmcmc_parameters_nested$names()

pmcmc_parameters_nested$populations()

pmcmc_parameters_nested$summary()

pmcmc_parameters_nested$initial()

pmcmc_parameters_nested$prior()

pmcmc_parameters_nested$propose()

pmcmc_parameters_nested$model()

pmcmc_parameters_nested$fix()

Method new()

Create the pmcmc_parameters object

Usage

pmcmc_parameters_nested$new(
  parameters,
  proposal_varied = NULL,
  proposal_fixed = NULL,
  populations = NULL,
  transform = NULL
)

Arguments

parameters

A list of pmcmc_parameter or pmcmc_varied_parameter objects, each of which describe a single (possibly-varying) parameter in your model. If parameters is named, then these names must match the $name element of each parameter that is used (this is verified).

proposal_varied, proposal_fixed

Square proposal matrices corresponding to the variance-covariance matrix of a multivariate gaussian distribution used to generate new varied and fixed parameters respectively.'. They must have the same number of rows and columns as there are varied and fixed parameters respectively. The names must correspond exactly to the names in parameters. Because it corresponds to a variance-covariance matrix it must be symmetric and positive definite.

populations

Specifies the names of the different populations that the varying parameters change according to. Only required if no pmcmc_varied_parameter objects are included in parameters. Otherwise population names are taken from those objects.

transform

An optional transformation function to apply to your parameter vector immediately before passing it to the model function. If not given, then as.list is used, as dust models require this. However, if you need to generate derived parameters from those being actively sampled you can do arbitrary transformations here.

Method names()

Return the names of the parameters

Usage

pmcmc_parameters_nested$names()

Method populations()

Return the names of the populations

Usage

pmcmc_parameters_nested$populations()

Method summary()

Return a data.frame with information about parameters (name, min, max, discrete, fixed or varied).

Usage

pmcmc_parameters_nested$summary(population = NULL)

Arguments

population

Specifies which population to summarise. If NULL then returns a summary of each population as a list.

Method initial()

Return the initial parameter values as a named matrix with rows corresponding to populations and columns are parameters.

Usage

pmcmc_parameters_nested$initial()

Method prior()

Compute the prior(s) for a parameter matrix. Returns a named vector with names corresponding to populations.

Usage

pmcmc_parameters_nested$prior(theta)

Arguments

theta

a parameter matrix with columns in the same order as $names() and rows in the same order as $populations().

Method propose()

This proposes a new parameter matrix given your current matrix and the variance-covariance matrices of the proposal kernels, discretises any discrete values, and reflects bounded parameters until they lie within min:max. Returns matrix with rows corresponding to populations and columns to parameters.

Usage

pmcmc_parameters_nested$propose(theta, type, scale = 1)

Arguments

theta

a parameter matrix with columns in the same order as $names() and rows in the same order as $populations().

type

specifies which type of parameters should be proposed, either fixed parameters only ("fixed"), varied only ("varied"), or both ("both") types. For 'fixed' and 'varied', parameters of the other type are left unchanged.

scale

an optional scaling factor to apply to the proposal distribution. This may be useful in sampling starting points. The parameter is equivalent to a multiplicative factor applied to the variance covariance matrix.

Method model()

Apply the model transformation function to a parameter matrix.

Usage

pmcmc_parameters_nested$model(theta)

Arguments

theta

a parameter matrix with columns in the same order as $names() and rows in the same order as $populations().

Method fix()

Set some parameters to fixed values. Use this to reduce the dimensionality of your system.

Usage

pmcmc_parameters_nested$fix(fixed)

Arguments

fixed

a named vector of parameters to fix

Examples

#Construct an object with two varied parameters, two fixed parameters,
#and two populations:
parameters <- list(pmcmc_varied_parameter("a", c("p1", "p2"), 2),
                   pmcmc_varied_parameter("b", c("p1", "p2"), 2),
                   pmcmc_parameter("c", 3),
                   pmcmc_parameter("d", 4))
proposal_fixed <- diag(2)
proposal_varied <- diag(2) + 1
pars <- mcstate::pmcmc_parameters_nested$new(parameters, proposal_varied,
                                             proposal_fixed)

# Initial parameters
p <- pars$initial()
p
#>    a b c d
#> p1 2 2 3 4
#> p2 2 2 3 4

# Propose a new parameter point
pars$propose(p, type = "both")
#>            a          b        c        d
#> p1 2.3553791  0.1494514 3.488047 3.603335
#> p2 0.1759955 -0.6924538 3.488047 3.603335

# Information about parameters:
pars$names()
#> [1] "a" "b" "c" "d"
pars$summary()
#> $p1
#>   name  min max discrete   type
#> 1    a -Inf Inf    FALSE varied
#> 2    b -Inf Inf    FALSE varied
#> 3    c -Inf Inf    FALSE  fixed
#> 4    d -Inf Inf    FALSE  fixed
#> 
#> $p2
#>   name  min max discrete   type
#> 1    a -Inf Inf    FALSE varied
#> 2    b -Inf Inf    FALSE varied
#> 3    c -Inf Inf    FALSE  fixed
#> 4    d -Inf Inf    FALSE  fixed
#> 

# Compute prior
pars$prior(p)
#> p1 p2 
#>  0  0 

# Transform data for your model
pars$model(p)
#> [[1]]
#> [[1]]$a
#> [1] 2
#> 
#> [[1]]$b
#> [1] 2
#> 
#> [[1]]$c
#> [1] 3
#> 
#> [[1]]$d
#> [1] 4
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1] 2
#> 
#> [[2]]$b
#> [1] 2
#> 
#> [[2]]$c
#> [1] 3
#> 
#> [[2]]$d
#> [1] 4
#> 
#>