Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 4 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.6234181 0.3373020
#>  [2,] 0.6395200 0.5048188
#>  [3,] 0.6839187 0.6355581
#>  [4,] 0.8461452 0.4726960
#>  [5,] 0.9414498 0.5457599
#>  [6,] 0.6240945 0.4831697
#>  [7,] 0.7723326 0.6115604
#>  [8,] 0.7617236 0.6542840
#>  [9,] 0.8033381 0.6487537
#> [10,] 0.8626510 0.5912032
#> [11,] 0.7048701 0.6053085
#> [12,] 0.7250251 0.6443289
#> [13,] 0.7125489 0.4683012
#> [14,] 0.6537150 0.4371837
#> [15,] 0.8286374 0.6582860
#> [16,] 0.8053705 0.5717765
#> [17,] 0.9376593 0.5109822
#> [18,] 0.6827583 0.5913065
#> [19,] 0.7313023 0.6551383
#> [20,] 0.7549054 0.6804744
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -274.4436     -274.4436 7.870105e-06
#>  [2,]         0      -270.4250     -270.4250 4.377481e-04
#>  [3,]         0      -276.8407     -276.8407 7.160343e-07
#>  [4,]         0      -280.1793     -280.1793 2.540862e-08
#>  [5,]         0      -294.8872     -294.8872 1.040932e-14
#>  [6,]         0      -262.7031     -262.7031 9.881176e-01
#>  [7,]         0      -267.9916     -267.9916 4.989391e-03
#>  [8,]         0      -271.0116     -271.0116 2.434750e-04
#>  [9,]         0      -272.1403     -272.1403 7.875821e-05
#> [10,]         0      -271.2243     -271.2243 1.968350e-04
#> [11,]         0      -275.0062     -275.0062 4.483798e-06
#> [12,]         0      -277.3026     -277.3026 4.511359e-07
#> [13,]         0      -271.5951     -271.5951 1.358450e-04
#> [14,]         0      -267.9108     -267.9108 5.409427e-03
#> [15,]         0      -270.7569     -270.7569 3.141088e-04
#> [16,]         0      -275.7490     -275.7490 2.133312e-06
#> [17,]         0      -291.5153     -291.5153 3.032834e-13
#> [18,]         0      -272.5349     -272.5349 5.307623e-05
#> [19,]         0      -274.4271     -274.4271 8.001343e-06
#> [20,]         0      -280.3328     -280.3328 2.179321e-08
#> 
#> $statistics
#> $statistics$ess
#>   [1] 15.732507 11.868600 10.920849 10.393475  9.201074 10.804517  8.276790
#>   [8] 13.763654 11.785464 10.159998  8.614089  6.365137 10.122098  7.889690
#>  [15] 13.976132 12.297198 10.012360  6.794154 12.301130  7.942866  5.356892
#>  [22]  9.532992  2.566999  4.299546  3.181808  3.895119  5.114685  7.987221
#>  [29]  9.283610 14.532961 10.380573  7.826367 17.184555 12.484384 12.528777
#>  [36] 11.351021 10.949340 10.376021  9.833956 19.148202 16.978123 13.469348
#>  [43] 14.861278 14.220665 14.108280 13.366425 14.588515 14.319825 14.339882
#>  [50] 11.703693 12.380811 11.548818 11.443330  7.404204 19.385837 19.208523
#>  [57] 18.979078 17.344709 17.946749 17.871910 16.045569 14.938465 13.749890
#>  [64] 12.668075 12.286740 11.004711  9.838441 19.422780 19.115187 18.387560
#>  [71] 16.409979 16.206907 15.594516 15.804389 14.863949 13.284438 11.175525
#>  [78] 11.110853 10.894148 10.134144 10.119972  9.851340 17.595379 15.557668
#>  [85] 14.406781 14.245937 15.234175 12.266100 11.333287  9.685388  9.807150
#>  [92] 16.440868 17.704926 16.146773 15.986584 12.523675 12.151394 12.920751
#>  [99] 10.033605  8.831434
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA   NA   NA 0.60   NA 0.65   NA   NA   NA 0.60 0.60   NA 0.50   NA
#>  [16]   NA   NA 0.20   NA 0.40 0.15 0.25 0.40 0.20 0.55 0.45 0.70 0.45 0.45   NA
#>  [31]   NA 0.50   NA   NA   NA   NA   NA   NA 0.50   NA   NA   NA   NA   NA   NA
#>  [46]   NA   NA   NA   NA   NA   NA   NA   NA 0.45   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA   NA   NA   NA   NA 0.35   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76]   NA   NA   NA   NA   NA   NA 0.25   NA   NA   NA   NA   NA   NA   NA 0.15
#>  [91] 0.20   NA   NA   NA   NA   NA   NA   NA   NA 0.30
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 193.4
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.6240945 0.4831697
#>  [2,] 0.6240945 0.4831697
#>  [3,] 0.6240945 0.4831697
#>  [4,] 0.6240945 0.4831697
#>  [5,] 0.6240945 0.4831697
#>  [6,] 0.6240945 0.4831697
#>  [7,] 0.6240945 0.4831697
#>  [8,] 0.6240945 0.4831697
#>  [9,] 0.6240945 0.4831697
#> [10,] 0.6240945 0.4831697
#> [11,] 0.6240945 0.4831697
#> [12,] 0.6240945 0.4831697
#> [13,] 0.6240945 0.4831697
#> [14,] 0.6240945 0.4831697
#> [15,] 0.6240945 0.4831697
#> [16,] 0.6240945 0.4831697
#> [17,] 0.6240945 0.4831697
#> [18,] 0.6240945 0.4831697
#> [19,] 0.6240945 0.4831697
#> [20,] 0.6240945 0.4831697