Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 6 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.5197561 0.4203240
#>  [2,] 0.9360745 0.8713031
#>  [3,] 0.5060630 0.3885833
#>  [4,] 0.5970785 0.5343210
#>  [5,] 0.5859347 0.8321037
#>  [6,] 0.7775521 0.5173928
#>  [7,] 0.3753704 0.2869705
#>  [8,] 0.5991985 0.3871409
#>  [9,] 0.5376235 0.4280069
#> [10,] 0.5789818 0.4582334
#> [11,] 0.6581683 0.6923047
#> [12,] 0.6015897 0.5564343
#> [13,] 0.4893063 0.3600380
#> [14,] 0.4693464 0.3813338
#> [15,] 0.5931343 0.3975371
#> [16,] 0.7578384 0.6308409
#> [17,] 0.5640883 0.5027904
#> [18,] 0.5341106 0.4799995
#> [19,] 0.4524607 0.4132473
#> [20,] 0.6858812 0.4805250
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -264.1154     -264.1154 3.038853e-02
#>  [2,]         0      -279.1774     -279.1774 8.737546e-09
#>  [3,]         0      -261.3013     -261.3013 5.068628e-01
#>  [4,]         0      -281.6458     -281.6458 7.402709e-10
#>  [5,]         0      -280.5710     -280.5710 2.168408e-09
#>  [6,]         0      -272.2325     -272.2325 9.067724e-06
#>  [7,]         0      -264.6333     -264.6333 1.810566e-02
#>  [8,]         0      -263.9379     -263.9379 3.629223e-02
#>  [9,]         0      -263.0195     -263.0195 9.092117e-02
#> [10,]         0      -263.9036     -263.9036 3.756006e-02
#> [11,]         0      -286.2269     -286.2269 7.582729e-12
#> [12,]         0      -272.1971     -272.1971 9.394985e-06
#> [13,]         0      -263.3627     -263.3627 6.451176e-02
#> [14,]         0      -262.6157     -262.6157 1.361532e-01
#> [15,]         0      -264.3886     -264.3886 2.312500e-02
#> [16,]         0      -274.1069     -274.1069 1.391476e-06
#> [17,]         0      -268.0129     -268.0129 6.166617e-04
#> [18,]         0      -263.5197     -263.5197 5.513520e-02
#> [19,]         0      -268.8261     -268.8261 2.734599e-04
#> [20,]         0      -270.9000     -270.9000 3.437269e-05
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.183342 11.882802  9.674453 17.671977 13.663395 12.206446  9.644357
#>   [8] 15.777671 14.320526 13.050810 12.212281  9.520993 16.673528 14.053620
#>  [15] 12.764683 11.381960 10.592515  9.389458 12.346041 11.437979  9.952211
#>  [22]  8.399516 10.375185  8.200510  8.034419  8.940084  9.958111 13.952458
#>  [29]  9.293102  9.805729 14.561567 12.148014  9.050292 17.874314  9.294458
#>  [36] 12.410034  6.280061  8.133529  6.903399  5.727166  5.275826  5.000346
#>  [43] 13.214830 12.271121  9.209150 17.979619  6.059395 14.283268 10.348492
#>  [50]  3.573619 19.297792 13.295990 13.955129  5.799482 19.577174 19.213633
#>  [57] 12.497738  6.185207 19.654561 19.339017 17.449361 13.272129  9.926418
#>  [64] 19.350643 18.619370 12.407120  6.818404 18.915902 19.572475 18.124177
#>  [71] 13.960781 11.854166 10.671816 12.723488  8.471413 18.836547 13.825313
#>  [78] 12.331096  9.672005 19.652533 19.457712 18.999489 17.541426 16.228450
#>  [85] 14.848440 15.527353 15.913181 13.576108 12.213379  7.729070 12.317634
#>  [92]  8.945315 19.575301 19.083439 18.675688 17.582337 16.264087 16.530564
#>  [99] 12.751573 11.170358
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.70   NA   NA   NA 0.45   NA   NA   NA   NA 0.65   NA   NA   NA
#>  [16]   NA   NA 0.60   NA   NA 0.40 0.45   NA 0.40 0.45 0.35 0.55   NA 0.60 0.40
#>  [31]   NA   NA 0.30   NA 0.65   NA 0.50 0.55 0.45 0.75 0.55 0.50   NA   NA 0.50
#>  [46]   NA 0.65   NA   NA 0.55   NA   NA   NA 0.50   NA   NA   NA 0.55   NA   NA
#>  [61]   NA   NA 0.30   NA   NA   NA 0.35   NA   NA   NA   NA   NA   NA   NA 0.45
#>  [76]   NA   NA   NA 0.35   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.50
#>  [91]   NA 0.50   NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 273.8
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.4693464 0.3813338
#>  [2,] 0.4893063 0.3600380
#>  [3,] 0.5376235 0.4280069
#>  [4,] 0.5376235 0.4280069
#>  [5,] 0.5197561 0.4203240
#>  [6,] 0.5931343 0.3975371
#>  [7,] 0.5197561 0.4203240
#>  [8,] 0.5060630 0.3885833
#>  [9,] 0.5060630 0.3885833
#> [10,] 0.4893063 0.3600380
#> [11,] 0.5060630 0.3885833
#> [12,] 0.4693464 0.3813338
#> [13,] 0.5060630 0.3885833
#> [14,] 0.4693464 0.3813338
#> [15,] 0.5060630 0.3885833
#> [16,] 0.5060630 0.3885833
#> [17,] 0.5376235 0.4280069
#> [18,] 0.5060630 0.3885833
#> [19,] 0.5376235 0.4280069
#> [20,] 0.5060630 0.3885833