Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)
#> Warning: 'initial_time' should be provided. I'm assuming '0' which is one time unit before the first time in your data (1), but this might not be appropriate. This will become an error in a future version of mcstate

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 4 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.9549419 0.9226837
#>  [2,] 0.9076964 0.8501227
#>  [3,] 0.6861023 0.5124258
#>  [4,] 0.4733528 0.2464084
#>  [5,] 0.3329879 0.2690083
#>  [6,] 0.4157179 0.2756735
#>  [7,] 0.6306680 0.3118022
#>  [8,] 0.8593084 0.5033835
#>  [9,] 0.4228297 0.2434155
#> [10,] 0.7613022 0.7118823
#> [11,] 0.5207209 0.3803546
#> [12,] 0.5668609 0.3817547
#> [13,] 0.4640919 0.2903486
#> [14,] 0.4114577 0.2744632
#> [15,] 0.6599055 0.5318405
#> [16,] 0.6671389 0.5235704
#> [17,] 0.9999649 0.9238741
#> [18,] 0.3008650 0.1644318
#> [19,] 0.3008650 0.1644318
#> [20,] 0.4343369 0.2844818
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -275.8553     -275.8553 1.606282e-10
#>  [2,]         0      -276.1228     -276.1228 1.229282e-10
#>  [3,]         0      -268.2317     -268.2317 3.286530e-07
#>  [4,]         0      -263.8450     -263.8450 2.641524e-05
#>  [5,]         0      -259.7607     -259.7607 1.569102e-03
#>  [6,]         0      -256.8073     -256.8073 3.007951e-02
#>  [7,]         0      -280.1212     -280.1212 2.255080e-12
#>  [8,]         0      -279.1445     -279.1445 5.988915e-12
#>  [9,]         0      -260.5388     -260.5388 7.206212e-04
#> [10,]         0      -273.1212     -273.1212 2.473192e-09
#> [11,]         0      -262.9907     -262.9907 6.206604e-05
#> [12,]         0      -261.2793     -261.2793 3.436519e-04
#> [13,]         0      -261.1184     -261.1184 4.036496e-04
#> [14,]         0      -259.7132     -259.7132 1.645360e-03
#> [15,]         0      -272.7429     -272.7429 3.610079e-09
#> [16,]         0      -263.5565     -263.5565 3.524887e-05
#> [17,]         0      -277.3034     -277.3034 3.775223e-11
#> [18,]         0      -253.9850     -253.9850 5.057990e-01
#> [19,]         0      -254.0964     -254.0964 4.524723e-01
#> [20,]         0      -258.2880     -258.2880 6.842746e-03
#> 
#> $statistics
#> $statistics$ess
#>   [1] 17.117092 11.961735  8.204669 16.256308 13.291454 12.131947 10.974376
#>   [8]  8.768810 15.286213 11.856751  9.683432 10.783284  8.977379 17.064429
#>  [15] 14.396185 12.353040 11.122073 10.168630  8.904058 15.043791 10.341715
#>  [22] 10.748471  8.600119 12.728187 11.201522  9.384770 13.645836 10.015987
#>  [29]  8.337727 17.039665 14.936260 13.069222 12.741636 10.283858 11.625582
#>  [36] 11.041593  9.974958 17.943411 15.750932 13.175514 11.198493  8.122256
#>  [43] 19.692267 16.705627 18.070309 17.005194 11.617834  9.992624 19.234045
#>  [50] 10.955990 12.796412  9.289936 19.631928 12.840949 14.427512 14.128495
#>  [57] 10.879213  8.960663 19.517741 19.137751 18.360640 15.771009 13.947639
#>  [64] 12.714465 12.623091  9.917313 18.506715 18.923427 18.367858 16.606272
#>  [71] 13.953784 13.828780 12.379156 15.944468 11.552038  8.111513 18.469461
#>  [78] 17.985455 14.914573 13.617201 14.537852 15.121365 10.204479  8.533563
#>  [85] 19.618287 19.285753 18.887039 17.524268 16.855423 14.658326  9.219110
#>  [92] 16.780858 17.718004 16.926841 17.405991 12.249930 11.211470 11.948223
#>  [99]  7.821049 19.418316
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.65   NA   NA   NA   NA 0.40   NA   NA 0.75   NA 0.45   NA   NA
#>  [16]   NA   NA   NA 0.40   NA   NA   NA 0.35   NA   NA 0.30   NA   NA 0.50   NA
#>  [31]   NA   NA   NA   NA   NA   NA 0.60   NA   NA   NA   NA 0.30   NA   NA   NA
#>  [46]   NA   NA 0.30   NA   NA   NA 0.45   NA   NA   NA   NA   NA 0.35   NA   NA
#>  [61]   NA   NA   NA   NA   NA 0.40   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76] 0.35   NA   NA   NA   NA   NA   NA   NA 0.25   NA   NA   NA   NA   NA   NA
#>  [91] 0.30   NA   NA   NA   NA   NA   NA   NA 0.30   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 177
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.3008650 0.1644318
#>  [2,] 0.3008650 0.1644318
#>  [3,] 0.4343369 0.2844818
#>  [4,] 0.3008650 0.1644318
#>  [5,] 0.4157179 0.2756735
#>  [6,] 0.3008650 0.1644318
#>  [7,] 0.3008650 0.1644318
#>  [8,] 0.3008650 0.1644318
#>  [9,] 0.3008650 0.1644318
#> [10,] 0.3008650 0.1644318
#> [11,] 0.3008650 0.1644318
#> [12,] 0.3008650 0.1644318
#> [13,] 0.3008650 0.1644318
#> [14,] 0.3008650 0.1644318
#> [15,] 0.3008650 0.1644318
#> [16,] 0.3008650 0.1644318
#> [17,] 0.3008650 0.1644318
#> [18,] 0.3008650 0.1644318
#> [19,] 0.3008650 0.1644318
#> [20,] 0.3008650 0.1644318