Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 5 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.7664672 0.4467154
#>  [2,] 0.8546322 0.3522865
#>  [3,] 0.9208576 0.3225362
#>  [4,] 0.9208576 0.3225362
#>  [5,] 0.7648966 0.2956328
#>  [6,] 0.8672826 0.3102614
#>  [7,] 0.9711873 0.3985973
#>  [8,] 0.8932340 0.4648248
#>  [9,] 0.8981123 0.2968052
#> [10,] 0.8981123 0.2968052
#> [11,] 0.8981123 0.2968052
#> [12,] 0.8981123 0.2968052
#> [13,] 0.9389731 0.4108697
#> [14,] 0.8086975 0.3965604
#> [15,] 0.7482737 0.2268260
#> [16,] 0.7482737 0.2268260
#> [17,] 0.7635517 0.2535270
#> [18,] 0.7854095 0.2391612
#> [19,] 0.7025768 0.2105838
#> [20,] 0.6266909 0.2413862
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior        weight
#>  [1,]         0      -274.5515     -274.5515  9.999999e-01
#>  [2,]         0      -389.6721     -389.6721  1.008742e-50
#>  [3,]         0      -736.3710     -736.3710 2.718483e-201
#>  [4,]         0      -748.4686     -748.4686 1.515082e-206
#>  [5,]         0      -442.4534     -442.4534  1.205448e-73
#>  [6,]         0      -718.9392     -718.9392 1.011320e-193
#>  [7,]         0      -417.1261     -417.1261  1.204079e-62
#>  [8,]         0      -301.0073     -301.0073  3.238797e-12
#>  [9,]         0      -917.5631     -917.5631 5.541050e-280
#> [10,]         0      -916.4646     -916.4646 1.662249e-279
#> [11,]         0      -915.2289     -915.2289 5.719094e-279
#> [12,]         0      -915.5842     -915.5842 4.009176e-279
#> [13,]         0      -472.2225     -472.2225  1.420911e-86
#> [14,]         0      -291.1874     -291.1874  5.958040e-08
#> [15,]         0      -844.5018     -844.5018 2.976580e-248
#> [16,]         0      -842.6810     -842.6810 1.838697e-247
#> [17,]         0      -482.8223     -482.8223  3.541119e-91
#> [18,]         0      -803.0033     -803.0033 3.135419e-230
#> [19,]         0      -853.2235     -853.2235 4.852165e-252
#> [20,]         0      -336.0285     -336.0285  1.999394e-27
#> 
#> $statistics
#> $statistics$ess
#>   [1] 11.519772  6.738048 13.317104  8.907025 11.770801  8.439787 11.492233
#>   [8]  9.574217 14.617692 12.393759 10.825429  8.258747 13.852669 12.079999
#>  [15] 10.920697  9.801042 13.965327 10.776075  9.131392 13.081511  8.834323
#>  [22] 14.577220  8.946775  8.710016 11.551450  8.854943 11.046806  7.029189
#>  [29] 11.415052 10.556876  9.223034 10.539798  8.794489  9.847069 14.897093
#>  [36] 13.562921 11.224665 10.048667  8.716416 18.244402 15.003842 11.453665
#>  [43] 12.308776 11.595312 11.493776 10.936708  8.615789 17.980478 17.116838
#>  [50] 14.047582 15.114203 13.702743 14.728659 11.777018 12.157346 12.756485
#>  [57] 11.742383 10.097394 11.295475 11.783240 11.548775 10.779943  9.953829
#>  [64] 17.718036 16.781422 13.582819  9.606440 13.911837 18.019828 16.393259
#>  [71] 13.873432 13.708524 13.167175 14.203029 12.308120 10.380356  8.856641
#>  [78] 16.113529 12.553308 11.474033 12.979000 13.508059 10.012251  8.048953
#>  [85] 15.362759 16.457475 17.383841 11.563265  7.977679 12.143819  6.046335
#>  [92]  8.265564 18.797919 14.233639 14.972300  7.368248 12.025090 13.297236
#>  [99]  5.031628 14.675584
#> 
#> $statistics$acceptance_rate
#>   [1]   NA 0.90   NA 0.70   NA 0.95   NA 0.45   NA   NA   NA 0.50   NA   NA   NA
#>  [16] 0.65   NA   NA 0.60   NA 0.55   NA 0.70 0.70   NA 0.65   NA 0.70   NA   NA
#>  [31] 0.35   NA 0.35 0.60   NA   NA   NA   NA 0.35   NA   NA   NA   NA   NA   NA
#>  [46]   NA 0.40   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA 0.50   NA   NA   NA 0.45   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76]   NA 0.40   NA   NA   NA   NA   NA   NA 0.70   NA   NA   NA   NA 0.50   NA
#>  [91] 0.45 0.45   NA   NA   NA 0.30   NA   NA 0.35   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 246.2
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.7664672 0.4467154
#>  [2,] 0.7664672 0.4467154
#>  [3,] 0.7664672 0.4467154
#>  [4,] 0.7664672 0.4467154
#>  [5,] 0.7664672 0.4467154
#>  [6,] 0.7664672 0.4467154
#>  [7,] 0.7664672 0.4467154
#>  [8,] 0.7664672 0.4467154
#>  [9,] 0.7664672 0.4467154
#> [10,] 0.7664672 0.4467154
#> [11,] 0.7664672 0.4467154
#> [12,] 0.7664672 0.4467154
#> [13,] 0.7664672 0.4467154
#> [14,] 0.7664672 0.4467154
#> [15,] 0.7664672 0.4467154
#> [16,] 0.7664672 0.4467154
#> [17,] 0.7664672 0.4467154
#> [18,] 0.7664672 0.4467154
#> [19,] 0.7664672 0.4467154
#> [20,] 0.7664672 0.4467154