Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 5 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.5696875 0.3574630
#>  [2,] 0.6241402 0.3585820
#>  [3,] 0.6129622 0.2643342
#>  [4,] 0.6906108 0.2873884
#>  [5,] 0.8045081 0.2653167
#>  [6,] 0.8194279 0.2700554
#>  [7,] 0.7272871 0.3716646
#>  [8,] 0.6209753 0.3795636
#>  [9,] 0.8457813 0.3437775
#> [10,] 0.8577606 0.3067575
#> [11,] 0.8160082 0.3838635
#> [12,] 0.8353097 0.3289259
#> [13,] 0.6520671 0.3505870
#> [14,] 0.7103166 0.3664761
#> [15,] 0.8552697 0.2977352
#> [16,] 0.6798036 0.3023436
#> [17,] 0.7559114 0.3941302
#> [18,] 0.8018514 0.3154747
#> [19,] 0.8552697 0.2977352
#> [20,] 0.5705562 0.2977946
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior        weight
#>  [1,]         0      -263.2819     -263.2819  9.847357e-01
#>  [2,]         0      -267.4533     -267.4533  1.519622e-02
#>  [3,]         0      -315.4067     -315.4067  2.268876e-23
#>  [4,]         0      -329.6166     -329.6166  1.529537e-29
#>  [5,]         0      -798.9071     -798.9071 2.367455e-233
#>  [6,]         0      -584.7868     -584.7868 2.320176e-140
#>  [7,]         0      -291.5366     -291.5366  5.277927e-13
#>  [8,]         0      -274.1678     -274.1678  1.843531e-05
#>  [9,]         0      -447.7982     -447.7982  7.226583e-81
#> [10,]         0      -668.1674     -668.1674 1.424938e-176
#> [11,]         0      -290.1496     -290.1496  2.112646e-12
#> [12,]         0      -341.7596     -341.7596  8.144967e-35
#> [13,]         0      -273.3040     -273.3040  4.373001e-05
#> [14,]         0      -298.2223     -298.2223  6.590483e-16
#> [15,]         0      -517.1080     -517.1080 5.728750e-111
#> [16,]         0      -310.4742     -310.4742  3.147723e-21
#> [17,]         0      -284.6498     -284.6498  5.168672e-10
#> [18,]         0      -523.0182     -523.0182 1.553426e-113
#> [19,]         0      -527.2677     -527.2677 2.216962e-115
#> [20,]         0      -275.3046     -275.3046  5.915030e-06
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.834072 11.492461  9.090395 12.001743  8.556956 12.853388  9.837557
#>   [8] 11.743339 10.267227  8.622964 11.315730  5.274939  9.824857  5.755733
#>  [15]  4.950421  1.634301  1.439768  1.070017  1.550871  1.584135  1.001872
#>  [22]  1.105564  1.055753  2.234193  4.011872  3.811058  2.954632  4.512354
#>  [29]  4.771255  6.650637  5.413187  7.007947  6.654386  3.597333 16.842065
#>  [36]  8.885743 19.457558 18.570698 17.028786 17.427567 17.072763 14.119912
#>  [43] 14.125799 14.149521 13.591484 12.827038 10.591303  9.479620 18.047527
#>  [50] 12.971992 13.734568 11.336853 12.824228  6.357805 19.073125 18.586840
#>  [57] 16.450100 14.184309 14.422038 14.709484 13.659490 12.539193 10.975468
#>  [64]  9.876637 18.755040 14.811505  9.776054 15.414289 17.425016 16.568837
#>  [71] 12.651190 11.854114 11.244977 12.918388 10.986565  9.154813 14.684388
#>  [78] 13.592804 12.200514 11.555715 11.654336 11.924969 10.548646  9.389930
#>  [85] 15.712651 16.515634 17.124391 11.444466  8.737769 13.519173  7.432811
#>  [92] 13.213183 14.997826 13.836083 14.633821 11.867591 11.048471 10.965201
#>  [99]  8.502770 17.613834
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.80   NA 0.65   NA 0.50   NA   NA 0.55   NA 0.50 0.65 0.45 0.60
#>  [16] 0.65 0.20 0.35 0.40 0.15 0.10 0.20 0.25 0.10 0.15 0.20 0.30 0.15 0.30 0.40
#>  [31] 0.20 0.35 0.25 0.45   NA 0.55   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [46]   NA   NA 0.30   NA   NA   NA   NA   NA 0.40   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA   NA 0.25   NA   NA 0.35   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76] 0.60   NA   NA   NA   NA   NA   NA   NA 0.45   NA   NA   NA   NA 0.50   NA
#>  [91] 0.50   NA   NA   NA   NA   NA   NA   NA 0.30   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 272.4
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta    gamma
#>  [1,] 0.5696875 0.357463
#>  [2,] 0.5696875 0.357463
#>  [3,] 0.5696875 0.357463
#>  [4,] 0.5696875 0.357463
#>  [5,] 0.5696875 0.357463
#>  [6,] 0.5696875 0.357463
#>  [7,] 0.5696875 0.357463
#>  [8,] 0.5696875 0.357463
#>  [9,] 0.5696875 0.357463
#> [10,] 0.5696875 0.357463
#> [11,] 0.5696875 0.357463
#> [12,] 0.5696875 0.357463
#> [13,] 0.5696875 0.357463
#> [14,] 0.5696875 0.357463
#> [15,] 0.5696875 0.357463
#> [16,] 0.5696875 0.357463
#> [17,] 0.5696875 0.357463
#> [18,] 0.5696875 0.357463
#> [19,] 0.5696875 0.357463
#> [20,] 0.5696875 0.357463