Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 3 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.6569610 0.5323431
#>  [2,] 0.6599226 0.5564885
#>  [3,] 0.2197714 0.1048892
#>  [4,] 0.3866828 0.2914423
#>  [5,] 0.4656262 0.3443504
#>  [6,] 0.3343995 0.2235048
#>  [7,] 0.4675299 0.4861413
#>  [8,] 0.3891111 0.2040878
#>  [9,] 0.3238292 0.2181048
#> [10,] 0.4100064 0.2680025
#> [11,] 0.2578350 0.1218363
#> [12,] 0.2905439 0.2230320
#> [13,] 0.5633807 0.3236886
#> [14,] 0.4032002 0.3372076
#> [15,] 0.2457441 0.1172546
#> [16,] 0.2692893 0.1412850
#> [17,] 0.3517143 0.2776872
#> [18,] 0.2063505 0.1109598
#> [19,] 0.3482362 0.1837804
#> [20,] 0.4825931 0.4130971
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -271.8051     -271.8051 1.528873e-12
#>  [2,]         0      -274.2094     -274.2094 1.381061e-13
#>  [3,]         0      -249.2328     -249.2328 9.714092e-03
#>  [4,]         0      -253.2247     -253.2247 1.793651e-04
#>  [5,]         0      -259.4315     -259.4315 3.615598e-07
#>  [6,]         0      -253.8008     -253.8008 1.008240e-04
#>  [7,]         0      -262.0261     -262.0261 2.699960e-08
#>  [8,]         0      -255.5986     -255.5986 1.670202e-05
#>  [9,]         0      -252.3815     -252.3815 4.168159e-04
#> [10,]         0      -256.6054     -256.6054 6.102395e-06
#> [11,]         0      -249.9117     -249.9117 4.926604e-03
#> [12,]         0      -255.6007     -255.6007 1.666681e-05
#> [13,]         0      -265.3934     -265.3934 9.309357e-10
#> [14,]         0      -269.0164     -269.0164 2.485832e-11
#> [15,]         0      -248.5916     -248.5916 1.844499e-02
#> [16,]         0      -249.1879     -249.1879 1.016023e-02
#> [17,]         0      -253.5382     -253.5382 1.310949e-04
#> [18,]         0      -244.6449     -244.6449 9.547711e-01
#> [19,]         0      -251.3975     -251.3975 1.114990e-03
#> [20,]         0      -274.0677     -274.0677 1.591325e-13
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.036757 11.787513  9.960445 15.506684 11.015293  9.644812 16.611247
#>   [8] 13.476080 10.866365  8.898992 12.360587 11.163609  8.876472 13.707969
#>  [15] 11.083310 10.237255  8.962372 11.059980  7.947548 10.324882  5.888481
#>  [22] 10.495051  9.208710  7.842737 10.889645  6.728939 12.009177  7.847677
#>  [29] 10.454136  9.276742 14.179471 10.471770 10.917125  8.795240 13.629347
#>  [36] 13.411133  9.797465 17.369082 15.508484 12.844118 11.428222  8.729115
#>  [43] 19.189851 13.983710 15.226732 14.495265 14.358046 13.149189 12.632618
#>  [50] 10.916968 10.959243 10.160769 10.104024  7.449730 17.896065 17.167030
#>  [57] 17.302631 15.948992 15.259109 14.437523 14.531466 13.904364 13.508615
#>  [64] 12.738877 12.501779 11.109435 10.044737 11.048340 10.477734  9.899069
#>  [71] 17.341569 17.921302 17.089221 17.067227 14.613738 12.936526 10.710105
#>  [78]  9.639934 18.274710 18.079534 18.837295 18.409382 17.202465 15.374935
#>  [85] 14.929990 16.070307 16.425841 14.169842 14.051472 10.596126  2.052930
#>  [92] 17.211877 17.972653 16.663484 17.064485 15.695938 14.707415 14.533603
#>  [99] 15.350614 14.551557
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.70   NA   NA 0.50   NA   NA   NA 0.55   NA   NA 0.35   NA   NA
#>  [16]   NA 0.70   NA 0.70   NA 0.60   NA 0.55 0.45   NA 0.80   NA 0.55   NA 0.70
#>  [31]   NA   NA   NA 0.50   NA   NA 0.60   NA   NA   NA   NA 0.30   NA   NA   NA
#>  [46]   NA   NA   NA   NA   NA   NA   NA   NA 0.45   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.25   NA   NA   NA   NA   NA
#>  [76]   NA   NA 0.30   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [91] 0.30   NA   NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 145.2
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.2063505 0.1109598
#>  [2,] 0.2063505 0.1109598
#>  [3,] 0.2063505 0.1109598
#>  [4,] 0.2578350 0.1218363
#>  [5,] 0.2063505 0.1109598
#>  [6,] 0.2063505 0.1109598
#>  [7,] 0.3866828 0.2914423
#>  [8,] 0.2063505 0.1109598
#>  [9,] 0.2063505 0.1109598
#> [10,] 0.2063505 0.1109598
#> [11,] 0.2063505 0.1109598
#> [12,] 0.2063505 0.1109598
#> [13,] 0.2063505 0.1109598
#> [14,] 0.2063505 0.1109598
#> [15,] 0.2063505 0.1109598
#> [16,] 0.2063505 0.1109598
#> [17,] 0.2063505 0.1109598
#> [18,] 0.2063505 0.1109598
#> [19,] 0.2063505 0.1109598
#> [20,] 0.2063505 0.1109598