Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4, 0)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 6 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.7294200 0.4105288
#>  [2,] 0.4978866 0.3547948
#>  [3,] 0.7953470 0.5482055
#>  [4,] 0.7715192 0.5339902
#>  [5,] 0.3101521 0.1404563
#>  [6,] 0.4326716 0.3064880
#>  [7,] 0.6674726 0.3734291
#>  [8,] 0.6022310 0.3297706
#>  [9,] 0.7038872 0.3349306
#> [10,] 0.6532491 0.4421937
#> [11,] 0.7295407 0.3047454
#> [12,] 0.8924913 0.5482182
#> [13,] 0.7066647 0.3772646
#> [14,] 0.7831483 0.3699999
#> [15,] 0.7575578 0.3979474
#> [16,] 0.8271948 0.4367486
#> [17,] 0.6387248 0.3596296
#> [18,] 0.6841375 0.4138903
#> [19,] 0.5193893 0.3649521
#> [20,] 0.7592926 0.5837226
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -277.3281     -277.3281 7.609478e-09
#>  [2,]         0      -263.5307     -263.5307 7.472996e-03
#>  [3,]         0      -269.8136     -269.8136 1.395950e-05
#>  [4,]         0      -268.8195     -268.8195 3.772158e-05
#>  [5,]         0      -261.1061     -261.1061 8.442853e-02
#>  [6,]         0      -259.0719     -259.0719 6.455417e-01
#>  [7,]         0      -268.4167     -268.4167 5.643379e-05
#>  [8,]         0      -268.6688     -268.6688 4.385727e-05
#>  [9,]         0      -284.2186     -284.2186 7.742489e-12
#> [10,]         0      -268.2451     -268.2451 6.699995e-05
#> [11,]         0      -322.9248     -322.9248 1.199437e-28
#> [12,]         0      -281.0559     -281.0559 1.829757e-10
#> [13,]         0      -282.9305     -282.9305 2.807262e-11
#> [14,]         0      -301.4602     -301.4602 2.517316e-19
#> [15,]         0      -283.8449     -283.8449 1.125025e-11
#> [16,]         0      -286.3753     -286.3753 8.957865e-13
#> [17,]         0      -272.1682     -272.1682 1.325200e-06
#> [18,]         0      -269.9032     -269.9032 1.276321e-05
#> [19,]         0      -259.9724     -259.9724 2.623233e-01
#> [20,]         0      -273.5174     -273.5174 3.438258e-07
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.134067 11.209646  9.737191 16.466810 11.120278  9.847807 16.104521
#>   [8] 13.809005 11.971037 10.446258  9.698621  8.223175  5.148712  9.559876
#>  [15]  9.428164 11.909255  8.016728  8.805900  9.370459  7.263300  2.969757
#>  [22]  8.396746  3.037304  6.108620  6.694802  5.456209  6.174905  5.071759
#>  [29]  6.904793  7.607803  6.311317  8.241298 12.109276  6.158373 17.054372
#>  [36] 14.059921 14.132934 13.676359 13.153579 12.743662 12.483349 11.956810
#>  [43] 12.073083 11.494545 11.119008 10.309578  8.690029 19.475964 18.952391
#>  [50] 15.037299 15.920379 14.391065 15.914559 10.262379 12.316855 12.735998
#>  [57]  9.799605 18.210529 18.878938 19.154222 17.880492 16.048133 14.671349
#>  [64] 14.029840 14.071234 12.448959 10.557695 11.362020 10.447980  9.212015
#>  [71] 18.762373 18.596948 18.579434 18.708350 17.219466 16.150990 14.732161
#>  [78] 13.746903 13.073037 12.359160 12.992375 13.654167 10.122856  7.642709
#>  [85] 19.149507 18.945733 18.555314 18.801099 17.605243 15.566072  5.817557
#>  [92] 16.598582 17.111899 15.749552 15.985010 10.945355 11.383179 12.128147
#>  [99]  8.022829 19.524858
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.80   NA   NA 0.75   NA   NA   NA   NA 0.55 0.70 0.45 0.80 0.80
#>  [16]   NA 0.45 0.35 0.55 0.35 0.30 0.25 0.55 0.50 0.60 0.40 0.50 0.45 0.25 0.65
#>  [31] 0.60 0.50   NA 0.35   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [46]   NA 0.60   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.60   NA   NA   NA
#>  [61]   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.45   NA   NA   NA   NA   NA
#>  [76]   NA   NA   NA   NA   NA   NA   NA   NA 0.25   NA   NA   NA   NA   NA   NA
#>  [91] 0.45   NA   NA   NA   NA   NA   NA   NA 0.30   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 209.6
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.4326716 0.3064880
#>  [2,] 0.5193893 0.3649521
#>  [3,] 0.4326716 0.3064880
#>  [4,] 0.4326716 0.3064880
#>  [5,] 0.4326716 0.3064880
#>  [6,] 0.4326716 0.3064880
#>  [7,] 0.4326716 0.3064880
#>  [8,] 0.4326716 0.3064880
#>  [9,] 0.4326716 0.3064880
#> [10,] 0.3101521 0.1404563
#> [11,] 0.4326716 0.3064880
#> [12,] 0.4326716 0.3064880
#> [13,] 0.4326716 0.3064880
#> [14,] 0.4326716 0.3064880
#> [15,] 0.4326716 0.3064880
#> [16,] 0.5193893 0.3649521
#> [17,] 0.4326716 0.3064880
#> [18,] 0.5193893 0.3649521
#> [19,] 0.5193893 0.3649521
#> [20,] 0.4326716 0.3064880