Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)
#> Warning: 'initial_time' should be provided. I'm assuming '0' which is one time unit before the first time in your data (1), but this might not be appropriate. This will become an error in a future version of mcstate

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 6 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta      gamma
#>  [1,] 0.4418851 0.28123026
#>  [2,] 0.3239820 0.17526245
#>  [3,] 0.1991936 0.11729941
#>  [4,] 0.2610502 0.13454510
#>  [5,] 0.2270018 0.13457145
#>  [6,] 0.4492536 0.26552855
#>  [7,] 0.3446401 0.18390169
#>  [8,] 0.6200668 0.32067321
#>  [9,] 0.2878217 0.14097331
#> [10,] 0.7624764 0.38740850
#> [11,] 0.6022278 0.36433840
#> [12,] 0.5551875 0.39779537
#> [13,] 0.2100341 0.05587271
#> [14,] 0.3348741 0.08377828
#> [15,] 0.2524407 0.04452127
#> [16,] 0.5894194 0.28066616
#> [17,] 0.3580242 0.21492076
#> [18,] 0.3732186 0.26043912
#> [19,] 0.1459721 0.06106597
#> [20,] 0.2616924 0.08501286
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -258.7682     -258.7682 2.465518e-06
#>  [2,]         0      -250.7326     -250.7326 7.616094e-03
#>  [3,]         0      -246.6116     -246.6116 4.693369e-01
#>  [4,]         0      -252.3305     -252.3305 1.540951e-03
#>  [5,]         0      -246.5194     -246.5194 5.146355e-01
#>  [6,]         0      -264.5839     -264.5839 7.348176e-09
#>  [7,]         0      -253.7712     -253.7712 3.648337e-04
#>  [8,]         0      -272.3785     -272.3785 3.027345e-12
#>  [9,]         0      -251.0659     -251.0659 5.457261e-03
#> [10,]         0      -281.8849     -281.8849 2.251576e-16
#> [11,]         0      -267.7904     -267.7904 2.976063e-10
#> [12,]         0      -262.2583     -262.2583 7.519868e-08
#> [13,]         0      -283.0790     -283.0790 6.821797e-17
#> [14,]         0      -339.4890     -339.4890 2.164436e-41
#> [15,]         0      -343.7167     -343.7167 3.156904e-43
#> [16,]         0      -280.4538     -280.4538 9.418470e-16
#> [17,]         0      -253.5801     -253.5801 4.416701e-04
#> [18,]         0      -254.3804     -254.3804 1.983890e-04
#> [19,]         0      -253.6646     -253.6646 4.058835e-04
#> [20,]         0      -273.3563     -273.3563 1.138651e-12
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.707501 11.210249  8.957282 12.857705  9.314145 15.489856 12.611136
#>   [8] 12.470643 11.019158 10.251319 10.173058  6.533471  8.724749 11.338064
#>  [15]  8.962647 12.253916  9.250176  7.422315  7.223400  6.844648  5.798044
#>  [22] 11.777400  7.930840  7.911735  8.670792  7.609419 10.559394  4.452653
#>  [29]  8.777507 11.569345  6.748967 10.921305  9.765402  5.942803 16.673150
#>  [36] 12.430254 11.096575 10.169770  9.389314 18.663811 16.631027 14.087918
#>  [43] 14.170804 13.439504 12.672523 11.949258 11.760461 11.851287 11.359755
#>  [50] 10.296517 10.308349  9.324673 16.931650 16.653587 17.364600 15.912557
#>  [57] 15.792880 14.747438 14.555634 13.362701 12.405150 10.408873  9.709516
#>  [64] 19.713042 18.844006 18.183263 15.518057 15.354167 14.720344 14.166092
#>  [71] 13.003477 13.156065 12.106025 13.405520 12.265321 10.689874  9.689506
#>  [78] 19.791883 18.560035 17.895789 18.494144 18.726600 17.199672 14.510901
#>  [85] 13.324985 13.585946 14.092594 10.886969  9.648965 17.778481  9.467498
#>  [92] 16.319266 16.362508 15.551296 15.885099 13.065130 12.627112 12.630387
#>  [99]  9.543546 18.175308
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.75   NA 0.80   NA   NA   NA   NA   NA   NA 0.75 0.70   NA 0.55
#>  [16]   NA 0.45 0.50 0.65 0.35 0.55   NA 0.60 0.70 0.45 0.75   NA 0.70 0.45   NA
#>  [31] 0.55   NA 0.65 0.40   NA   NA   NA   NA 0.50   NA   NA   NA   NA   NA   NA
#>  [46]   NA   NA   NA   NA   NA   NA 0.65   NA   NA   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA 0.35   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76]   NA 0.60   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.40   NA
#>  [91] 0.40   NA   NA   NA   NA   NA   NA   NA 0.25   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 201.2
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.2270018 0.1345714
#>  [2,] 0.1991936 0.1172994
#>  [3,] 0.2270018 0.1345714
#>  [4,] 0.2270018 0.1345714
#>  [5,] 0.1991936 0.1172994
#>  [6,] 0.1991936 0.1172994
#>  [7,] 0.2270018 0.1345714
#>  [8,] 0.1991936 0.1172994
#>  [9,] 0.2270018 0.1345714
#> [10,] 0.1991936 0.1172994
#> [11,] 0.1991936 0.1172994
#> [12,] 0.1991936 0.1172994
#> [13,] 0.1991936 0.1172994
#> [14,] 0.2270018 0.1345714
#> [15,] 0.1991936 0.1172994
#> [16,] 0.1991936 0.1172994
#> [17,] 0.1991936 0.1172994
#> [18,] 0.2270018 0.1345714
#> [19,] 0.2270018 0.1345714
#> [20,] 0.1991936 0.1172994