Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4, 0)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 3 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta      gamma
#>  [1,] 0.3513022 0.23989581
#>  [2,] 0.3513022 0.23989581
#>  [3,] 0.2362629 0.11015619
#>  [4,] 0.2768475 0.17529531
#>  [5,] 0.7546645 0.45079520
#>  [6,] 0.3567854 0.20557758
#>  [7,] 0.6560249 0.28025540
#>  [8,] 0.2056105 0.09407039
#>  [9,] 0.2213673 0.09461608
#> [10,] 0.7809097 0.55856628
#> [11,] 0.3264273 0.21089623
#> [12,] 0.6592066 0.33336556
#> [13,] 0.8284439 0.55851111
#> [14,] 0.4155104 0.23725366
#> [15,] 0.7996129 0.76598710
#> [16,] 0.5160616 0.43670680
#> [17,] 0.3287224 0.18505256
#> [18,] 0.3859665 0.25348588
#> [19,] 0.3859665 0.25348588
#> [20,] 0.7462383 0.77570845
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -258.5926     -258.5926 3.996974e-06
#>  [2,]         0      -259.0544     -259.0544 2.518720e-06
#>  [3,]         0      -250.1680     -250.1680 1.821908e-02
#>  [4,]         0      -249.7474     -249.7474 2.774529e-02
#>  [5,]         0      -282.8327     -282.8327 1.186836e-16
#>  [6,]         0      -257.6067     -257.6067 1.071285e-05
#>  [7,]         0      -288.2703     -288.2703 5.162732e-19
#>  [8,]         0      -246.2728     -246.2728 8.956917e-01
#>  [9,]         0      -249.1662     -249.1662 4.961176e-02
#> [10,]         0      -277.6741     -277.6741 2.064159e-14
#> [11,]         0      -253.3147     -253.3147 7.832726e-04
#> [12,]         0      -277.9227     -277.9227 1.609943e-14
#> [13,]         0      -276.1772     -276.1772 9.222504e-14
#> [14,]         0      -261.6832     -261.6832 1.817688e-07
#> [15,]         0      -275.2654     -275.2654 2.295345e-13
#> [16,]         0      -267.4528     -267.4528 5.672841e-10
#> [17,]         0      -251.0013     -251.0013 7.917900e-03
#> [18,]         0      -257.8751     -257.8751 8.191432e-06
#> [19,]         0      -258.2978     -258.2978 5.367420e-06
#> [20,]         0      -283.5267     -283.5267 5.929494e-17
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.208447 11.516647  9.442669 16.192207 13.441174 10.160999  7.814529
#>   [8] 11.289341  7.864565 13.501211  8.659062 10.479627  8.346978 17.426817
#>  [15] 15.708607 13.364903 11.755372 10.273074  8.296187 16.812286 13.060177
#>  [22] 11.224667  9.760602 15.012308 14.361050 11.312589  8.725827 12.028095
#>  [29]  9.578179 16.142473 13.689489 12.433259 11.606991 10.012028  8.456172
#>  [36] 18.384038 17.781391 16.283629 15.081713 14.083611 12.725672 10.397264
#>  [43]  9.887702 16.472642 17.434994 16.804801 14.614294 12.958030 11.980646
#>  [50]  9.296699 18.747935 19.412781 17.898589 15.344375 15.201640 14.842874
#>  [57] 14.313906 13.256223 12.920069 12.411498 12.378160 12.085353 11.833493
#>  [64] 11.616740 11.322755  9.477134 16.493896 18.146121 17.384918 15.078870
#>  [71] 11.390160 11.253766 10.958414 12.714458  9.003162 18.729752 15.172859
#>  [78] 14.562275 12.451460 11.550590 12.269602 13.448524  9.698037 18.619405
#>  [85] 17.626565 18.001076 18.157039 13.070096 10.754451  8.245777 12.704146
#>  [92]  7.352199 19.336146 18.882901 18.540505 16.660496 14.821178 15.317425
#>  [99] 11.351613 10.409544
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.80   NA   NA   NA 0.70   NA 0.60   NA 0.95   NA 0.45   NA   NA
#>  [16]   NA   NA   NA 0.65   NA   NA   NA 0.55   NA   NA   NA 0.30   NA 0.30   NA
#>  [31]   NA   NA   NA   NA 0.45   NA   NA   NA   NA   NA   NA   NA 0.25   NA   NA
#>  [46]   NA   NA   NA   NA 0.30   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA   NA   NA   NA 0.30   NA   NA   NA   NA   NA   NA   NA   NA 0.40
#>  [76]   NA   NA   NA   NA   NA   NA   NA 0.30   NA   NA   NA   NA   NA   NA 0.30
#>  [91]   NA 0.25   NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 155
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta      gamma
#>  [1,] 0.2056105 0.09407039
#>  [2,] 0.2056105 0.09407039
#>  [3,] 0.2056105 0.09407039
#>  [4,] 0.2056105 0.09407039
#>  [5,] 0.2056105 0.09407039
#>  [6,] 0.2056105 0.09407039
#>  [7,] 0.2056105 0.09407039
#>  [8,] 0.2056105 0.09407039
#>  [9,] 0.2056105 0.09407039
#> [10,] 0.2213673 0.09461608
#> [11,] 0.2056105 0.09407039
#> [12,] 0.2056105 0.09407039
#> [13,] 0.2056105 0.09407039
#> [14,] 0.2056105 0.09407039
#> [15,] 0.2056105 0.09407039
#> [16,] 0.2056105 0.09407039
#> [17,] 0.2056105 0.09407039
#> [18,] 0.2056105 0.09407039
#> [19,] 0.2056105 0.09407039
#> [20,] 0.2056105 0.09407039