Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 6 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta      gamma
#>  [1,] 0.2838700 0.15073490
#>  [2,] 0.6822658 0.50013005
#>  [3,] 0.7161032 0.86217943
#>  [4,] 0.5541732 0.30053760
#>  [5,] 0.6618246 0.57956418
#>  [6,] 0.7453775 0.29895423
#>  [7,] 0.7453775 0.29895423
#>  [8,] 0.7633150 0.33760979
#>  [9,] 0.5214001 0.35655529
#> [10,] 0.7590833 0.36294430
#> [11,] 0.2043390 0.08111642
#> [12,] 0.2919712 0.20965735
#> [13,] 0.8889428 0.66808919
#> [14,] 0.5510789 0.31222311
#> [15,] 0.7170711 0.46811997
#> [16,] 0.6077868 0.34482640
#> [17,] 0.5061663 0.36570673
#> [18,] 0.6155670 0.60171388
#> [19,] 0.8466557 0.75472107
#> [20,] 0.1905987 0.13682942
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -247.5124     -247.5124 9.697532e-01
#>  [2,]         0      -274.2474     -274.2474 2.375771e-12
#>  [3,]         0      -287.5487     -287.5487 3.972939e-18
#>  [4,]         0      -265.0234     -265.0234 2.408528e-08
#>  [5,]         0      -272.0519     -272.0519 2.134563e-11
#>  [6,]         0      -392.6894     -392.6894 8.651798e-64
#>  [7,]         0      -390.7471     -390.7471 6.034079e-63
#>  [8,]         0      -311.8421     -311.8421 1.118485e-28
#>  [9,]         0      -268.3882     -268.3882 8.325931e-10
#> [10,]         0      -290.0015     -290.0015 3.418859e-19
#> [11,]         0      -253.4128     -253.4128 2.655545e-03
#> [12,]         0      -251.0810     -251.0810 2.734220e-02
#> [13,]         0      -276.4089     -276.4089 2.735774e-13
#> [14,]         0      -264.9608     -264.9608 2.564174e-08
#> [15,]         0      -270.4443     -270.4443 1.065273e-10
#> [16,]         0      -266.8538     -266.8538 3.861844e-09
#> [17,]         0      -260.6873     -260.6873 1.840352e-06
#> [18,]         0      -275.9622     -275.9622 4.276275e-13
#> [19,]         0      -274.7044     -274.7044 1.504258e-12
#> [20,]         0      -255.7873     -255.7873 2.471304e-04
#> 
#> $statistics
#> $statistics$ess
#>   [1] 15.395717  8.775674 14.225324 10.162643  7.597229 15.113452 12.259736
#>   [8] 11.049315 10.116283  8.889372 11.211255  8.314408 10.678812  8.289028
#>  [15] 12.989783 11.044609  8.843580 11.501726 10.101069  7.920735  4.987493
#>  [22]  4.430444  2.129467  3.817966  6.072177  3.158242  3.680363  4.003572
#>  [29]  7.204311  7.980917  9.452202 10.994986 10.828873  7.128849 17.071641
#>  [36] 15.164526 13.639911 12.784131 11.590972 11.008245 10.773542  9.799536
#>  [43] 19.302527 13.521898 15.199671 14.177788 13.177248 11.815495 10.704039
#>  [50]  6.436972 19.403354 18.738811 18.531689 12.853342 14.051064 13.669082
#>  [57] 12.486577 10.443159 11.273068 12.137116 10.137299  8.313381 18.847900
#>  [64] 17.301185 17.138767 13.200394 11.036907 12.440994 10.411346  8.472786
#>  [71] 18.160605 17.617047 16.532542 17.647159 14.898326 12.764342  9.425915
#>  [78] 19.495705 16.786944 15.419364 15.914997 16.947648 12.079515  8.911821
#>  [85] 19.189805 19.459078 19.516240 13.982750 12.152939  9.865851 10.469510
#>  [92]  8.275679 18.927700 19.007571 18.439676 16.134304 14.267770 13.710799
#>  [99]  8.286423 19.234151
#> 
#> $statistics$acceptance_rate
#>   [1]   NA 0.70   NA   NA 0.70   NA   NA   NA   NA 0.55   NA 0.75   NA 0.60   NA
#>  [16]   NA 0.65   NA   NA 0.65 0.35 0.55 0.60 0.30 0.40 0.50 0.50 0.55 0.60 0.65
#>  [31] 0.60   NA   NA 0.50   NA   NA   NA   NA   NA   NA   NA 0.30   NA   NA   NA
#>  [46]   NA   NA   NA   NA 0.45   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [61]   NA 0.35   NA   NA   NA   NA   NA   NA   NA 0.25   NA   NA   NA   NA   NA
#>  [76]   NA 0.10   NA   NA   NA   NA   NA   NA 0.25   NA   NA   NA   NA   NA 0.35
#>  [91]   NA 0.25   NA   NA   NA   NA   NA   NA 0.30   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 233.2
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>          beta     gamma
#>  [1,] 0.28387 0.1507349
#>  [2,] 0.28387 0.1507349
#>  [3,] 0.28387 0.1507349
#>  [4,] 0.28387 0.1507349
#>  [5,] 0.28387 0.1507349
#>  [6,] 0.28387 0.1507349
#>  [7,] 0.28387 0.1507349
#>  [8,] 0.28387 0.1507349
#>  [9,] 0.28387 0.1507349
#> [10,] 0.28387 0.1507349
#> [11,] 0.28387 0.1507349
#> [12,] 0.28387 0.1507349
#> [13,] 0.28387 0.1507349
#> [14,] 0.28387 0.1507349
#> [15,] 0.28387 0.1507349
#> [16,] 0.28387 0.1507349
#> [17,] 0.28387 0.1507349
#> [18,] 0.28387 0.1507349
#> [19,] 0.28387 0.1507349
#> [20,] 0.28387 0.1507349