Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4, 0)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 3 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.6461590 0.5279180
#>  [2,] 0.6229844 0.3910483
#>  [3,] 0.7546427 0.4185920
#>  [4,] 0.4728749 0.2125778
#>  [5,] 0.4980403 0.3792280
#>  [6,] 0.8791577 0.7139000
#>  [7,] 0.5675444 0.2857376
#>  [8,] 0.4470699 0.2130174
#>  [9,] 0.4303635 0.2172559
#> [10,] 0.4625769 0.1969925
#> [11,] 0.4224140 0.2551989
#> [12,] 0.6686283 0.4977436
#> [13,] 0.3836934 0.3165366
#> [14,] 0.7944169 0.4372227
#> [15,] 0.7438867 0.4815718
#> [16,] 0.8653324 0.4311370
#> [17,] 0.8430423 0.5878276
#> [18,] 0.8352137 0.6311886
#> [19,] 0.8228349 0.6603169
#> [20,] 0.4076082 0.3345655
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -267.6852     -267.6852 4.926874e-05
#>  [2,]         0      -266.3903     -266.3903 1.798477e-04
#>  [3,]         0      -287.5740     -287.5740 1.134936e-13
#>  [4,]         0      -281.7717     -281.7717 3.757108e-11
#>  [5,]         0      -259.4392     -259.4392 1.878232e-01
#>  [6,]         0      -277.7423     -277.7423 2.112600e-09
#>  [7,]         0      -266.4620     -266.4620 1.674098e-04
#>  [8,]         0      -265.7090     -265.7090 3.554818e-04
#>  [9,]         0      -260.7221     -260.7221 5.207218e-02
#> [10,]         0      -278.2743     -278.2743 1.240982e-09
#> [11,]         0      -258.0431     -258.0431 7.586891e-01
#> [12,]         0      -265.1669     -265.1669 6.112911e-04
#> [13,]         0      -271.2399     -271.2399 1.408558e-06
#> [14,]         0      -285.1817     -285.1817 1.241419e-12
#> [15,]         0      -270.7105     -270.7105 2.391514e-06
#> [16,]         0      -315.2847     -315.2847 1.047970e-25
#> [17,]         0      -274.1915     -274.1915 7.360202e-08
#> [18,]         0      -272.1240     -272.1240 5.818243e-07
#> [19,]         0      -273.7094     -273.7094 1.192012e-07
#> [20,]         0      -267.7184     -267.7184 4.765578e-05
#> 
#> $statistics
#> $statistics$ess
#>   [1] 17.078466 12.147351  9.222005 15.816190 11.631713 11.742856 10.179598
#>   [8] 12.103535 12.593925 11.682749  9.743812  8.787973 12.929988 11.436428
#>  [15] 10.845993 10.754866  9.684147 11.706034 10.899881  8.833030 10.385429
#>  [22] 10.474455  7.598490 12.405469 10.883331  8.355660 10.786058  7.192133
#>  [29]  7.884290  9.645351 10.627039  7.096780 14.100683  8.673378 17.373278
#>  [36] 14.142166 12.938766 11.766953 11.109123 10.375218  9.444964 16.079288
#>  [43] 17.756099 14.137654 14.671179 12.915283 14.299579 13.929570 13.367502
#>  [50] 11.567968 12.881589 11.167461 12.016570  9.109217 18.083444 17.342561
#>  [57] 16.422783 14.495997 14.830969 15.005023 13.036183 11.416639 11.029924
#>  [64] 10.641223 10.591442  9.093361 16.568994 17.951550 16.142148 14.409618
#>  [71] 11.700934 12.382267 11.863257 12.612132 10.519150  9.758315 17.751340
#>  [78] 16.720921 15.078629 13.379271 14.488098 15.393825 11.831465  9.385832
#>  [85] 19.548215 19.519645 19.422561 14.487789 13.283791 12.243848  8.952238
#>  [92] 17.120635 18.114168 17.028014 17.365674 15.077640 13.422750 13.555606
#>  [99]  9.329971 19.675713
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.80   NA   NA   NA   NA   NA   NA   NA 0.80 0.60   NA   NA   NA
#>  [16]   NA 0.50   NA   NA 0.70   NA   NA 0.55   NA   NA 0.50   NA 0.35 0.50 0.50
#>  [31]   NA 0.70   NA 0.45   NA   NA   NA   NA   NA   NA 0.45   NA   NA   NA   NA
#>  [46]   NA   NA   NA   NA   NA   NA   NA   NA 0.60   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA   NA   NA   NA 0.50   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76] 0.50   NA   NA   NA   NA   NA   NA   NA 0.45   NA   NA   NA   NA   NA   NA
#>  [91] 0.40   NA   NA   NA   NA   NA   NA   NA 0.40   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 175.2
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.4980403 0.3792280
#>  [2,] 0.4224140 0.2551989
#>  [3,] 0.4224140 0.2551989
#>  [4,] 0.4980403 0.3792280
#>  [5,] 0.4224140 0.2551989
#>  [6,] 0.4224140 0.2551989
#>  [7,] 0.4224140 0.2551989
#>  [8,] 0.4224140 0.2551989
#>  [9,] 0.4224140 0.2551989
#> [10,] 0.4224140 0.2551989
#> [11,] 0.4224140 0.2551989
#> [12,] 0.4980403 0.3792280
#> [13,] 0.4224140 0.2551989
#> [14,] 0.4224140 0.2551989
#> [15,] 0.4224140 0.2551989
#> [16,] 0.4224140 0.2551989
#> [17,] 0.4303635 0.2172559
#> [18,] 0.4224140 0.2551989
#> [19,] 0.4980403 0.3792280
#> [20,] 0.4224140 0.2551989