Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 5 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>             beta       gamma
#>  [1,] 0.23388720 0.055672596
#>  [2,] 0.29859446 0.133219682
#>  [3,] 0.28807712 0.125449156
#>  [4,] 0.13218688 0.038174537
#>  [5,] 0.36743086 0.146642774
#>  [6,] 0.12958452 0.020454883
#>  [7,] 0.16487336 0.026744091
#>  [8,] 0.20249236 0.069712123
#>  [9,] 0.10953090 0.081636860
#> [10,] 0.16468840 0.011262289
#> [11,] 0.14111629 0.015812379
#> [12,] 0.03918494 0.001035505
#> [13,] 0.13128208 0.043266049
#> [14,] 0.27871048 0.106941950
#> [15,] 0.32227490 0.159625911
#> [16,] 0.33821362 0.173455754
#> [17,] 0.21683167 0.135958837
#> [18,] 0.20158635 0.100552265
#> [19,] 0.39502013 0.104020185
#> [20,] 0.39502013 0.104020185
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -293.2694     -293.2694 7.288894e-22
#>  [2,]         0      -256.6647     -256.6647 5.752843e-06
#>  [3,]         0      -256.1804     -256.1804 9.337086e-06
#>  [4,]         0      -262.4277     -262.4277 1.807322e-08
#>  [5,]         0      -267.4858     -267.4858 1.149006e-10
#>  [6,]         0      -289.1663     -289.1663 4.412003e-20
#>  [7,]         0      -298.2268     -298.2268 5.125117e-24
#>  [8,]         0      -257.5296     -257.5296 2.422525e-06
#>  [9,]         0      -286.3747     -286.3747 7.194544e-19
#> [10,]         0      -412.0143     -412.0143 1.960737e-73
#> [11,]         0      -306.6055     -306.6055 1.177274e-27
#> [12,]         0      -468.3120     -468.3120 6.960401e-98
#> [13,]         0      -269.0367     -269.0367 2.436710e-11
#> [14,]         0      -266.8304     -266.8304 2.212865e-10
#> [15,]         0      -250.4146     -250.4146 2.980405e-03
#> [16,]         0      -252.7864     -252.7864 2.781013e-04
#> [17,]         0      -246.5488     -246.5488 1.422880e-01
#> [18,]         0      -244.7562     -244.7562 8.544359e-01
#> [19,]         0      -350.2530     -350.2530 1.303161e-46
#> [20,]         0      -352.0213     -352.0213 2.223575e-47
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.562999 12.011400  9.497425 15.298380 12.302792  9.820250 15.013484
#>   [8] 13.336612 12.241544 10.861279  9.301047  4.547160  6.881059  5.751026
#>  [15]  4.553067  6.135523  4.321723  3.142656  2.009107  3.290470  2.524090
#>  [22]  4.770715  3.410934  4.988127  7.416070  2.546905  2.399316  3.022691
#>  [29]  3.844345  5.849165  3.542820  5.502815  7.915658  4.042499 15.380093
#>  [36]  8.804227 16.386921 13.703710 11.385886 10.893083 10.822354  8.998256
#>  [43] 16.155238  6.608456 16.150654 11.221732 12.673543 12.034203 11.083240
#>  [50] 10.275156 10.360153  9.973651 14.403671 16.939309 14.625798 12.570747
#>  [57] 14.157019 13.301938 13.254054 11.862853 11.955925 11.856094 11.791402
#>  [64] 11.212127 10.961772 10.548883 10.555396 11.219014 10.906934 10.948498
#>  [71] 10.137298  9.986567 17.133845 12.826817 13.604654 14.582709 13.372348
#>  [78] 12.571638 11.942084 10.908468 11.276756 11.100852 11.095239  9.773365
#>  [85] 17.713574 16.465438 14.691310 15.614357 15.373724 14.308529  6.520116
#>  [92] 11.150771 15.314566 14.191593 15.492608 11.089134  8.827871 18.890680
#>  [99] 14.829174 13.550601
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.75   NA   NA 0.75   NA   NA   NA   NA 0.60 0.40 0.65 0.60 0.45
#>  [16] 0.60 0.55 0.60 0.75 0.40 0.40 0.45 0.55 0.40 0.65 0.60 0.65 0.55 0.45 0.35
#>  [31] 0.35 0.45 0.65 0.45   NA 0.35   NA   NA   NA   NA   NA 0.50   NA 0.30   NA
#>  [46]   NA   NA   NA   NA   NA   NA 0.40   NA   NA   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.50   NA   NA   NA
#>  [76]   NA   NA   NA   NA   NA   NA   NA   NA 0.55   NA   NA   NA   NA   NA   NA
#>  [91] 0.30   NA   NA   NA   NA   NA 0.45   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 233.4
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.2168317 0.1359588
#>  [2,] 0.3222749 0.1596259
#>  [3,] 0.2015864 0.1005523
#>  [4,] 0.2015864 0.1005523
#>  [5,] 0.2015864 0.1005523
#>  [6,] 0.2015864 0.1005523
#>  [7,] 0.2015864 0.1005523
#>  [8,] 0.2015864 0.1005523
#>  [9,] 0.2015864 0.1005523
#> [10,] 0.2015864 0.1005523
#> [11,] 0.2015864 0.1005523
#> [12,] 0.2015864 0.1005523
#> [13,] 0.2015864 0.1005523
#> [14,] 0.2168317 0.1359588
#> [15,] 0.2015864 0.1005523
#> [16,] 0.2015864 0.1005523
#> [17,] 0.2015864 0.1005523
#> [18,] 0.2015864 0.1005523
#> [19,] 0.2015864 0.1005523
#> [20,] 0.2015864 0.1005523