Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 5 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta      gamma
#>  [1,] 0.9347916 0.92693591
#>  [2,] 0.9321590 0.81805935
#>  [3,] 0.6581651 0.87271726
#>  [4,] 0.8752852 0.68914781
#>  [5,] 0.8864686 0.83310204
#>  [6,] 0.8923311 0.70635126
#>  [7,] 0.1374828 0.04443101
#>  [8,] 0.3390174 0.23857812
#>  [9,] 0.7538802 0.59159210
#> [10,] 0.9451216 0.86845092
#> [11,] 0.7437442 0.78605934
#> [12,] 0.7592917 0.74749809
#> [13,] 0.9505602 0.54495831
#> [14,] 0.7954475 0.95742481
#> [15,] 0.9601672 0.84463243
#> [16,] 0.9312359 0.97702461
#> [17,] 0.6659760 0.72415820
#> [18,] 0.4308665 0.30472509
#> [19,] 0.5382744 0.38727625
#> [20,] 0.7649870 0.86093729
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -277.8110     -277.8110 9.880623e-11
#>  [2,]         0      -274.5916     -274.5916 2.471409e-09
#>  [3,]         0      -300.3334     -300.3334 1.634704e-20
#>  [4,]         0      -271.6077     -271.6077 4.885036e-08
#>  [5,]         0      -285.1737     -285.1737 6.269007e-14
#>  [6,]         0      -277.0836     -277.0836 2.045123e-10
#>  [7,]         0      -264.8963     -264.8963 4.014122e-05
#>  [8,]         0      -254.9158     -254.9158 8.670997e-01
#>  [9,]         0      -272.2023     -272.2023 2.695476e-08
#> [10,]         0      -281.9634     -281.9634 1.553947e-12
#> [11,]         0      -281.2707     -281.2707 3.106359e-12
#> [12,]         0      -285.1947     -285.1947 6.138782e-14
#> [13,]         0      -298.6773     -298.6773 8.564261e-20
#> [14,]         0      -302.9398     -302.9398 1.206374e-21
#> [15,]         0      -282.5512     -282.5512 8.632613e-13
#> [16,]         0      -286.7992     -286.7992 1.233916e-14
#> [17,]         0      -288.3294     -288.3294 2.671349e-15
#> [18,]         0      -256.7916     -256.7916 1.328568e-01
#> [19,]         0      -267.3959     -267.3959 3.296070e-06
#> [20,]         0      -280.5977     -280.5977 6.088833e-12
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.817909 13.211396 11.506221  9.729999 16.538956 14.283244 12.612312
#>   [8] 10.720595  9.190234 16.283431 10.554356  8.525001 17.491250 13.654786
#>  [15] 10.790841  8.360820 11.860760  8.174307 10.539852  9.326070 14.591831
#>  [22]  8.945685 12.282522  9.184919  7.137523 13.388256  9.273165 16.139542
#>  [29] 11.837273  6.993648 15.345849 11.947941  8.436076 18.776633  6.226271
#>  [36] 13.636198  4.783688  6.187559 10.652387  2.811522  7.120555  5.336889
#>  [43] 14.270353 14.550817  9.293572 17.690957  8.283915 14.083222 10.994744
#>  [50]  6.467146 18.752421 12.700430 12.144680  3.454438 19.125534 18.300246
#>  [57] 13.550948  9.285380 19.647142 19.498298 16.888893 12.568330  9.430402
#>  [64] 19.599420 19.229914 15.096438 11.120940 14.658498 11.018804  9.284838
#>  [71] 18.589300 17.926655 16.164629 16.958165 13.019100  9.657350 18.593031
#>  [78] 17.766257 16.092308 15.919307 16.619994 17.444788 12.934285 11.622860
#>  [85] 10.408356 11.411569 11.785643  9.441912 18.077282 15.078607  8.450788
#>  [92] 12.893182 17.196631 13.754716 16.462302  7.662626 18.730832 18.618273
#>  [99] 11.126472  8.307529
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA   NA 0.85   NA   NA   NA   NA 0.80   NA   NA 0.75   NA   NA   NA
#>  [16] 0.50   NA 0.45   NA 0.55   NA 0.60   NA 0.50 0.65   NA 0.50   NA   NA 0.45
#>  [31]   NA   NA 0.70   NA 0.30   NA 0.60 0.45   NA 0.60 0.40 0.60   NA   NA 0.55
#>  [46]   NA 0.45   NA   NA 0.40   NA   NA   NA 0.50   NA   NA   NA 0.50   NA   NA
#>  [61]   NA   NA 0.45   NA   NA   NA   NA   NA   NA 0.40   NA   NA   NA   NA   NA
#>  [76] 0.40   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.45   NA   NA
#>  [91] 0.20   NA   NA   NA   NA 0.25   NA   NA   NA 0.30
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 282.2
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.4308665 0.3047251
#>  [2,] 0.4308665 0.3047251
#>  [3,] 0.3390174 0.2385781
#>  [4,] 0.3390174 0.2385781
#>  [5,] 0.3390174 0.2385781
#>  [6,] 0.3390174 0.2385781
#>  [7,] 0.3390174 0.2385781
#>  [8,] 0.3390174 0.2385781
#>  [9,] 0.3390174 0.2385781
#> [10,] 0.3390174 0.2385781
#> [11,] 0.3390174 0.2385781
#> [12,] 0.3390174 0.2385781
#> [13,] 0.3390174 0.2385781
#> [14,] 0.3390174 0.2385781
#> [15,] 0.3390174 0.2385781
#> [16,] 0.3390174 0.2385781
#> [17,] 0.3390174 0.2385781
#> [18,] 0.3390174 0.2385781
#> [19,] 0.3390174 0.2385781
#> [20,] 0.3390174 0.2385781