Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 4 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.7430756 0.4081703
#>  [2,] 0.8580454 0.6260028
#>  [3,] 0.7586262 0.5321209
#>  [4,] 0.7102032 0.4012541
#>  [5,] 0.7514589 0.5152573
#>  [6,] 0.8580364 0.4797174
#>  [7,] 0.7901931 0.4322945
#>  [8,] 0.7266088 0.5395045
#>  [9,] 0.8146892 0.6084439
#> [10,] 0.7602237 0.6258859
#> [11,] 0.7733803 0.4680888
#> [12,] 0.7699753 0.4841217
#> [13,] 0.7067089 0.5562002
#> [14,] 0.8122339 0.3891210
#> [15,] 0.6444117 0.4835253
#> [16,] 0.7544105 0.4948171
#> [17,] 0.7884671 0.4777716
#> [18,] 0.8293784 0.6297100
#> [19,] 0.8342070 0.5468938
#> [20,] 0.8363316 0.8456827
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -283.1053     -283.1053 1.938880e-08
#>  [2,]         0      -278.4491     -278.4491 2.040314e-06
#>  [3,]         0      -274.5986     -274.5986 9.593475e-05
#>  [4,]         0      -278.7840     -278.7840 1.459739e-06
#>  [5,]         0      -265.4749     -265.4749 8.796567e-01
#>  [6,]         0      -284.5203     -284.5203 4.710017e-09
#>  [7,]         0      -280.2812     -280.2812 3.266257e-07
#>  [8,]         0      -272.1724     -272.1724 1.085589e-03
#>  [9,]         0      -272.2476     -272.2476 1.006901e-03
#> [10,]         0      -273.2061     -273.2061 3.860964e-04
#> [11,]         0      -272.8861     -272.8861 5.317196e-04
#> [12,]         0      -271.3101     -271.3101 2.571126e-03
#> [13,]         0      -267.5538     -267.5538 1.100165e-01
#> [14,]         0      -303.1499     -303.1499 3.821862e-17
#> [15,]         0      -270.8285     -270.8285 4.162118e-03
#> [16,]         0      -274.0824     -274.0824 1.607418e-04
#> [17,]         0      -278.6408     -278.6408 1.684441e-06
#> [18,]         0      -273.4281     -273.4281 3.092346e-04
#> [19,]         0      -277.0451     -277.0451 8.307440e-06
#> [20,]         0      -277.9048     -277.9048 3.516468e-06
#> 
#> $statistics
#> $statistics$ess
#>   [1] 17.659590 13.369420 11.642229 10.686332 10.092968  8.864485 15.156703
#>   [8] 13.107327 11.015527  9.234509  9.889515  4.721903  7.665092  8.630003
#>  [15]  7.256911  4.559025  3.096227  2.303203  3.589189  3.309333  2.842562
#>  [22]  7.341813  4.283384  4.516588  5.450717  2.281586  5.984785  5.378704
#>  [29]  7.235952  9.446705  9.225688 10.314917  9.177165 13.138215 14.434084
#>  [36] 13.303910 13.818242 13.385798 12.769022 12.445334 12.496561 13.376060
#>  [43] 13.199693 12.141172 12.031628 12.171272 10.168351  9.423370 19.757665
#>  [50] 16.023539 17.296471 15.693920 16.728484 11.673103 12.687414 13.481554
#>  [57] 11.294329 10.119933 10.227040 10.967034 10.440891  8.952650 19.535536
#>  [64] 19.025448 19.088881 17.057201 14.355162 16.020220 14.781163 12.306784
#>  [71] 10.211380  9.713670 19.534045 19.043496 18.429522 17.637379 15.900727
#>  [78] 15.185895 13.250494 12.625462 12.383491 12.750351  9.982172 18.997236
#>  [85] 17.624627 18.043607 18.044055 16.725450 16.033353 14.727339 10.933343
#>  [92]  8.875574 19.601127 19.390690 19.278728 17.096607 16.052918 15.550445
#>  [99] 10.281665 10.367241
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA   NA   NA   NA 0.65   NA   NA   NA 0.55 0.50 0.50 0.40 0.50 0.70
#>  [16] 0.60 0.40 0.55 0.40 0.35 0.50 0.40 0.30 0.45 0.25 0.45 0.40 0.65 0.60 0.35
#>  [31] 0.50   NA 0.55   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [46]   NA   NA 0.50   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [61]   NA 0.35   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.55   NA   NA   NA
#>  [76]   NA   NA   NA   NA   NA   NA   NA 0.50   NA   NA   NA   NA   NA   NA   NA
#>  [91]   NA 0.55   NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 189.4
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.7514589 0.5152573
#>  [2,] 0.7514589 0.5152573
#>  [3,] 0.7514589 0.5152573
#>  [4,] 0.7514589 0.5152573
#>  [5,] 0.7514589 0.5152573
#>  [6,] 0.7067089 0.5562002
#>  [7,] 0.7514589 0.5152573
#>  [8,] 0.7514589 0.5152573
#>  [9,] 0.7514589 0.5152573
#> [10,] 0.7514589 0.5152573
#> [11,] 0.7514589 0.5152573
#> [12,] 0.7514589 0.5152573
#> [13,] 0.7514589 0.5152573
#> [14,] 0.7514589 0.5152573
#> [15,] 0.7514589 0.5152573
#> [16,] 0.7067089 0.5562002
#> [17,] 0.7514589 0.5152573
#> [18,] 0.7514589 0.5152573
#> [19,] 0.7514589 0.5152573
#> [20,] 0.7514589 0.5152573