Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 5 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta      gamma
#>  [1,] 0.1258925 0.03765022
#>  [2,] 0.2888811 0.17287626
#>  [3,] 0.2820228 0.11147064
#>  [4,] 0.1753075 0.07180743
#>  [5,] 0.1315252 0.06949941
#>  [6,] 0.1708548 0.08801029
#>  [7,] 0.1545015 0.04375641
#>  [8,] 0.1432555 0.02453426
#>  [9,] 0.1342194 0.04201007
#> [10,] 0.1818885 0.03486860
#> [11,] 0.1062452 0.02178842
#> [12,] 0.1038991 0.02394936
#> [13,] 0.2063068 0.10298559
#> [14,] 0.1194532 0.04539881
#> [15,] 0.1687290 0.06571409
#> [16,] 0.1300091 0.02505222
#> [17,] 0.1696455 0.04961134
#> [18,] 0.1553965 0.07255072
#> [19,] 0.1891617 0.05210490
#> [20,] 0.1612732 0.07441230
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -259.7851     -259.7851 8.270861e-08
#>  [2,]         0      -247.9629     -247.9629 1.126828e-02
#>  [3,]         0      -267.7726     -267.7726 2.809442e-11
#>  [4,]         0      -250.7546     -250.7546 6.909994e-04
#>  [5,]         0      -256.7141     -256.7141 1.783481e-06
#>  [6,]         0      -245.4422     -245.4422 1.401513e-01
#>  [7,]         0      -267.1860     -267.1860 5.050959e-11
#>  [8,]         0      -281.8459     -281.8459 2.171066e-17
#>  [9,]         0      -261.6059     -261.6059 1.339041e-08
#> [10,]         0      -288.7536     -288.7536 2.171244e-20
#> [11,]         0      -288.5940     -288.5940 2.546921e-20
#> [12,]         0      -297.1845     -297.1845 4.733549e-24
#> [13,]         0      -243.6761     -243.6761 8.195803e-01
#> [14,]         0      -260.5483     -260.5483 3.855764e-08
#> [15,]         0      -248.9960     -248.9960 4.010707e-03
#> [16,]         0      -273.4997     -273.4997 9.149087e-14
#> [17,]         0      -266.9188     -266.9188 6.598320e-11
#> [18,]         0      -247.8175     -247.8175 1.303185e-02
#> [19,]         0      -276.6127     -276.6127 4.068410e-15
#> [20,]         0      -247.9633     -247.9633 1.126457e-02
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.372576 10.280562  8.106656 13.375853 10.567672  9.964159  8.030180
#>   [8]  7.492343  6.055786  7.199830  6.178175  1.693066  1.815350  1.271532
#>  [15]  1.110237  1.057466  1.108257  1.764026  1.000000  1.000016  1.049498
#>  [22]  2.025300  1.064815  1.164829  1.193399  3.828357  2.733867  1.335440
#>  [29]  2.322615  1.884151  1.835233  2.539284  1.845452  1.542716  6.633714
#>  [36]  4.168909 10.827272  7.061925 10.867620  9.159377 16.539651 15.069280
#>  [43] 11.323854  3.531190 14.322240  7.730473 18.874427 15.260539 11.770717
#>  [50] 11.840810  8.660425 18.344277  7.849712 18.398076 15.980557  9.171127
#>  [57] 19.363210 18.355499 11.855021  7.215275 18.910440 17.283893 15.219072
#>  [64] 13.177014 11.216732 11.511113 12.366137  9.100924 19.162958 18.006680
#>  [71] 17.588042 15.301598 11.964188  8.000717 19.410830 18.374225 16.730894
#>  [78] 16.334635 15.777334 15.142854 14.093250 13.156507 13.450838 13.506876
#>  [85] 13.056970 13.148739 12.570267 12.844262 12.401861 11.353276 12.955375
#>  [92] 13.752302 13.462055 12.920104 11.806904 12.100580 11.869994 11.719438
#>  [99] 10.107835 10.477176
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.70   NA   NA 0.65 0.70 0.45 0.55 0.45 0.30 0.45 0.35 0.45 0.55
#>  [16] 0.25 0.40 0.50 0.55 0.60 0.45 0.40 0.55 0.45 0.50 0.40 0.35 0.65 0.50 0.80
#>  [31] 0.45 0.40 0.55 0.45 0.50 0.45   NA 0.40   NA 0.45   NA   NA   NA 0.35   NA
#>  [46] 0.70   NA   NA   NA   NA 0.60   NA 0.40   NA   NA 0.50   NA   NA   NA 0.35
#>  [61]   NA   NA   NA   NA   NA   NA   NA 0.55   NA   NA   NA   NA   NA 0.60   NA
#>  [76]   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [91]   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 256.8
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta      gamma
#>  [1,] 0.2063068 0.10298559
#>  [2,] 0.2063068 0.10298559
#>  [3,] 0.2063068 0.10298559
#>  [4,] 0.2888811 0.17287626
#>  [5,] 0.2063068 0.10298559
#>  [6,] 0.2063068 0.10298559
#>  [7,] 0.2063068 0.10298559
#>  [8,] 0.2888811 0.17287626
#>  [9,] 0.2063068 0.10298559
#> [10,] 0.2063068 0.10298559
#> [11,] 0.2063068 0.10298559
#> [12,] 0.2063068 0.10298559
#> [13,] 0.2063068 0.10298559
#> [14,] 0.2063068 0.10298559
#> [15,] 0.1708548 0.08801029
#> [16,] 0.2063068 0.10298559
#> [17,] 0.2063068 0.10298559
#> [18,] 0.2063068 0.10298559
#> [19,] 0.2063068 0.10298559
#> [20,] 0.2063068 0.10298559