Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 6 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.5602810 0.1567041
#>  [2,] 0.5602810 0.1567041
#>  [3,] 0.4466340 0.2819846
#>  [4,] 0.4927105 0.2641582
#>  [5,] 0.7911999 0.4241131
#>  [6,] 0.8862471 0.3328685
#>  [7,] 0.8056242 0.3179150
#>  [8,] 0.9600979 0.5716144
#>  [9,] 0.8017928 0.2596612
#> [10,] 0.7409148 0.2394125
#> [11,] 0.9297048 0.3639087
#> [12,] 0.8017928 0.2596612
#> [13,] 0.7616927 0.3042058
#> [14,] 0.7959200 0.2946272
#> [15,] 0.7959200 0.2946272
#> [16,] 0.6769285 0.2468760
#> [17,] 0.6096826 0.5753850
#> [18,] 0.7522796 0.4636340
#> [19,] 0.5517148 0.3377767
#> [20,] 0.8429801 0.6659662
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior        weight
#>  [1,]         0      -453.7102     -453.7102  2.649441e-82
#>  [2,]         0      -446.2500     -446.2500  4.603351e-79
#>  [3,]         0      -268.5246     -268.5246  7.050910e-02
#>  [4,]         0      -267.3928     -267.3928  2.186687e-01
#>  [5,]         0      -301.5724     -301.5724  3.131502e-16
#>  [6,]         0      -453.1344     -453.1344  4.712036e-82
#>  [7,]         0      -457.2363     -457.2363  7.794182e-84
#>  [8,]         0      -291.3776     -291.3776  8.381297e-12
#>  [9,]         0      -695.5525     -695.5525 2.468100e-187
#> [10,]         0      -552.5354     -552.5354 3.190898e-125
#> [11,]         0      -544.0928     -544.0928 1.480771e-121
#> [12,]         0      -679.8144     -679.8144 1.687899e-180
#> [13,]         0      -407.2919     -407.2919  3.822671e-62
#> [14,]         0      -457.6581     -457.6581  5.112255e-84
#> [15,]         0      -459.1520     -459.1520  1.147637e-84
#> [16,]         0      -432.3374     -432.3374  5.072801e-73
#> [17,]         0      -267.8256     -267.8256  1.418461e-01
#> [18,]         0      -280.5689     -280.5689  4.144438e-07
#> [19,]         0      -266.4390     -266.4390  5.675640e-01
#> [20,]         0      -272.4355     -272.4355  1.411687e-03
#> 
#> $statistics
#> $statistics$ess
#>   [1] 14.443363 10.049069  8.869172 16.400075 12.579791 11.232548  9.214649
#>   [8] 12.830325 10.235040  8.804012 10.062934  8.782916 11.702382  9.044320
#>  [15] 11.857969 10.283915  9.266750  9.098214 10.969251  6.941363  4.566524
#>  [22]  9.472221  4.969873  6.640764  5.740806  6.016167  6.701089  7.346616
#>  [29]  8.537421  8.053867  8.196376  8.014953 10.983648  4.428173 17.846816
#>  [36] 13.705789 13.592634 12.911448 12.043587 11.606932 11.830962 11.003507
#>  [43] 11.372551 10.174728 10.036037  9.239376 15.822380 14.896047 14.371713
#>  [50] 10.499521 11.708588 10.310518 10.608276  8.727086 19.000420 18.661473
#>  [57] 16.499828 13.971384 14.244847 14.287865 12.331359 11.011735  9.628259
#>  [64] 19.243998 18.608930 16.361057 12.218379 14.181548 11.904623 10.637579
#>  [71]  9.024352 19.284037 17.909729 18.494970 16.051345 14.338343 11.940004
#>  [78] 11.529470 10.504637 10.453115 11.199709 11.648652  9.115528 15.704297
#>  [85] 13.317690 14.189428 14.430617 12.170902 11.059977 10.007725  6.954846
#>  [92] 11.465699 12.032510 10.329812 10.657860  9.133527 15.596855 16.477333
#>  [99] 10.288675  9.108682
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.80   NA   NA   NA 0.85   NA   NA 0.45   NA 0.75   NA 0.70   NA
#>  [16]   NA 0.65 0.45   NA 0.35 0.30 0.35 0.55 0.45 0.35 0.35 0.40 0.35 0.20 0.50
#>  [31] 0.35 0.45   NA 0.60   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [46] 0.55   NA   NA   NA   NA   NA   NA   NA 0.45   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA 0.35   NA   NA   NA   NA   NA   NA   NA 0.20   NA   NA   NA   NA
#>  [76]   NA   NA   NA   NA   NA   NA   NA 0.50   NA   NA   NA   NA   NA   NA   NA
#>  [91] 0.20   NA   NA   NA   NA 0.15   NA   NA   NA 0.40
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 231.4
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.5517148 0.3377767
#>  [2,] 0.5517148 0.3377767
#>  [3,] 0.5517148 0.3377767
#>  [4,] 0.5517148 0.3377767
#>  [5,] 0.5517148 0.3377767
#>  [6,] 0.6096826 0.5753850
#>  [7,] 0.5517148 0.3377767
#>  [8,] 0.5517148 0.3377767
#>  [9,] 0.5517148 0.3377767
#> [10,] 0.5517148 0.3377767
#> [11,] 0.4927105 0.2641582
#> [12,] 0.5517148 0.3377767
#> [13,] 0.4927105 0.2641582
#> [14,] 0.5517148 0.3377767
#> [15,] 0.5517148 0.3377767
#> [16,] 0.5517148 0.3377767
#> [17,] 0.4927105 0.2641582
#> [18,] 0.5517148 0.3377767
#> [19,] 0.4927105 0.2641582
#> [20,] 0.4927105 0.2641582