Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars
A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter
A particle_filter object

control
A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 6 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>             beta      gamma
#>  [1,] 0.53449304 0.15590641
#>  [2,] 0.49258869 0.15782694
#>  [3,] 0.64810577 0.19872710
#>  [4,] 0.64810577 0.19872710
#>  [5,] 0.64810577 0.19872710
#>  [6,] 0.25670449 0.11606310
#>  [7,] 0.20986742 0.07413075
#>  [8,] 0.21549173 0.10597094
#>  [9,] 0.12161279 0.01638613
#> [10,] 0.19031429 0.06484157
#> [11,] 0.13986388 0.03515864
#> [12,] 0.14814767 0.04751126
#> [13,] 0.27246339 0.07905619
#> [14,] 0.14611066 0.01355994
#> [15,] 0.16118551 0.04624763
#> [16,] 0.16380769 0.06257263
#> [17,] 0.19194894 0.05291276
#> [18,] 0.09933406 0.02855456
#> [19,] 0.24989225 0.11815328
#> [20,] 0.16989496 0.08659757
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -426.1232     -426.1232 1.030607e-79
#>  [2,]         0      -363.6450     -363.6450 1.402876e-52
#>  [3,]         0      -454.1331     -454.1331 7.056168e-92
#>  [4,]         0      -442.6456     -442.6456 6.878940e-87
#>  [5,]         0      -424.1512     -424.1512 7.405338e-79
#>  [6,]         0      -246.5634     -246.5634 9.884340e-02
#>  [7,]         0      -258.2798     -258.2798 8.063950e-07
#>  [8,]         0      -244.4114     -244.4114 8.501985e-01
#>  [9,]         0      -281.0422     -281.0422 1.049490e-16
#> [10,]         0      -257.6858     -257.6858 1.460544e-06
#> [11,]         0      -270.9079     -270.9079 2.643972e-12
#> [12,]         0      -257.0059     -257.0059 2.882846e-06
#> [13,]         0      -291.2678     -291.2678 3.802268e-21
#> [14,]         0      -288.8409     -288.8409 4.305781e-20
#> [15,]         0      -271.6560     -271.6560 1.251233e-12
#> [16,]         0      -252.5739     -252.5739 2.424337e-04
#> [17,]         0      -277.5860     -277.5860 3.326677e-15
#> [18,]         0      -278.6644     -278.6644 1.131517e-15
#> [19,]         0      -247.5301     -247.5301 3.759387e-02
#> [20,]         0      -248.5830     -248.5830 1.311663e-02
#> 
#> $statistics
#> $statistics$ess
#>   [1] 15.802789 11.063093  9.636482 14.421550 12.805213 11.567877 10.600346
#>   [8] 10.563826 10.386998  9.846043 10.674595  7.683640  9.541610  5.659775
#>  [15]  4.922349  4.221245  3.764605  1.594171  3.905487  2.542549  3.059956
#>  [22]  6.654848  4.996842  6.726930  7.161412  5.310174  4.201296  1.883427
#>  [29]  3.100887  5.545400  3.025752  5.231645  5.537049  2.149514 15.882553
#>  [36]  7.125690 16.602494 12.632937  9.737791 19.510321 18.908707 17.550165
#>  [43] 17.494460 12.907152 11.438165  7.902885 16.541938 15.556641 13.118412
#>  [50] 12.268403  8.603064 17.116458 11.345517 12.097185  9.492641 13.835336
#>  [57] 11.204073 10.049341  7.382858 12.935249 10.850657  9.344836 16.356780
#>  [64] 12.957705  9.936239 15.931290 13.203205  9.826123 16.446845 12.394283
#>  [71] 10.637337  8.882786 16.636348  8.822380 16.252800 12.931603 12.067957
#>  [78] 10.435602  9.925891 17.382547 13.726384  9.510890 16.338923 13.828998
#>  [85] 13.771422 13.160940 12.319997 11.705724 11.178724 10.839634  5.587768
#>  [92] 11.215093 15.795475 15.145229 14.967774 13.516220 12.733710 13.180550
#>  [99] 10.045474  9.938452
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.90   NA   NA   NA   NA   NA   NA 0.70   NA 0.75 0.50 0.45 0.55
#>  [16] 0.55 0.55 0.35 0.40 0.70 0.60 0.55 0.60 0.50 0.55 0.80 0.60 0.70 0.45 0.40
#>  [31] 0.40 0.50 0.25 0.50   NA 0.55   NA   NA 0.40   NA   NA   NA   NA   NA   NA
#>  [46] 0.30   NA   NA   NA   NA 0.30   NA   NA   NA 0.15   NA   NA   NA 0.45   NA
#>  [61]   NA 0.20   NA   NA 0.35   NA   NA 0.40   NA   NA   NA 0.30   NA 0.10   NA
#>  [76]   NA   NA   NA 0.40   NA   NA 0.20   NA   NA   NA   NA   NA   NA   NA   NA
#>  [91] 0.30   NA   NA   NA   NA   NA   NA   NA   NA 0.25
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 324.2
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.2154917 0.1059709
#>  [2,] 0.2154917 0.1059709
#>  [3,] 0.2154917 0.1059709
#>  [4,] 0.2567045 0.1160631
#>  [5,] 0.2154917 0.1059709
#>  [6,] 0.2154917 0.1059709
#>  [7,] 0.2154917 0.1059709
#>  [8,] 0.2154917 0.1059709
#>  [9,] 0.2567045 0.1160631
#> [10,] 0.2154917 0.1059709
#> [11,] 0.2567045 0.1160631
#> [12,] 0.2154917 0.1059709
#> [13,] 0.2154917 0.1059709
#> [14,] 0.2498922 0.1181533
#> [15,] 0.2154917 0.1059709
#> [16,] 0.2154917 0.1059709
#> [17,] 0.2567045 0.1160631
#> [18,] 0.2154917 0.1059709
#> [19,] 0.2154917 0.1059709
#> [20,] 0.2154917 0.1059709