Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 4 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta      gamma
#>  [1,] 0.3448765 0.20844541
#>  [2,] 0.2612363 0.18521201
#>  [3,] 0.2992830 0.24343941
#>  [4,] 0.2704340 0.22177796
#>  [5,] 0.8280062 0.90345932
#>  [6,] 0.9779795 0.89344217
#>  [7,] 0.2872151 0.34621282
#>  [8,] 0.2428995 0.13975823
#>  [9,] 0.4685683 0.26906649
#> [10,] 0.1855952 0.10091031
#> [11,] 0.3547067 0.19059536
#> [12,] 0.2098122 0.10755919
#> [13,] 0.3386358 0.18170566
#> [14,] 0.2216313 0.12506220
#> [15,] 0.1487188 0.09195377
#> [16,] 0.7131663 0.57644683
#> [17,] 0.8374169 0.82357113
#> [18,] 0.8767720 0.76341089
#> [19,] 0.7051786 0.62010536
#> [20,] 0.7975978 0.75982710
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -252.3916     -252.3916 8.923420e-06
#>  [2,]         0      -255.4016     -255.4016 4.398463e-07
#>  [3,]         0      -265.2229     -265.2229 2.387666e-11
#>  [4,]         0      -269.2666     -269.2666 4.185985e-13
#>  [5,]         0      -282.3967     -282.3967 8.307336e-19
#>  [6,]         0      -280.4881     -280.4881 5.602072e-18
#>  [7,]         0      -299.4365     -299.4365 3.304899e-26
#>  [8,]         0      -248.1085     -248.1085 6.466435e-04
#>  [9,]         0      -257.3618     -257.3618 6.193998e-08
#> [10,]         0      -240.7988     -240.7988 9.664854e-01
#> [11,]         0      -251.5257     -251.5257 2.121221e-05
#> [12,]         0      -244.1815     -244.1815 3.281886e-02
#> [13,]         0      -252.3782     -252.3782 9.043936e-06
#> [14,]         0      -252.4594     -252.4594 8.338108e-06
#> [15,]         0      -254.5420     -254.5420 1.039007e-06
#> [16,]         0      -266.9784     -266.9784 4.126067e-12
#> [17,]         0      -281.2111     -281.2111 2.718732e-18
#> [18,]         0      -268.4430     -268.4430 9.538402e-13
#> [19,]         0      -270.2045     -270.2045 1.638653e-13
#> [20,]         0      -270.1183     -270.1183 1.786158e-13
#> 
#> $statistics
#> $statistics$ess
#>   [1] 14.189379  9.271280 12.567617  9.542120 11.357804  9.626751 13.957411
#>   [8] 11.025514  9.272585 14.142231 10.510184  9.455581 13.460765 10.920916
#>  [15]  9.481157 17.533970 14.854564 11.404179  9.839928 15.528062 11.288765
#>  [22] 10.633868  9.085862 13.657719 13.021755 11.497944 10.278401  9.120229
#>  [29] 11.582002 10.467569  9.499046 14.371430 12.473493 11.170428 10.002732
#>  [36]  8.831135 15.804792 13.581670 12.492487 11.479335 10.671644  9.759955
#>  [43] 17.631396 12.783875 13.725771 12.552743  9.911585 17.128314 15.443421
#>  [50] 11.338158 12.509753 10.909548 11.709559  8.745134 18.342594 18.342293
#>  [57] 16.224718 13.812081 14.048570 14.286703 12.875602 12.003577 11.201411
#>  [64] 10.936421 10.600423  9.635123 15.232280 18.565195 16.977546 14.709703
#>  [71] 11.665704 11.427139 11.367395 13.203203 10.610216  9.141755 15.645083
#>  [78] 15.236839 12.145184 11.042425 12.590443 12.220947  7.630407 16.677108
#>  [85] 14.744939 16.542838 17.670160 10.118127  8.876065 17.869399 11.768279
#>  [92]  8.282977 19.252214 18.454940 18.192904 16.731400 15.264291 16.270355
#>  [99] 10.739214  9.948110
#> 
#> $statistics$acceptance_rate
#>   [1]   NA 0.85   NA 0.75   NA 0.45   NA   NA 0.70   NA   NA 0.80   NA   NA 0.70
#>  [16]   NA   NA   NA 0.70   NA   NA   NA 0.55   NA   NA   NA   NA 0.55   NA   NA
#>  [31] 0.65   NA   NA   NA   NA 0.65   NA   NA   NA   NA   NA 0.35   NA   NA   NA
#>  [46]   NA 0.50   NA   NA   NA   NA   NA   NA 0.35   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA   NA   NA   NA 0.20   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76] 0.35   NA   NA   NA   NA   NA   NA 0.50   NA   NA   NA   NA   NA 0.40   NA
#>  [91]   NA 0.25   NA   NA   NA   NA   NA   NA   NA 0.20
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 186.8
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.1855952 0.1009103
#>  [2,] 0.1855952 0.1009103
#>  [3,] 0.1855952 0.1009103
#>  [4,] 0.1855952 0.1009103
#>  [5,] 0.1855952 0.1009103
#>  [6,] 0.1855952 0.1009103
#>  [7,] 0.1855952 0.1009103
#>  [8,] 0.1855952 0.1009103
#>  [9,] 0.1855952 0.1009103
#> [10,] 0.1855952 0.1009103
#> [11,] 0.1855952 0.1009103
#> [12,] 0.1855952 0.1009103
#> [13,] 0.1855952 0.1009103
#> [14,] 0.1855952 0.1009103
#> [15,] 0.1855952 0.1009103
#> [16,] 0.1855952 0.1009103
#> [17,] 0.1855952 0.1009103
#> [18,] 0.1855952 0.1009103
#> [19,] 0.1855952 0.1009103
#> [20,] 0.1855952 0.1009103