Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars
A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter
A particle_filter object

control
A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)

#> Finished 100 steps in 5 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res

#> $pars
#>            beta      gamma
#>  [1,] 0.1720582 0.06592886
#>  [2,] 0.2864904 0.11776156
#>  [3,] 0.2452870 0.14133770
#>  [4,] 0.1719065 0.08753887
#>  [5,] 0.1966343 0.06952398
#>  [6,] 0.1062925 0.02161079
#>  [7,] 0.2473570 0.06902634
#>  [8,] 0.1384636 0.03493597
#>  [9,] 0.1539858 0.05458405
#> [10,] 0.2032423 0.05489765
#> [11,] 0.3853402 0.14629138
#> [12,] 0.2509530 0.08857231
#> [13,] 0.2134426 0.03081528
#> [14,] 0.6034143 0.18363509
#> [15,] 0.2685695 0.12998487
#> [16,] 0.2244792 0.10983304
#> [17,] 0.1129770 0.05581241
#> [18,] 0.1387729 0.04772492
#> [19,] 0.1752658 0.07467236
#> [20,] 0.2196667 0.04602956
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -255.0983     -255.0983 2.679283e-05
#>  [2,]         0      -259.2762     -259.2762 4.107607e-07
#>  [3,]         0      -244.7154     -244.7154 8.655329e-01
#>  [4,]         0      -250.2389     -250.2389 3.455083e-03
#>  [5,]         0      -255.8384     -255.8384 1.278188e-05
#>  [6,]         0      -299.8159     -299.8159 1.017225e-24
#>  [7,]         0      -283.2269     -283.2269 1.629060e-17
#>  [8,]         0      -267.0829     -267.0829 1.671785e-10
#>  [9,]         0      -254.8873     -254.8873 3.308726e-05
#> [10,]         0      -277.4470     -277.4470 5.273711e-15
#> [11,]         0      -283.7963     -283.7963 9.218661e-18
#> [12,]         0      -261.0035     -261.0035 7.301880e-08
#> [13,]         0      -336.5325     -336.5325 1.152466e-40
#> [14,]         0      -379.2537     -379.2537 3.221379e-59
#> [15,]         0      -246.8719     -246.8719 1.001665e-01
#> [16,]         0      -251.0053     -251.0053 1.605499e-03
#> [17,]         0      -266.1529     -266.1529 4.237112e-10
#> [18,]         0      -259.4875     -259.4875 3.325346e-07
#> [19,]         0      -248.1057     -248.1057 2.916653e-02
#> [20,]         0      -306.0430     -306.0430 2.009198e-27
#> 
#> $statistics
#> $statistics$ess
#>   [1] 15.474439 10.857070  9.643540 16.572875 13.289478 12.336846 11.393570
#>   [8] 10.263624  9.467652 10.070968  8.395634  5.650497  4.871045  6.506163
#>  [15]  2.947299  3.089128  2.791894  2.988008  2.088045  1.014332  1.705373
#>  [22]  1.800052  1.352154  1.784680  3.357645  3.507483  3.533213  4.147229
#>  [29]  4.754798  7.314291  6.979115  5.669050  9.053186  6.686989 15.306337
#>  [36] 10.075869  9.032352 16.885789 14.072912 13.162275 12.535458 12.077404
#>  [43] 11.313261 11.321196 10.747069  9.561539 17.962191 18.089835 16.947012
#>  [50] 16.589036 14.233233 13.867378 11.068070 12.572094 10.539479  9.559392
#>  [57] 18.478528 15.774979 12.143925 11.499679 11.033838 10.479213  9.768146
#>  [64] 17.053905 13.667881 14.241819 13.146077  9.468175 18.578165 16.925325
#>  [71] 15.624214 14.069685 13.222896 11.874716 12.333434 11.774823 10.798028
#>  [78] 10.682089 10.147204  9.804613 15.218184 11.397984 14.031339 14.577828
#>  [85] 14.280234 13.777314 13.520845 13.344705 12.972840 13.143762 10.963391
#>  [92]  6.641130 18.468603 18.970394 17.285812 18.024057 17.300812 16.908049
#>  [99] 12.861805 11.407750
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.85   NA   NA   NA   NA   NA 0.70   NA 0.70 0.55 0.45 0.55 0.50
#>  [16] 0.50 0.65 0.45 0.35 0.45 0.25 0.30 0.45 0.50 0.65 0.60 0.50 0.40 0.70 0.55
#>  [31] 0.60 0.70 0.45 0.45   NA   NA 0.35   NA   NA   NA   NA   NA   NA   NA   NA
#>  [46] 0.30   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.20   NA   NA   NA   NA
#>  [61]   NA   NA 0.35   NA   NA   NA   NA 0.40   NA   NA   NA   NA   NA   NA   NA
#>  [76]   NA   NA   NA   NA 0.20   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [91]   NA 0.40   NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 218.8
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)

#>            beta      gamma
#>  [1,] 0.2452870 0.14133770
#>  [2,] 0.1752658 0.07467236
#>  [3,] 0.2452870 0.14133770
#>  [4,] 0.2452870 0.14133770
#>  [5,] 0.2452870 0.14133770
#>  [6,] 0.2452870 0.14133770
#>  [7,] 0.2452870 0.14133770
#>  [8,] 0.2452870 0.14133770
#>  [9,] 0.2452870 0.14133770
#> [10,] 0.2452870 0.14133770
#> [11,] 0.2685695 0.12998487
#> [12,] 0.2452870 0.14133770
#> [13,] 0.2452870 0.14133770
#> [14,] 0.2452870 0.14133770
#> [15,] 0.2685695 0.12998487
#> [16,] 0.2452870 0.14133770
#> [17,] 0.2452870 0.14133770
#> [18,] 0.2452870 0.14133770
#> [19,] 0.2452870 0.14133770
#> [20,] 0.2452870 0.14133770