Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 5 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.3298741 0.4621726
#>  [2,] 0.9012454 0.9535128
#>  [3,] 0.3468398 0.6421845
#>  [4,] 0.6734746 0.8332144
#>  [5,] 0.3481295 0.5708221
#>  [6,] 0.6863129 0.9866762
#>  [7,] 0.3886158 0.2536544
#>  [8,] 0.2201724 0.1133226
#>  [9,] 0.7820894 0.6932292
#> [10,] 0.5061620 0.4135199
#> [11,] 0.4874888 0.4036201
#> [12,] 0.5783467 0.4225906
#> [13,] 0.9539760 0.8047446
#> [14,] 0.9894061 0.7934015
#> [15,] 0.7078709 0.7704429
#> [16,] 0.9372391 0.7980750
#> [17,] 0.3338701 0.1987294
#> [18,] 0.2329761 0.1257331
#> [19,] 0.5822707 0.4006299
#> [20,] 0.6025178 0.4920489
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -338.1467     -338.1467 1.534233e-40
#>  [2,]         0      -295.9177     -295.9177 3.355290e-22
#>  [3,]         0      -379.7065     -379.7065 1.370003e-58
#>  [4,]         0      -298.4692     -298.4692 2.616018e-23
#>  [5,]         0      -365.8007     -365.8007 1.499459e-52
#>  [6,]         0      -315.5939     -315.5939 9.560616e-31
#>  [7,]         0      -253.6592     -253.6592 7.557011e-04
#>  [8,]         0      -250.9718     -250.9718 1.110426e-02
#>  [9,]         0      -278.2847     -278.2847 1.526261e-14
#> [10,]         0      -274.8317     -274.8317 4.822632e-13
#> [11,]         0      -266.7766     -266.7766 1.519043e-09
#> [12,]         0      -265.2194     -265.2194 7.208459e-09
#> [13,]         0      -278.8036     -278.8036 9.083887e-15
#> [14,]         0      -279.1266     -279.1266 6.576695e-15
#> [15,]         0      -292.6299     -292.6299 8.987061e-21
#> [16,]         0      -284.9917     -284.9917 1.865638e-17
#> [17,]         0      -250.5956     -250.5956 1.617591e-02
#> [18,]         0      -246.4998     -246.4998 9.719638e-01
#> [19,]         0      -261.3765     -261.3765 3.363302e-07
#> [20,]         0      -268.8492     -268.8492 1.911760e-10
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.831861 12.149368 10.012053  8.585619 13.693518 11.456379 10.245187
#>   [8]  8.720465 14.043470 10.531456  9.146778 10.866894  8.745115 12.385331
#>  [15]  9.751495 12.190974  9.034081 11.312039  8.755433 11.358981  8.595698
#>  [22] 11.911303  9.343229 15.828381 12.105288  9.915294 14.228256 10.319288
#>  [29]  8.859726 14.035290 11.314869  9.747737 15.165732 13.744489 10.509392
#>  [36]  9.994392 11.419824  8.714848 14.885979  9.540711 10.434369  3.218147
#>  [43] 17.413131 16.887442 14.807791 13.197437  7.022195 14.386417 10.256438
#>  [50]  3.657781 19.322157 14.778019 15.009968  6.616290 19.013815 18.197544
#>  [57] 12.888488  9.390982 18.769864 18.139731 16.526948 12.152702  9.586704
#>  [64] 18.705427 17.679646 10.549827  5.796271 16.912790 17.824520 15.986195
#>  [71] 12.228230 10.639609 10.289831 11.704149  7.568511 16.202545 10.628659
#>  [78]  9.252105 17.960766 15.160257 16.654755 17.420970 10.957248  9.430016
#>  [85] 18.930300 19.383259 19.155267 13.982627 11.944497 11.038468  6.876163
#>  [92] 13.767285 15.068593 16.046321 16.244761 10.579118  9.272929 19.784337
#>  [99] 14.940073 13.339730
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA   NA 0.70   NA   NA   NA 0.60   NA   NA 0.50   NA 0.70   NA 0.55
#>  [16]   NA 0.40   NA 0.60   NA 0.50   NA 0.55   NA   NA 0.30   NA   NA 0.50   NA
#>  [31]   NA 0.60   NA   NA   NA 0.50   NA 0.45   NA 0.60   NA 0.55   NA   NA   NA
#>  [46]   NA 0.30   NA   NA 0.50   NA   NA   NA 0.35   NA   NA   NA 0.35   NA   NA
#>  [61]   NA   NA 0.25   NA   NA   NA 0.30   NA   NA   NA   NA   NA   NA   NA 0.35
#>  [76]   NA   NA 0.35   NA   NA   NA   NA   NA 0.25   NA   NA   NA   NA   NA   NA
#>  [91] 0.40   NA   NA   NA   NA   NA 0.30   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 247.6
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.2329761 0.1257331
#>  [2,] 0.2329761 0.1257331
#>  [3,] 0.2329761 0.1257331
#>  [4,] 0.2329761 0.1257331
#>  [5,] 0.2329761 0.1257331
#>  [6,] 0.2329761 0.1257331
#>  [7,] 0.2329761 0.1257331
#>  [8,] 0.2329761 0.1257331
#>  [9,] 0.2329761 0.1257331
#> [10,] 0.2329761 0.1257331
#> [11,] 0.2329761 0.1257331
#> [12,] 0.2329761 0.1257331
#> [13,] 0.2329761 0.1257331
#> [14,] 0.2329761 0.1257331
#> [15,] 0.2329761 0.1257331
#> [16,] 0.2329761 0.1257331
#> [17,] 0.2329761 0.1257331
#> [18,] 0.2329761 0.1257331
#> [19,] 0.2329761 0.1257331
#> [20,] 0.2329761 0.1257331