Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars
A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter
A particle_filter object

control
A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 4 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.8786062 0.6437136
#>  [2,] 0.9214060 0.7206962
#>  [3,] 0.7794009 0.4554216
#>  [4,] 0.9031054 0.3912214
#>  [5,] 0.9848284 0.5753385
#>  [6,] 0.9683735 0.6812068
#>  [7,] 0.9665211 0.4632114
#>  [8,] 0.9304471 0.2533553
#>  [9,] 0.9304471 0.2533553
#> [10,] 0.8417721 0.3962091
#> [11,] 0.8908930 0.4795897
#> [12,] 0.8129269 0.3707690
#> [13,] 0.9935318 0.7692723
#> [14,] 0.9212062 0.7203307
#> [15,] 0.8415764 0.5830177
#> [16,] 0.5277357 0.2802811
#> [17,] 0.4028228 0.3333174
#> [18,] 0.8161944 0.4787353
#> [19,] 0.6855155 0.5679502
#> [20,] 0.6855155 0.5679502
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior        weight
#>  [1,]         0      -278.7735     -278.7735  2.859476e-05
#>  [2,]         0      -278.0422     -278.0422  5.941255e-05
#>  [3,]         0      -277.0136     -277.0136  1.661937e-04
#>  [4,]         0      -350.8526     -350.8526  1.421469e-36
#>  [5,]         0      -286.5330     -286.5330  1.220080e-08
#>  [6,]         0      -280.7715     -280.7715  3.877727e-06
#>  [7,]         0      -317.9326     -317.9326  2.816504e-22
#>  [8,]         0      -608.5820     -608.5820 1.668316e-148
#>  [9,]         0      -596.4430     -596.4430 3.120076e-143
#> [10,]         0      -344.8674     -344.8674  5.650361e-34
#> [11,]         0      -295.5379     -295.5379  1.498378e-12
#> [12,]         0      -317.7717     -317.7717  3.308255e-22
#> [13,]         0      -278.1509     -278.1509  5.329317e-05
#> [14,]         0      -274.7557     -274.7557  1.589230e-03
#> [15,]         0      -278.7987     -278.7987  2.788502e-05
#> [16,]         0      -271.2130     -271.2130  5.492417e-02
#> [17,]         0      -274.1308     -274.1308  2.968949e-03
#> [18,]         0      -282.8102     -282.8102  5.048456e-07
#> [19,]         0      -269.2382     -269.2382  3.957467e-01
#> [20,]         0      -268.9192     -268.9192  5.444312e-01
#> 
#> $statistics
#> $statistics$ess
#>   [1] 15.640930 10.775486  9.537136 15.205944 12.313101 12.119734 10.451670
#>   [8]  9.566562 12.447094  9.683855 10.935335  4.253650  9.891268  9.084123
#>  [15]  7.209652  8.681930  7.257702  3.807916  5.315583  4.614897  3.200618
#>  [22]  5.443249  1.368950  1.743904  3.647132  2.199062  2.356738  2.232530
#>  [29]  2.859938  8.504770  6.573243  9.261848 14.853334  9.588694 18.801232
#>  [36] 17.723347 18.388918 18.032706 17.375412 16.467848 16.065108 14.738381
#>  [43] 15.622150 13.482905 13.155182 11.727117 13.730075 13.846506 13.874801
#>  [50] 14.004910 14.413095 13.474362 14.668527  9.540205 18.880216 18.020297
#>  [57] 17.775742 16.306056 16.621497 16.876997 15.329050 13.763136 13.014696
#>  [64] 11.998772 11.349339  8.966360 17.046908 17.768898 16.355222 14.700180
#>  [71] 13.935344 13.562562 12.992445 12.817996 10.891351  9.562253 17.190674
#>  [78] 16.500475 14.630750 14.734692 15.731766 15.777220 14.400955 12.398735
#>  [85] 11.150965 11.847114 12.193318  9.354824 18.845585 15.779387 10.876748
#>  [92]  8.177334 19.224516 19.504540 19.244692 17.179103 15.726231 15.974851
#>  [99] 15.526603 13.804860
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.80   NA   NA   NA   NA 0.75   NA 0.70   NA 0.70 0.45 0.50 0.40
#>  [16] 0.35 0.15 0.30 0.35 0.45 0.40 0.15 0.05 0.25 0.20 0.35 0.40 0.50 0.30 0.55
#>  [31] 0.30 0.50   NA 0.50   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [46]   NA   NA   NA   NA   NA   NA   NA   NA 0.35   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA   NA   NA   NA 0.30   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76] 0.35   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.40   NA   NA
#>  [91]   NA 0.30   NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 198.6
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.6855155 0.5679502
#>  [2,] 0.6855155 0.5679502
#>  [3,] 0.6855155 0.5679502
#>  [4,] 0.5277357 0.2802811
#>  [5,] 0.6855155 0.5679502
#>  [6,] 0.6855155 0.5679502
#>  [7,] 0.6855155 0.5679502
#>  [8,] 0.6855155 0.5679502
#>  [9,] 0.6855155 0.5679502
#> [10,] 0.6855155 0.5679502
#> [11,] 0.6855155 0.5679502
#> [12,] 0.6855155 0.5679502
#> [13,] 0.6855155 0.5679502
#> [14,] 0.6855155 0.5679502
#> [15,] 0.6855155 0.5679502
#> [16,] 0.6855155 0.5679502
#> [17,] 0.6855155 0.5679502
#> [18,] 0.6855155 0.5679502
#> [19,] 0.6855155 0.5679502
#> [20,] 0.6855155 0.5679502