Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 4 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta      gamma
#>  [1,] 0.4653023 0.28292571
#>  [2,] 0.2252273 0.12346243
#>  [3,] 0.7184876 0.31667874
#>  [4,] 0.6087611 0.30200149
#>  [5,] 0.2590465 0.16207320
#>  [6,] 0.3872166 0.21216487
#>  [7,] 0.5647553 0.21257608
#>  [8,] 0.5612190 0.20950850
#>  [9,] 0.3787192 0.16600937
#> [10,] 0.7049833 0.25370523
#> [11,] 0.7049833 0.25370523
#> [12,] 0.5149776 0.30016689
#> [13,] 0.8210590 0.32393379
#> [14,] 0.3186302 0.24772640
#> [15,] 0.1686909 0.08603967
#> [16,] 0.3643739 0.19844256
#> [17,] 0.2816995 0.17697766
#> [18,] 0.6900185 0.30659963
#> [19,] 0.4205434 0.26811106
#> [20,] 0.2686988 0.12969438
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -257.3132     -257.3132 2.026203e-06
#>  [2,]         0      -244.4263     -244.4263 8.004871e-01
#>  [3,]         0      -323.3704     -323.3704 4.153441e-35
#>  [4,]         0      -288.3719     -288.3719 6.577034e-20
#>  [5,]         0      -248.2309     -248.2309 1.782551e-02
#>  [6,]         0      -252.2181     -252.2181 3.306910e-04
#>  [7,]         0      -324.4279     -324.4279 1.442454e-35
#>  [8,]         0      -341.4836     -341.4836 5.648589e-43
#>  [9,]         0      -267.5946     -267.5946 6.942729e-11
#> [10,]         0      -432.0056     -432.0056 2.746033e-82
#> [11,]         0      -432.0056     -432.0056 2.746033e-82
#> [12,]         0      -264.8647     -264.8647 1.064351e-09
#> [13,]         0      -458.3813     -458.3813 9.635952e-94
#> [14,]         0      -258.4774     -258.4774 6.325039e-07
#> [15,]         0      -246.1496     -246.1496 1.428721e-01
#> [16,]         0      -256.5853     -256.5853 4.195537e-06
#> [17,]         0      -247.5689     -247.5689 3.455766e-02
#> [18,]         0      -304.5168     -304.5168 6.403425e-27
#> [19,]         0      -258.4587     -258.4587 6.444636e-07
#> [20,]         0      -249.7456     -249.7456 3.919392e-03
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.043524 10.558297  8.298736 15.597856 11.943107 11.273715  9.699628
#>   [8] 16.160924 14.418917 11.946892 10.014597  7.722709 11.089033  8.553432
#>  [15] 11.903158  9.234312 10.247960  6.129669  7.438555  5.548535  3.609377
#>  [22]  4.357120  2.005882  1.470289  1.198560  1.124104  1.093529  1.137435
#>  [29]  1.946819  4.772020  3.389005  3.257879  9.411688  2.922129 15.565284
#>  [36]  9.832834 18.838170 17.446090 16.574270 16.525713 14.907253 13.178865
#>  [43] 12.957293 13.103275 12.527879 12.301484 10.706531 10.318279 10.571279
#>  [50]  7.415629 18.901439 17.578607 16.363089 14.830729 14.370044 14.206379
#>  [57] 14.015349 13.424980 13.648431 12.969633 12.674543 11.623129 11.207160
#>  [64] 10.398564 10.023287  8.180587 15.636139 17.159858 16.278611 15.150565
#>  [71] 13.418911 13.281122 12.542965 13.960808 12.450619 10.653152  9.605077
#>  [78] 18.929924 17.478087 16.575411 16.858201 16.960703 14.983149 13.707649
#>  [85] 12.803098 13.081105 13.138664 11.906523 10.745784  9.353229 11.685812
#>  [92]  6.822545 18.918771 18.560970 18.322954 16.167297 14.590544 14.809356
#>  [99] 11.103178  9.737715
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.85   NA   NA   NA 0.55   NA   NA   NA   NA 0.65   NA 0.70   NA
#>  [16] 0.65   NA 0.45 0.40 0.20 0.25 0.45 0.25 0.35 0.30 0.15 0.35 0.20 0.35 0.20
#>  [31] 0.20 0.15 0.45 0.40   NA 0.30   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [46]   NA   NA   NA   NA 0.35   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA   NA   NA   NA 0.60   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76]   NA 0.45   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.40
#>  [91]   NA 0.50   NA   NA   NA   NA   NA   NA   NA 0.45
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 221
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta      gamma
#>  [1,] 0.2252273 0.12346243
#>  [2,] 0.2252273 0.12346243
#>  [3,] 0.2252273 0.12346243
#>  [4,] 0.2252273 0.12346243
#>  [5,] 0.2252273 0.12346243
#>  [6,] 0.1686909 0.08603967
#>  [7,] 0.2252273 0.12346243
#>  [8,] 0.2252273 0.12346243
#>  [9,] 0.2252273 0.12346243
#> [10,] 0.2252273 0.12346243
#> [11,] 0.1686909 0.08603967
#> [12,] 0.2816995 0.17697766
#> [13,] 0.2252273 0.12346243
#> [14,] 0.2252273 0.12346243
#> [15,] 0.2252273 0.12346243
#> [16,] 0.2252273 0.12346243
#> [17,] 0.1686909 0.08603967
#> [18,] 0.1686909 0.08603967
#> [19,] 0.2252273 0.12346243
#> [20,] 0.2252273 0.12346243