Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars
A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter
A particle_filter object

control
A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)

#> Finished 100 steps in 5 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res

#> $pars
#>            beta      gamma
#>  [1,] 0.8529808 0.82271630
#>  [2,] 0.5019392 0.36262686
#>  [3,] 0.7088655 0.45160493
#>  [4,] 0.3718913 0.30541369
#>  [5,] 0.5138369 0.29747261
#>  [6,] 0.7230215 0.71669743
#>  [7,] 0.9704239 0.79570790
#>  [8,] 0.7602990 0.44353587
#>  [9,] 0.6410545 0.41356840
#> [10,] 0.7132891 0.47319350
#> [11,] 0.8612085 0.74894807
#> [12,] 0.9322045 0.94221703
#> [13,] 0.6444064 0.64193176
#> [14,] 0.3886257 0.27387225
#> [15,] 0.4284959 0.21028878
#> [16,] 0.8409835 0.74630544
#> [17,] 0.2908220 0.19695225
#> [18,] 0.1417851 0.07752534
#> [19,] 0.1893133 0.08980550
#> [20,] 0.1655558 0.05653590
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -273.5886     -273.5886 3.346226e-12
#>  [2,]         0      -272.9555     -272.9555 6.302877e-12
#>  [3,]         0      -271.6019     -271.6019 2.439909e-11
#>  [4,]         0      -260.4545     -260.4545 1.693020e-06
#>  [5,]         0      -261.9861     -261.9861 3.659826e-07
#>  [6,]         0      -293.0942     -293.0942 1.130882e-20
#>  [7,]         0      -277.4731     -277.4731 6.879558e-14
#>  [8,]         0      -279.2568     -279.2568 1.155918e-14
#>  [9,]         0      -268.8560     -268.8560 3.801357e-10
#> [10,]         0      -267.8890     -267.8890 9.996935e-10
#> [11,]         0      -278.4928     -278.4928 2.481497e-14
#> [12,]         0      -278.1806     -278.1806 3.390856e-14
#> [13,]         0      -269.8150     -269.8150 1.456864e-10
#> [14,]         0      -262.5304     -262.5304 2.123613e-07
#> [15,]         0      -261.7579     -261.7579 4.598385e-07
#> [16,]         0      -279.1488     -279.1488 1.287682e-14
#> [17,]         0      -254.2030     -254.2030 8.782924e-04
#> [18,]         0      -251.0389     -251.0389 2.078606e-02
#> [19,]         0      -247.1874     -247.1874 9.783328e-01
#> [20,]         0      -262.9840     -262.9840 1.349202e-07
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.871188 14.921658 13.681847 13.272595 11.528596 11.510550 10.881768
#>   [8]  7.753661 11.978985  8.525308 10.370269  5.367469  6.263176  6.616754
#>  [15]  8.389671  6.161487  3.379756  1.450065  3.828938  4.690120  2.913379
#>  [22]  2.119341  1.678654  1.477156  1.688529  4.410192  4.832904  4.717994
#>  [29]  4.710634  7.730618  9.225760 11.346089 11.388823  8.482291 17.607354
#>  [36] 14.380328 12.431403 10.884378  9.752358 18.104629 16.095759 13.281538
#>  [43] 12.418866 13.316674 12.204947 10.986894 10.563128 10.135661  9.963626
#>  [50] 13.201058 18.250573 15.998188 15.748128 11.258450 10.605084  9.758228
#>  [57] 15.676447 10.510578 15.035716 14.416450 13.178863 11.860860 10.633425
#>  [64] 10.000657  9.735537 16.042438  7.658109 17.283896 18.493013 17.735928
#>  [71] 14.497570 13.719245 12.761786 15.514833  9.932228 16.288495  8.305940
#>  [78] 19.711568 18.172018 17.193140 18.420826 18.328732 14.360867 11.436816
#>  [85] 10.351674 10.343591 10.515660  6.867772 19.254949 15.675782  8.026961
#>  [92] 17.128136 17.947777 16.816252 17.006963 16.168842 14.667839 14.917178
#>  [99]  8.666995 19.181091
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA   NA   NA   NA   NA   NA 0.70   NA 0.85   NA 0.45 0.40 0.40 0.40
#>  [16] 0.40 0.35 0.65 0.35 0.35 0.30 0.35 0.20 0.20 0.25 0.25 0.50 0.45 0.50 0.60
#>  [31] 0.55   NA   NA 0.45   NA   NA   NA   NA 0.55   NA   NA   NA   NA   NA   NA
#>  [46]   NA   NA   NA 0.55   NA   NA   NA   NA   NA   NA 0.35   NA   NA   NA   NA
#>  [61]   NA   NA   NA   NA 0.30   NA 0.25   NA   NA   NA   NA   NA   NA   NA 0.20
#>  [76]   NA 0.35   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.35   NA   NA
#>  [91] 0.30   NA   NA   NA   NA   NA   NA   NA 0.25   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 257.6
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)

#>            beta      gamma
#>  [1,] 0.1417851 0.07752534
#>  [2,] 0.1893133 0.08980550
#>  [3,] 0.1893133 0.08980550
#>  [4,] 0.1893133 0.08980550
#>  [5,] 0.1893133 0.08980550
#>  [6,] 0.1893133 0.08980550
#>  [7,] 0.1893133 0.08980550
#>  [8,] 0.1893133 0.08980550
#>  [9,] 0.1893133 0.08980550
#> [10,] 0.1893133 0.08980550
#> [11,] 0.1893133 0.08980550
#> [12,] 0.1893133 0.08980550
#> [13,] 0.1893133 0.08980550
#> [14,] 0.1893133 0.08980550
#> [15,] 0.1893133 0.08980550
#> [16,] 0.1893133 0.08980550
#> [17,] 0.1893133 0.08980550
#> [18,] 0.1893133 0.08980550
#> [19,] 0.1893133 0.08980550
#> [20,] 0.1417851 0.07752534