Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 5 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.5638697 0.3549382
#>  [2,] 0.7794116 0.2773594
#>  [3,] 0.5967252 0.1781254
#>  [4,] 0.6375952 0.1715966
#>  [5,] 0.5813593 0.2009096
#>  [6,] 0.6645388 0.2299377
#>  [7,] 0.6375952 0.1715966
#>  [8,] 0.5164777 0.1975853
#>  [9,] 0.5987234 0.1701817
#> [10,] 0.5816989 0.1431217
#> [11,] 0.5548261 0.1435435
#> [12,] 0.5919411 0.2051464
#> [13,] 0.5816989 0.1431217
#> [14,] 0.5816989 0.1431217
#> [15,] 0.5816989 0.1431217
#> [16,] 0.5969086 0.1607611
#> [17,] 0.6401319 0.2732567
#> [18,] 0.6951961 0.3173584
#> [19,] 0.5733350 0.4281890
#> [20,] 0.5180847 0.3840221
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior        weight
#>  [1,]         0      -260.2386     -260.2386  9.708824e-01
#>  [2,]         0      -489.2143     -489.2143 3.501726e-100
#>  [3,]         0      -516.9903     -516.9903 3.029378e-112
#>  [4,]         0      -741.1780     -741.1780 1.311701e-209
#>  [5,]         0      -353.8479     -353.8479  2.153561e-41
#>  [6,]         0      -442.8056     -442.8056  5.004328e-80
#>  [7,]         0      -739.5244     -739.5244 6.855090e-209
#>  [8,]         0      -291.5379     -291.5379  2.477870e-14
#>  [9,]         0      -468.6327     -468.6327  3.039308e-91
#> [10,]         0     -1177.2890    -1177.2890  0.000000e+00
#> [11,]         0     -1077.8871    -1077.8871  0.000000e+00
#> [12,]         0      -371.8885     -371.8885  3.149638e-49
#> [13,]         0     -1175.4855    -1175.4855  0.000000e+00
#> [14,]         0     -1175.1596    -1175.1596  0.000000e+00
#> [15,]         0     -1175.3661    -1175.3661  0.000000e+00
#> [16,]         0      -644.0810     -644.0810 1.934390e-167
#> [17,]         0      -313.3133     -313.3133  8.652673e-24
#> [18,]         0      -303.6150     -303.6150  1.409428e-19
#> [19,]         0      -263.9509     -263.9509  2.371207e-02
#> [20,]         0      -265.4294     -265.4294  5.405526e-03
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.543501 12.412510  9.521904 14.031795 10.101699  9.508019 14.254014
#>   [8] 12.253283 10.554813  9.504475 11.687458  9.416281  8.404882  8.085365
#>  [15]  5.261197  1.996055  2.315369  1.827814  2.993185  1.884269  2.865070
#>  [22]  5.947337  3.613091  2.267674  5.249366  3.655842  5.759681  4.420845
#>  [29]  6.078691  9.515440  8.613943  9.227683 12.925000  8.105587 18.146253
#>  [36] 16.075431 14.002469 13.154154 12.743110 11.651110  9.914777 14.944218
#>  [43] 15.967818 16.129541 15.769939 14.688400 12.655112 11.466297 10.603275
#>  [50]  9.499285 19.015276 16.522359 16.254977 12.151977 11.860242 11.613621
#>  [57] 10.456950 10.016704 10.179845  9.731486 15.876827 12.347895 10.118835
#>  [64]  9.105153 18.815315 15.107384 11.489564 12.491256 10.155458  8.920474
#>  [71] 17.399378 16.781853 16.254969 16.629201 13.280489 10.832684  7.797804
#>  [78] 17.453789 16.676645 16.260691 16.363985 16.394464 13.633071 12.340053
#>  [85] 10.198156 10.305703 10.782523 10.205713  8.376998 16.373970 11.194570
#>  [92]  9.255019 17.900072 16.125688 16.194289 13.418149 11.440781 12.477454
#>  [99]  7.933294 14.975177
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.70   NA   NA 0.65   NA   NA   NA 0.70   NA 0.45 0.50 0.40 0.55
#>  [16] 0.65 0.45 0.35 0.35 0.60 0.25 0.40 0.60 0.35 0.40 0.55 0.50 0.55 0.70 0.50
#>  [31] 0.45 0.45   NA 0.40   NA   NA   NA   NA   NA   NA 0.70   NA   NA   NA   NA
#>  [46]   NA   NA   NA   NA 0.55   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.55
#>  [61]   NA   NA   NA 0.55   NA   NA   NA   NA   NA 0.30   NA   NA   NA   NA   NA
#>  [76]   NA 0.45   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.45   NA
#>  [91]   NA 0.60   NA   NA   NA   NA   NA   NA 0.35   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 251.4
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.5638697 0.3549382
#>  [2,] 0.5638697 0.3549382
#>  [3,] 0.5638697 0.3549382
#>  [4,] 0.5638697 0.3549382
#>  [5,] 0.5638697 0.3549382
#>  [6,] 0.5638697 0.3549382
#>  [7,] 0.5733350 0.4281890
#>  [8,] 0.5638697 0.3549382
#>  [9,] 0.5638697 0.3549382
#> [10,] 0.5638697 0.3549382
#> [11,] 0.5638697 0.3549382
#> [12,] 0.5638697 0.3549382
#> [13,] 0.5638697 0.3549382
#> [14,] 0.5638697 0.3549382
#> [15,] 0.5638697 0.3549382
#> [16,] 0.5638697 0.3549382
#> [17,] 0.5638697 0.3549382
#> [18,] 0.5638697 0.3549382
#> [19,] 0.5638697 0.3549382
#> [20,] 0.5638697 0.3549382