Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 4 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.2979329 0.2028387
#>  [2,] 0.4356365 0.3138683
#>  [3,] 0.7875211 0.2866836
#>  [4,] 0.6385621 0.2512303
#>  [5,] 0.4347711 0.2690180
#>  [6,] 0.8044366 0.2400268
#>  [7,] 0.3184256 0.2120904
#>  [8,] 0.6106574 0.3859329
#>  [9,] 0.4583496 0.3832904
#> [10,] 0.4218176 0.1840554
#> [11,] 0.2285596 0.1593934
#> [12,] 0.4950961 0.3722796
#> [13,] 0.2453759 0.1771210
#> [14,] 0.4277116 0.2387346
#> [15,] 0.5085384 0.3543030
#> [16,] 0.2118208 0.1580547
#> [17,] 0.1774907 0.1133360
#> [18,] 0.4715265 0.3266530
#> [19,] 0.3507396 0.2457326
#> [20,] 0.5355642 0.2364348
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -256.0696     -256.0696 3.133938e-03
#>  [2,]         0      -259.2704     -259.2704 1.276428e-04
#>  [3,]         0      -432.7265     -432.7265 5.956387e-80
#>  [4,]         0      -327.9696     -327.9696 1.863482e-34
#>  [5,]         0      -259.8994     -259.8994 6.805141e-05
#>  [6,]         0      -433.5099     -433.5099 2.721300e-80
#>  [7,]         0      -250.4591     -250.4591 8.564872e-01
#>  [8,]         0      -264.6211     -264.6211 6.056394e-07
#>  [9,]         0      -268.9965     -268.9965 7.620758e-09
#> [10,]         0      -269.4365     -269.4365 4.908066e-09
#> [11,]         0      -255.8565     -255.8565 3.878455e-03
#> [12,]         0      -263.8026     -263.8026 1.373095e-06
#> [13,]         0      -255.1403     -255.1403 7.937770e-03
#> [14,]         0      -260.8804     -260.8804 2.551479e-05
#> [15,]         0      -260.0914     -260.0914 5.616393e-05
#> [16,]         0      -262.5257     -262.5257 4.923450e-06
#> [17,]         0      -252.9587     -252.9587 7.032961e-02
#> [18,]         0      -261.0503     -261.0503 2.152732e-05
#> [19,]         0      -253.1527     -253.1527 5.792717e-02
#> [20,]         0      -283.4973     -283.4973 3.840459e-15
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.093391 11.712240  9.419512 13.694331 10.113122  9.060563 11.510099
#>   [8] 10.215513  8.466480 15.655966 13.375653  9.582053 11.793811  9.509436
#>  [15] 12.062433 10.382918  8.476714  8.118369  7.037422  3.449619  1.976621
#>  [22]  3.785630  1.817085  2.193029  5.729548  5.019575  5.279442  4.458814
#>  [29]  8.756097  9.231431 10.141424  7.466284 16.791321 11.137185 12.239930
#>  [36] 11.699644 11.725090 11.579040 10.852908 10.633003  9.952157 13.941230
#>  [43] 15.350465 17.383564 17.159899 16.823230 15.332467 14.277025 14.197482
#>  [50] 11.121126 11.966418 10.638879 11.746968  7.149704 19.251857 18.734874
#>  [57] 19.056349 17.754027 17.793192 17.421578 17.044143 15.844718 15.582983
#>  [64] 14.775555 14.888357 14.180573 13.604940 14.162263 13.439397 12.585674
#>  [71] 11.448716 10.848997 10.049480 11.107836  9.486624 18.520350 15.539995
#>  [78] 14.908590 12.451700 11.836759 13.698269 14.368545 11.699488  9.465604
#>  [85] 19.412066 19.600473 19.670754 18.113726 16.632963 14.831072 10.375741
#>  [92]  6.648641 19.356400 19.311716 19.406803 17.356682 16.500514 16.663957
#>  [99] 12.820131 11.281180
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.85   NA   NA 0.55   NA   NA 0.45   NA   NA 0.55   NA 0.35   NA
#>  [16]   NA 0.35 0.40 0.35 0.25 0.35 0.15 0.30 0.45 0.40 0.60 0.45 0.60 0.45 0.60
#>  [31]   NA 0.55   NA   NA   NA   NA   NA   NA   NA   NA 0.30   NA   NA   NA   NA
#>  [46]   NA   NA   NA   NA   NA   NA   NA   NA 0.35   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.30
#>  [76]   NA   NA   NA   NA   NA   NA   NA   NA 0.35   NA   NA   NA   NA   NA   NA
#>  [91]   NA 0.40   NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 170.2
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.3184256 0.2120904
#>  [2,] 0.3184256 0.2120904
#>  [3,] 0.3184256 0.2120904
#>  [4,] 0.3184256 0.2120904
#>  [5,] 0.3184256 0.2120904
#>  [6,] 0.3184256 0.2120904
#>  [7,] 0.3184256 0.2120904
#>  [8,] 0.3184256 0.2120904
#>  [9,] 0.3184256 0.2120904
#> [10,] 0.3184256 0.2120904
#> [11,] 0.3184256 0.2120904
#> [12,] 0.3184256 0.2120904
#> [13,] 0.3184256 0.2120904
#> [14,] 0.3184256 0.2120904
#> [15,] 0.3184256 0.2120904
#> [16,] 0.3184256 0.2120904
#> [17,] 0.3184256 0.2120904
#> [18,] 0.3184256 0.2120904
#> [19,] 0.3184256 0.2120904
#> [20,] 0.3184256 0.2120904