Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 10 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.5635217 0.3806904
#>  [2,] 0.6747048 0.6962108
#>  [3,] 0.4345065 0.4135748
#>  [4,] 0.2059686 0.1385882
#>  [5,] 0.3798210 0.4619731
#>  [6,] 0.6074880 0.9622521
#>  [7,] 0.6739778 0.6743388
#>  [8,] 0.3804015 0.6088137
#>  [9,] 0.3087617 0.5766475
#> [10,] 0.2295567 0.3962910
#> [11,] 0.4331036 0.3570816
#> [12,] 0.3346086 0.2011267
#> [13,] 0.6114944 0.6792161
#> [14,] 0.7443958 0.8604939
#> [15,] 0.4267086 0.5961160
#> [16,] 0.4429949 0.3895651
#> [17,] 0.8637295 0.7331694
#> [18,] 0.6471032 0.9200694
#> [19,] 0.4361667 0.2699596
#> [20,] 0.3813327 0.2538055
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -268.0174     -268.0174 9.407472e-08
#>  [2,]         0      -286.6440     -286.6440 7.656720e-16
#>  [3,]         0      -279.1301     -279.1301 1.403776e-12
#>  [4,]         0      -254.4899     -254.4899 7.053290e-02
#>  [5,]         0      -295.7184     -295.7184 8.771958e-20
#>  [6,]         0      -335.7434     -335.7434 3.634326e-37
#>  [7,]         0      -279.7957     -279.7957 7.214352e-13
#>  [8,]         0      -348.8019     -348.8019 7.748610e-43
#>  [9,]         0      -401.2324     -401.2324 1.315050e-65
#> [10,]         0      -383.4282     -383.4282 7.098949e-58
#> [11,]         0      -271.5349     -271.5349 2.791438e-09
#> [12,]         0      -252.8504     -252.8504 3.634270e-01
#> [13,]         0      -299.2212     -299.2212 2.641295e-21
#> [14,]         0      -293.7529     -293.7529 6.261415e-19
#> [15,]         0      -330.2138     -330.2138 9.160295e-35
#> [16,]         0      -272.4887     -272.4887 1.075512e-09
#> [17,]         0      -280.3804     -280.3804 4.020375e-13
#> [18,]         0      -324.1500     -324.1500 3.938930e-32
#> [19,]         0      -259.8110     -259.8110 3.447238e-04
#> [20,]         0      -252.4079     -252.4079 5.656952e-01
#> 
#> $statistics
#> $statistics$ess
#>   [1] 15.433599  9.787993 14.956608 11.742703 10.081981  9.743570 17.377690
#>   [8] 15.649668 14.758356 13.903637 11.838732  9.834962 10.677855  8.919454
#>  [15]  5.106778  4.206457  4.820195  3.951844  2.947653  4.121352  4.057099
#>  [22]  1.070254  3.457004  3.163097  2.215245  2.015562  2.037817  1.301242
#>  [29]  2.435536  1.099430  3.275616  5.750491  5.081310  9.780877  1.559513
#>  [36]  5.070164  2.192325  3.086214  2.747099  2.485244  1.869416  1.669380
#>  [43]  4.821538  7.401920  4.294139  6.025251  4.113651  4.611707  5.841768
#>  [50]  3.171081 11.127256  2.818589 18.037493  1.610512 18.757875 14.411791
#>  [57]  4.925858  8.131257 19.014084 17.170886  8.728267 11.736162  5.956677
#>  [64] 18.159797 15.530948  8.873909 11.371822 14.500679  9.200715 17.384793
#>  [71]  8.818736 19.488484 18.555955 18.842791 14.345336  9.141588 17.238048
#>  [78] 16.158951 12.352998  9.658832 19.654040 19.139045 15.614179 12.209540
#>  [85] 11.276320 11.509926 11.575215  9.355636 18.383136 15.568017  9.971585
#>  [92]  7.005473 19.017258 17.849642 17.340119 13.603880 11.024513 11.287700
#>  [99]  5.940481 18.446528
#> 
#> $statistics$acceptance_rate
#>   [1]   NA 0.70   NA   NA   NA 0.55   NA   NA   NA   NA   NA 0.60   NA 0.95 0.80
#>  [16] 0.70 0.85 0.50 0.60 0.45 0.50 0.45 0.45 0.35 0.70 0.40 0.50 0.60 0.60 0.40
#>  [31] 0.45 0.50 0.50 0.35 0.60 0.40 0.30 0.55 0.50 0.50 0.55 0.40 0.45 0.50 0.65
#>  [46] 0.60 0.60 0.55 0.50 0.55   NA 0.40   NA 0.30   NA   NA 0.50 0.45   NA   NA
#>  [61] 0.50   NA 0.35   NA   NA 0.40   NA   NA 0.40   NA 0.15   NA   NA   NA   NA
#>  [76] 0.65   NA   NA   NA 0.60   NA   NA   NA   NA   NA   NA   NA 0.40   NA   NA
#>  [91] 0.55 0.35   NA   NA   NA   NA   NA   NA 0.50   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 476.2
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.3346086 0.2011267
#>  [2,] 0.3346086 0.2011267
#>  [3,] 0.3346086 0.2011267
#>  [4,] 0.3813327 0.2538055
#>  [5,] 0.3813327 0.2538055
#>  [6,] 0.3346086 0.2011267
#>  [7,] 0.3346086 0.2011267
#>  [8,] 0.3813327 0.2538055
#>  [9,] 0.3346086 0.2011267
#> [10,] 0.2059686 0.1385882
#> [11,] 0.3813327 0.2538055
#> [12,] 0.3346086 0.2011267
#> [13,] 0.3346086 0.2011267
#> [14,] 0.3346086 0.2011267
#> [15,] 0.3346086 0.2011267
#> [16,] 0.3813327 0.2538055
#> [17,] 0.3813327 0.2538055
#> [18,] 0.3346086 0.2011267
#> [19,] 0.2059686 0.1385882
#> [20,] 0.3813327 0.2538055