Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 4 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.6266245 0.4770313
#>  [2,] 0.5348160 0.3957876
#>  [3,] 0.5766346 0.3796365
#>  [4,] 0.6831072 0.4519022
#>  [5,] 0.4924612 0.3540175
#>  [6,] 0.6065127 0.2467650
#>  [7,] 0.4571027 0.3163489
#>  [8,] 0.6572835 0.2781478
#>  [9,] 0.5299988 0.3583517
#> [10,] 0.6438695 0.3307542
#> [11,] 0.5759681 0.3492902
#> [12,] 0.5926762 0.3435696
#> [13,] 0.7193898 0.5181804
#> [14,] 0.5288198 0.4192230
#> [15,] 0.5972312 0.3131388
#> [16,] 0.8443727 0.5878887
#> [17,] 0.7870448 0.3263883
#> [18,] 0.7808404 0.2477519
#> [19,] 0.6305017 0.4899910
#> [20,] 0.4829823 0.2670600
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -263.2230     -263.2230 5.081057e-02
#>  [2,]         0      -262.3689     -262.3689 1.193743e-01
#>  [3,]         0      -262.9673     -262.9673 6.561876e-02
#>  [4,]         0      -268.3886     -268.3886 2.901301e-04
#>  [5,]         0      -263.8788     -263.8788 2.637265e-02
#>  [6,]         0      -338.4323     -338.4323 1.104000e-34
#>  [7,]         0      -263.3527     -263.3527 4.463352e-02
#>  [8,]         0      -307.0580     -307.0580 4.663075e-21
#>  [9,]         0      -261.2430     -261.2430 3.680109e-01
#> [10,]         0      -274.9492     -274.9492 4.105447e-07
#> [11,]         0      -262.6511     -262.6511 9.002039e-02
#> [12,]         0      -264.7232     -264.7232 1.133588e-02
#> [13,]         0      -272.8273     -272.8273 3.426891e-06
#> [14,]         0      -269.2581     -269.2581 1.216119e-04
#> [15,]         0      -277.2476     -277.2476 4.122408e-08
#> [16,]         0      -271.2170     -271.2170 1.714808e-05
#> [17,]         0      -306.1733     -306.1733 1.129533e-20
#> [18,]         0      -415.9225     -415.9225 2.451417e-68
#> [19,]         0      -266.2484     -266.2484 2.466499e-03
#> [20,]         0      -261.7533     -261.7533 2.209238e-01
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.075225 12.520454 10.648528  9.551672 15.475182 14.221776 11.810956
#>   [8] 11.174832  9.842754 12.172172  8.981795  7.207607 11.185471  9.613683
#>  [15]  8.655433  7.213655  7.828752  8.523092  5.896405  6.503821  3.752980
#>  [22]  4.274000  3.138316  3.707030  6.713567  7.157987  6.711276  7.305442
#>  [29]  8.072226 10.768780  8.824094 10.454477  9.844477  9.893360 17.615287
#>  [36] 16.497018 14.816667 13.752914 12.897784 12.357141 12.095668 11.934563
#>  [43] 11.380133 10.292750 10.013671  9.276664 16.248330 15.402345 14.894912
#>  [50] 12.393279 13.410844 12.323828 13.360768  7.959530 18.906873 18.327592
#>  [57] 19.127481 17.832574 18.288272 18.232508 16.320221 14.354482 13.100074
#>  [64] 13.105261 12.630216  9.984432 17.538213 18.095948 16.935953 15.880105
#>  [71] 14.965015 14.634836 13.997743 14.119778 12.882511 11.386958 10.237851
#>  [78]  9.646607 19.384790 18.607399 18.871627 18.766831 17.270507 15.717310
#>  [85] 14.653243 14.748761 14.866692 11.928165 13.180438 11.772631  6.442746
#>  [92] 16.113869 17.032095 16.006502 15.774693 12.997812 12.963889 13.679003
#>  [99] 11.491809 10.433892
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA   NA 0.60   NA   NA   NA   NA 0.55   NA 0.60 0.40   NA 0.50 0.45
#>  [16] 0.50 0.30 0.45 0.20 0.35 0.35 0.40 0.45 0.55 0.50 0.50 0.60 0.55 0.40   NA
#>  [31] 0.70   NA 0.45 0.55   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [46] 0.60   NA   NA   NA   NA   NA   NA   NA 0.35   NA   NA   NA   NA   NA   NA
#>  [61]   NA   NA   NA   NA   NA 0.40   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76]   NA   NA 0.50   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [91] 0.25   NA   NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 182.6
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.4571027 0.3163489
#>  [2,] 0.5766346 0.3796365
#>  [3,] 0.5759681 0.3492902
#>  [4,] 0.4829823 0.2670600
#>  [5,] 0.4829823 0.2670600
#>  [6,] 0.5299988 0.3583517
#>  [7,] 0.5759681 0.3492902
#>  [8,] 0.4829823 0.2670600
#>  [9,] 0.5766346 0.3796365
#> [10,] 0.5299988 0.3583517
#> [11,] 0.5348160 0.3957876
#> [12,] 0.5299988 0.3583517
#> [13,] 0.4829823 0.2670600
#> [14,] 0.4829823 0.2670600
#> [15,] 0.5759681 0.3492902
#> [16,] 0.4829823 0.2670600
#> [17,] 0.5299988 0.3583517
#> [18,] 0.4829823 0.2670600
#> [19,] 0.5759681 0.3492902
#> [20,] 0.4829823 0.2670600