Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 4 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.5489669 0.3637664
#>  [2,] 0.6406957 0.4364599
#>  [3,] 0.5127247 0.1957208
#>  [4,] 0.7489883 0.3652438
#>  [5,] 0.8641901 0.5416322
#>  [6,] 0.7524941 0.3085992
#>  [7,] 0.7761599 0.4718165
#>  [8,] 0.6990801 0.3735866
#>  [9,] 0.6685559 0.4473855
#> [10,] 0.7637622 0.6344609
#> [11,] 0.6815587 0.4509838
#> [12,] 0.5632089 0.2510618
#> [13,] 0.7033602 0.3352163
#> [14,] 0.5037243 0.3586787
#> [15,] 0.4585063 0.3312808
#> [16,] 0.5442838 0.3972794
#> [17,] 0.6500983 0.2092604
#> [18,] 0.6721500 0.2572125
#> [19,] 0.5076777 0.2252409
#> [20,] 0.3959983 0.2783685
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -264.0924     -264.0924 4.112245e-04
#>  [2,]         0      -270.6680     -270.6680 5.732430e-07
#>  [3,]         0      -305.8728     -305.8728 2.944928e-22
#>  [4,]         0      -290.2061     -290.2061 1.875241e-15
#>  [5,]         0      -277.1562     -277.1562 8.720697e-10
#>  [6,]         0      -356.5691     -356.5691 2.831223e-44
#>  [7,]         0      -285.2103     -285.2103 2.771308e-13
#>  [8,]         0      -288.6593     -288.6593 8.806858e-15
#>  [9,]         0      -270.3072     -270.3072 8.223177e-07
#> [10,]         0      -274.4375     -274.4375 1.322159e-08
#> [11,]         0      -266.0498     -266.0498 5.807948e-05
#> [12,]         0      -290.8897     -290.8897 9.466501e-16
#> [13,]         0      -289.1974     -289.1974 5.141908e-15
#> [14,]         0      -261.4881     -261.4881 5.560497e-03
#> [15,]         0      -262.0622     -262.0622 3.131714e-03
#> [16,]         0      -263.6953     -263.6953 6.117026e-04
#> [17,]         0      -430.7903     -430.7903 1.652323e-76
#> [18,]         0      -328.1823     -328.1823 6.028266e-32
#> [19,]         0      -302.1145     -302.1145 1.262651e-20
#> [20,]         0      -256.3059     -256.3059 9.902254e-01
#> 
#> $statistics
#> $statistics$ess
#>   [1] 17.078057 13.577659 11.126239  8.777764 16.003803 15.401537 14.342380
#>   [8] 12.745875 11.506346  9.258556 13.524214  9.977395  7.554286  7.403734
#>  [15]  8.371626  7.814480  4.838270  4.088525  2.594820  1.278993  1.033943
#>  [22]  1.519691  1.786546  1.120538  3.173256  2.516832  3.890375  6.888629
#>  [29]  3.462290  6.149600  3.882741  4.278827 10.902279  4.726946 15.289778
#>  [36] 12.774553 12.688730 12.447474 11.925864 11.687911 11.402996 10.119210
#>  [43] 10.172842 10.912087 10.683922 10.092618  8.614559 17.319866 16.248189
#>  [50] 12.474633 13.148463 11.772992 11.749597 10.054029 10.158814 10.097556
#>  [57]  9.175067 15.064827 16.129748 16.328050 14.116910 12.280000 11.754166
#>  [64] 11.209490 11.315416  9.552387 14.915989 17.724799 16.255418 14.023623
#>  [71] 10.490745 10.546047 11.408375 14.746876 12.615905 10.525903  7.788806
#>  [78] 19.682681 16.955058 15.209446 17.762417 18.040583 15.942067 13.995362
#>  [85] 13.056916 13.568260 13.698739 12.714144 12.323652  9.818502 12.504506
#>  [92]  6.112893 19.233136 19.199264 18.720393 17.421975 16.894262 17.411313
#>  [99] 13.805836 14.002588
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA   NA 0.60   NA   NA   NA   NA   NA 0.70   NA 0.60 0.40 0.50 0.40
#>  [16] 0.45 0.30 0.45 0.30 0.30 0.45 0.10 0.35 0.55 0.35 0.50 0.60 0.30 0.50 0.45
#>  [31] 0.45 0.55   NA 0.35   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [46]   NA 0.35   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.65   NA   NA   NA
#>  [61]   NA   NA   NA   NA   NA 0.45   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76]   NA 0.55   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.60
#>  [91]   NA 0.45   NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 207.8
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.3959983 0.2783685
#>  [2,] 0.3959983 0.2783685
#>  [3,] 0.3959983 0.2783685
#>  [4,] 0.3959983 0.2783685
#>  [5,] 0.3959983 0.2783685
#>  [6,] 0.3959983 0.2783685
#>  [7,] 0.3959983 0.2783685
#>  [8,] 0.3959983 0.2783685
#>  [9,] 0.3959983 0.2783685
#> [10,] 0.3959983 0.2783685
#> [11,] 0.3959983 0.2783685
#> [12,] 0.3959983 0.2783685
#> [13,] 0.3959983 0.2783685
#> [14,] 0.3959983 0.2783685
#> [15,] 0.3959983 0.2783685
#> [16,] 0.3959983 0.2783685
#> [17,] 0.3959983 0.2783685
#> [18,] 0.3959983 0.2783685
#> [19,] 0.3959983 0.2783685
#> [20,] 0.3959983 0.2783685