Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 5 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta      gamma
#>  [1,] 0.2672479 0.15429515
#>  [2,] 0.6406604 0.19248095
#>  [3,] 0.6577874 0.22844108
#>  [4,] 0.5542338 0.17419692
#>  [5,] 0.4737157 0.15535867
#>  [6,] 0.9051280 0.33172378
#>  [7,] 0.6440128 0.25347288
#>  [8,] 0.7923908 0.31396420
#>  [9,] 0.9532320 0.41899886
#> [10,] 0.7797637 0.44011986
#> [11,] 0.3488672 0.16452152
#> [12,] 0.4485232 0.29682077
#> [13,] 0.8053436 0.44895227
#> [14,] 0.6812566 0.55136681
#> [15,] 0.3554548 0.17326733
#> [16,] 0.2230750 0.11416599
#> [17,] 0.2603285 0.15383611
#> [18,] 0.2853387 0.15755944
#> [19,] 0.2131908 0.05548211
#> [20,] 0.2472189 0.11643353
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior        weight
#>  [1,]         0      -245.7462     -245.7462  4.125883e-01
#>  [2,]         0      -506.8078     -506.8078 1.729297e-114
#>  [3,]         0      -453.4836     -453.4836  2.490640e-91
#>  [4,]         0      -435.8965     -435.8965  1.082141e-83
#>  [5,]         0      -338.1606     -338.1606  3.022822e-41
#>  [6,]         0      -504.9106     -504.9106 1.153021e-113
#>  [7,]         0      -326.0939     -326.0939  5.259352e-36
#>  [8,]         0      -410.8896     -410.8896  7.845495e-73
#>  [9,]         0      -428.8507     -428.8507  1.242245e-80
#> [10,]         0      -283.2043     -283.2043  2.226574e-17
#> [11,]         0      -253.9065     -253.9065  1.179017e-04
#> [12,]         0      -253.8424     -253.8424  1.257019e-04
#> [13,]         0      -290.5334     -290.5334  1.461005e-20
#> [14,]         0      -272.2419     -272.2419  1.284037e-12
#> [15,]         0      -259.4036     -259.4036  4.832313e-07
#> [16,]         0      -245.6057     -245.6057  4.748043e-01
#> [17,]         0      -247.5640     -247.5640  6.699117e-02
#> [18,]         0      -249.3499     -249.3499  1.123107e-02
#> [19,]         0      -272.0717     -272.0717  1.522263e-12
#> [20,]         0      -248.2381     -248.2381  3.414111e-02
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.758775 11.440959  8.993842 15.887222 11.914206 10.572762  9.635389
#>   [8] 14.006550 11.654793 10.127090  7.990271  3.605332  7.787383  6.065883
#>  [15]  6.482978  8.378646  3.520470  2.925213  2.968727  3.028992  2.111189
#>  [22]  3.754267  1.263031  2.684396  2.566763  4.605891  1.318789  1.813653
#>  [29]  3.596290  4.715076  3.803582  4.712530  9.905592  5.770817 15.463358
#>  [36] 11.259354 11.095331 10.796725 10.386230 10.417371  9.962388 16.839656
#>  [43] 17.249570 14.352182 13.957423 12.477130 12.341837 11.763423 11.579233
#>  [50] 10.184398 10.413091  9.638646 15.790502 15.186833 13.330488 12.836271
#>  [57] 11.913751 10.411852 11.128074 11.426804 10.030177  8.909976 17.847463
#>  [64] 15.261428 14.233059 12.072298  8.910473 12.845072 15.078133 14.211994
#>  [71] 13.536939 13.207017 12.592556 13.049927 11.982016 11.787541 10.445426
#>  [78]  9.558947 17.092672 16.231133 16.262893 15.961255 11.376784  9.560934
#>  [85] 17.862878 17.857880 17.168117 14.638340 12.957025 10.942440  5.783923
#>  [92]  9.610884 18.053330 17.555906 17.737612 12.300216 10.461840 10.707741
#>  [99]  7.311551 17.813412
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.70   NA   NA   NA 0.85   NA   NA   NA 0.65 0.80 0.70 0.50 0.50
#>  [16] 0.70 0.45 0.50 0.55 0.60 0.30 0.45 0.40 0.40 0.40 0.40 0.25 0.60 0.45 0.50
#>  [31] 0.35 0.75 0.45 0.55   NA   NA   NA   NA   NA   NA 0.50   NA   NA   NA   NA
#>  [46]   NA   NA   NA   NA   NA   NA 0.30   NA   NA   NA   NA   NA   NA   NA   NA
#>  [61]   NA 0.45   NA   NA   NA   NA 0.35   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76]   NA   NA 0.50   NA   NA   NA   NA   NA 0.50   NA   NA   NA   NA   NA   NA
#>  [91] 0.25 0.45   NA   NA   NA   NA   NA   NA 0.35   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 263.2
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.2603285 0.1538361
#>  [2,] 0.2672479 0.1542952
#>  [3,] 0.2672479 0.1542952
#>  [4,] 0.2230750 0.1141660
#>  [5,] 0.2472189 0.1164335
#>  [6,] 0.2672479 0.1542952
#>  [7,] 0.2672479 0.1542952
#>  [8,] 0.2230750 0.1141660
#>  [9,] 0.2230750 0.1141660
#> [10,] 0.2230750 0.1141660
#> [11,] 0.2230750 0.1141660
#> [12,] 0.2672479 0.1542952
#> [13,] 0.2230750 0.1141660
#> [14,] 0.2230750 0.1141660
#> [15,] 0.2472189 0.1164335
#> [16,] 0.2603285 0.1538361
#> [17,] 0.2672479 0.1542952
#> [18,] 0.2603285 0.1538361
#> [19,] 0.2603285 0.1538361
#> [20,] 0.2230750 0.1141660