Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4, 0)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 9 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.7205076 0.5152818
#>  [2,] 0.6737100 0.4739463
#>  [3,] 0.3986721 0.2507847
#>  [4,] 0.8811124 0.6657855
#>  [5,] 0.8536868 0.9171112
#>  [6,] 0.6273758 0.3833057
#>  [7,] 0.9562266 0.7262537
#>  [8,] 0.9137904 0.8259329
#>  [9,] 0.9992540 0.9721307
#> [10,] 0.5856941 0.6199171
#> [11,] 0.5089493 0.6311209
#> [12,] 0.3700838 0.2348021
#> [13,] 0.7903663 0.6629748
#> [14,] 0.9279070 0.6393224
#> [15,] 0.9383732 0.9362328
#> [16,] 0.6217106 0.3285917
#> [17,] 0.6847381 0.6313767
#> [18,] 0.9809256 0.9267186
#> [19,] 0.9960983 0.8546837
#> [20,] 0.4465820 0.3699207
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -269.3881     -269.3881 3.650761e-07
#>  [2,]         0      -268.3946     -268.3946 9.859505e-07
#>  [3,]         0      -254.5796     -254.5796 9.854483e-01
#>  [4,]         0      -276.0490     -276.0490 4.673217e-10
#>  [5,]         0      -279.4540     -279.4540 1.551788e-11
#>  [6,]         0      -264.6138     -264.6138 4.323550e-05
#>  [7,]         0      -280.1004     -280.1004 8.129911e-12
#>  [8,]         0      -281.7730     -281.7730 1.526552e-12
#>  [9,]         0      -277.6791     -277.6791 9.155265e-11
#> [10,]         0      -280.7785     -280.7785 4.126912e-12
#> [11,]         0      -293.6270     -293.6270 1.085359e-17
#> [12,]         0      -258.7981     -258.7981 1.450703e-02
#> [13,]         0      -272.5808     -272.5808 1.499117e-08
#> [14,]         0      -281.8461     -281.8461 1.418879e-12
#> [15,]         0      -278.4891     -278.4891 4.072775e-11
#> [16,]         0      -272.2580     -272.2580 2.070169e-08
#> [17,]         0      -281.2409     -281.2409 2.598856e-12
#> [18,]         0      -285.1319     -285.1319 5.308060e-14
#> [19,]         0      -280.1715     -280.1715 7.571926e-12
#> [20,]         0      -274.2461     -274.2461 2.835225e-09
#> 
#> $statistics
#> $statistics$ess
#>   [1] 15.353289 11.261335  9.200434 15.889100 13.377050 11.714301 10.275916
#>   [8]  7.897779 14.155452 11.324463  8.014034 14.831008 12.557837 12.238665
#>  [15] 10.903208  8.577360 11.151588  8.487574  9.041484 13.777816 12.031938
#>  [22]  7.114499 12.560480 10.305798  6.770238 14.017597 11.957421 12.092802
#>  [29] 10.557329  7.824657 14.857308 12.325503  8.155165 16.531655  7.042463
#>  [36] 10.959541  6.628089 10.388101  7.590232  8.027993  6.687423  4.971399
#>  [43] 15.131134 15.464659  9.877830 18.137781  6.301512 14.483561  9.690452
#>  [50]  5.888396 19.513178 14.633669 15.293451  4.545940 19.704592 19.087163
#>  [57] 15.333112 10.911274 12.030084 11.872385  9.905870 18.023761 13.906598
#>  [64] 11.527080 10.813359  5.867063 16.703253 17.713326 15.329358 10.914761
#>  [71]  6.364880 19.718259 18.815475 18.584733 15.574285 12.231116  9.569255
#>  [78] 19.793150 18.178540 16.689211 17.229496 17.880814 12.932479 10.173034
#>  [85]  8.389409 19.718739 19.212202 15.998045 12.865958 11.942651  6.534807
#>  [92] 14.983661 16.848333 16.610055 17.251148 12.344311 10.142478 11.813930
#>  [99]  5.098407 19.381317
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA 0.60   NA   NA   NA   NA 0.75   NA   NA 0.50   NA   NA   NA   NA
#>  [16] 0.60   NA 0.55 0.70   NA   NA 0.55   NA   NA 0.45   NA   NA   NA   NA 0.55
#>  [31]   NA   NA 0.70   NA 0.30   NA 0.40   NA 0.25 0.45 0.55 0.40   NA   NA 0.60
#>  [46]   NA 0.50   NA 0.35 0.45   NA   NA   NA 0.50   NA   NA   NA   NA   NA   NA
#>  [61] 0.55   NA   NA   NA   NA 0.35   NA   NA   NA   NA 0.40   NA   NA   NA   NA
#>  [76]   NA 0.20   NA   NA   NA   NA   NA   NA   NA 0.35   NA   NA   NA   NA   NA
#>  [91] 0.40   NA   NA   NA   NA   NA   NA   NA 0.25   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 262.8
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.3986721 0.2507847
#>  [2,] 0.3986721 0.2507847
#>  [3,] 0.3986721 0.2507847
#>  [4,] 0.3986721 0.2507847
#>  [5,] 0.3986721 0.2507847
#>  [6,] 0.3986721 0.2507847
#>  [7,] 0.3986721 0.2507847
#>  [8,] 0.3986721 0.2507847
#>  [9,] 0.3986721 0.2507847
#> [10,] 0.3986721 0.2507847
#> [11,] 0.3986721 0.2507847
#> [12,] 0.3986721 0.2507847
#> [13,] 0.3986721 0.2507847
#> [14,] 0.3986721 0.2507847
#> [15,] 0.3986721 0.2507847
#> [16,] 0.3986721 0.2507847
#> [17,] 0.3986721 0.2507847
#> [18,] 0.3986721 0.2507847
#> [19,] 0.3986721 0.2507847
#> [20,] 0.3700838 0.2348021