Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4, 0)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 8 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.7661182 0.3708942
#>  [2,] 0.7486132 0.2900644
#>  [3,] 0.8648048 0.5770543
#>  [4,] 0.4308009 0.3321630
#>  [5,] 0.8650630 0.4853768
#>  [6,] 0.5089963 0.2409458
#>  [7,] 0.5523483 0.3706970
#>  [8,] 0.3548023 0.2389735
#>  [9,] 0.9286125 0.5622455
#> [10,] 0.8287790 0.4271743
#> [11,] 0.8746073 0.7786284
#> [12,] 0.5594999 0.2623045
#> [13,] 0.7663018 0.3450418
#> [14,] 0.9262394 0.4979161
#> [15,] 0.8832824 0.4951422
#> [16,] 0.8848188 0.8047369
#> [17,] 0.8437165 0.5440800
#> [18,] 0.7832452 0.3229276
#> [19,] 0.9342575 0.5025524
#> [20,] 0.8436630 0.4041191
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -292.2544     -292.2544 3.240416e-17
#>  [2,]         0      -359.8632     -359.8632 1.407624e-46
#>  [3,]         0      -279.5900     -279.5900 1.024976e-11
#>  [4,]         0      -259.8820     -259.8820 3.713520e-03
#>  [5,]         0      -294.6035     -294.6035 3.093426e-18
#>  [6,]         0      -274.2314     -274.2314 2.177331e-09
#>  [7,]         0      -262.7240     -262.7240 2.165349e-04
#>  [8,]         0      -254.2901     -254.2901 9.960699e-01
#>  [9,]         0      -287.5844     -287.5844 3.457530e-15
#> [10,]         0      -322.6306     -322.6306 2.081541e-30
#> [11,]         0      -279.0865     -279.0865 1.695702e-11
#> [12,]         0      -278.2150     -278.2150 4.053731e-11
#> [13,]         0      -316.0234     -316.0234 1.541289e-27
#> [14,]         0      -315.6498     -315.6498 2.239282e-27
#> [15,]         0      -308.9050     -308.9050 1.902561e-24
#> [16,]         0      -280.0879     -280.0879 6.229434e-12
#> [17,]         0      -271.6360     -271.6360 2.917940e-08
#> [18,]         0      -366.4665     -366.4665 1.908517e-49
#> [19,]         0      -297.3364     -297.3364 2.011659e-19
#> [20,]         0      -311.2848     -311.2848 1.761213e-25
#> 
#> $statistics
#> $statistics$ess
#>   [1] 15.327437  8.874555 13.406932 10.778494  9.445535 16.604673 13.637797
#>   [8] 11.887638 10.134512  8.288416 11.039786  6.737429  7.871434 10.713844
#>  [15]  8.688497  7.591590  6.349628  2.902345  3.933034  1.626564  1.868233
#>  [22]  3.546099  1.644774  2.089268  2.030629  1.209732  1.016999  1.005773
#>  [29]  1.017837  3.216492  3.940684  6.373303  7.362011  6.609455 11.000082
#>  [36]  7.393836 13.338263 10.433064  8.926647 12.072052  7.254219  6.068412
#>  [43] 13.478692 13.765299 10.826816  9.725025  7.952297 11.714716  9.567417
#>  [50]  6.428430 16.896327 11.043463 11.210814  4.664209 17.554578 15.046522
#>  [57] 12.688314 11.096013 11.386793 11.435905 10.515831  9.182563 16.507828
#>  [64] 13.699295 13.095267  9.578880 13.802212 15.994793 13.491108 12.307365
#>  [71] 10.389607 10.264632  9.486541 18.717634 17.602722 14.112456 10.602136
#>  [78]  9.510799 18.522423 17.694024 18.265273 18.389426 15.060176 12.227395
#>  [85] 11.623121 12.260785 12.388268 10.283130  9.301748 17.634766 10.775593
#>  [92]  8.476083 19.062236 18.206589 17.995503 15.076473 13.656955 14.099944
#>  [99] 10.543502 10.056367
#> 
#> $statistics$acceptance_rate
#>   [1]   NA 0.75   NA   NA 0.50   NA   NA   NA   NA 0.35   NA 0.40 0.65   NA 0.50
#>  [16] 0.30 0.55 0.50 0.30 0.35 0.20 0.25 0.25 0.15 0.25 0.15 0.20 0.10 0.15 0.10
#>  [31] 0.30 0.15 0.20 0.15   NA 0.15   NA   NA 0.30   NA 0.60 0.40   NA   NA   NA
#>  [46] 0.40 0.45   NA 0.35 0.45   NA   NA   NA 0.50   NA   NA   NA   NA   NA   NA
#>  [61]   NA 0.50   NA   NA   NA 0.40   NA   NA   NA   NA   NA   NA 0.55   NA   NA
#>  [76]   NA   NA 0.75   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.60   NA
#>  [91]   NA 0.45   NA   NA   NA   NA   NA   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 299.2
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.3548023 0.2389735
#>  [2,] 0.3548023 0.2389735
#>  [3,] 0.3548023 0.2389735
#>  [4,] 0.3548023 0.2389735
#>  [5,] 0.3548023 0.2389735
#>  [6,] 0.3548023 0.2389735
#>  [7,] 0.3548023 0.2389735
#>  [8,] 0.3548023 0.2389735
#>  [9,] 0.3548023 0.2389735
#> [10,] 0.3548023 0.2389735
#> [11,] 0.3548023 0.2389735
#> [12,] 0.3548023 0.2389735
#> [13,] 0.3548023 0.2389735
#> [14,] 0.3548023 0.2389735
#> [15,] 0.3548023 0.2389735
#> [16,] 0.3548023 0.2389735
#> [17,] 0.3548023 0.2389735
#> [18,] 0.3548023 0.2389735
#> [19,] 0.3548023 0.2389735
#> [20,] 0.3548023 0.2389735