Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars
A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter
A particle_filter object

control
A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)

#> Finished 100 steps in 4 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res

#> $pars
#>            beta     gamma
#>  [1,] 0.4829121 0.4011267
#>  [2,] 0.4241000 0.3639147
#>  [3,] 0.4340332 0.2633166
#>  [4,] 0.9121954 0.6878734
#>  [5,] 0.7407507 0.6535312
#>  [6,] 0.5687379 0.3256722
#>  [7,] 0.8835054 0.6366582
#>  [8,] 0.8920677 0.4197919
#>  [9,] 0.9226416 0.3549579
#> [10,] 0.9226416 0.3549579
#> [11,] 0.6875824 0.3231193
#> [12,] 0.5060184 0.2779617
#> [13,] 0.3479212 0.2335346
#> [14,] 0.3390703 0.1604942
#> [15,] 0.3532544 0.1491114
#> [16,] 0.4248062 0.3095722
#> [17,] 0.5494787 0.4148284
#> [18,] 0.2406211 0.1681895
#> [19,] 0.2451263 0.1196684
#> [20,] 0.2072258 0.1164764
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -271.1556     -271.1556 4.684386e-12
#>  [2,]         0      -266.3466     -266.3466 5.743865e-10
#>  [3,]         0      -260.1522     -260.1522 2.814332e-07
#>  [4,]         0      -280.1432     -280.1432 5.853148e-16
#>  [5,]         0      -278.3232     -278.3232 3.612450e-15
#>  [6,]         0      -264.8012     -264.8012 2.693722e-09
#>  [7,]         0      -277.7147     -277.7147 6.638987e-15
#>  [8,]         0      -318.1731     -318.1731 1.783294e-32
#>  [9,]         0      -398.3550     -398.3550 2.683366e-67
#> [10,]         0      -398.8927     -398.8927 1.567273e-67
#> [11,]         0      -297.7937     -297.7937 1.264416e-23
#> [12,]         0      -267.1708     -267.1708 2.518996e-10
#> [13,]         0      -260.9473     -260.9473 1.270775e-07
#> [14,]         0      -256.3677     -256.3677 1.238688e-05
#> [15,]         0      -271.6563     -271.6563 2.839340e-12
#> [16,]         0      -257.0670     -257.0670 6.155375e-06
#> [17,]         0      -273.6019     -273.6019 4.057475e-13
#> [18,]         0      -254.2264     -254.2264 1.054175e-04
#> [19,]         0      -250.4120     -250.4120 4.780835e-03
#> [20,]         0      -245.0738     -245.0738 9.950948e-01
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.976534 12.152704 10.010423  9.213990 12.997432 12.154671 11.046970
#>   [8] 11.735006 10.998086  9.780050 10.889909  7.769627 10.377537  8.010670
#>  [15] 10.019553  8.083564  6.753084  6.646334  7.626026  6.955333  7.768321
#>  [22] 13.462964  7.247355  8.286881 11.598888  7.903272  8.749200  6.912666
#>  [29]  9.228536  9.557397  9.224916  9.938760 15.267198  8.908731 18.019958
#>  [36] 17.331755 16.627554 16.192270 16.001456 15.493291 14.421699 11.790520
#>  [43] 13.018779 14.370205 14.076723 13.709510 12.884755 13.020823 12.573232
#>  [50]  7.953598 19.066675 19.054808 18.374495 13.886651 15.212792 15.774929
#>  [57] 12.194608  8.975382 19.474475 18.634310 18.686990 17.198632 16.484817
#>  [64] 15.624706 15.361712 11.790984  7.674439 18.783252 19.056695 19.193718
#>  [71] 18.023093 17.508701 16.863898 17.155507 15.697066 13.587367  9.565267
#>  [78] 19.795651 18.763238 18.302957 18.548949 18.606463 16.882174 13.558054
#>  [85] 13.590921 13.875334 14.321481 10.062367  9.275341 18.159602  7.435706
#>  [92] 14.678943 16.700819 14.883198 15.798492 11.245265  9.484621 19.745415
#>  [99] 14.830910 12.813488
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA   NA 0.60   NA   NA   NA   NA   NA 0.70   NA 0.70   NA 0.50   NA
#>  [16] 0.65 0.80 0.75 0.60 0.70 0.80   NA 0.60 0.55   NA 0.50 0.60 0.40 0.55 0.40
#>  [31] 0.55 0.45   NA 0.40   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA
#>  [46]   NA   NA   NA   NA 0.30   NA   NA   NA   NA   NA   NA   NA 0.40   NA   NA
#>  [61]   NA   NA   NA   NA   NA   NA 0.30   NA   NA   NA   NA   NA   NA   NA   NA
#>  [76]   NA 0.50   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA 0.50   NA
#>  [91] 0.35   NA   NA   NA   NA   NA 0.35   NA   NA   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 212.8
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)

#>            beta     gamma
#>  [1,] 0.2072258 0.1164764
#>  [2,] 0.2072258 0.1164764
#>  [3,] 0.2072258 0.1164764
#>  [4,] 0.2072258 0.1164764
#>  [5,] 0.2072258 0.1164764
#>  [6,] 0.2072258 0.1164764
#>  [7,] 0.2072258 0.1164764
#>  [8,] 0.2072258 0.1164764
#>  [9,] 0.2072258 0.1164764
#> [10,] 0.2072258 0.1164764
#> [11,] 0.2072258 0.1164764
#> [12,] 0.2072258 0.1164764
#> [13,] 0.2072258 0.1164764
#> [14,] 0.2072258 0.1164764
#> [15,] 0.2072258 0.1164764
#> [16,] 0.2072258 0.1164764
#> [17,] 0.2072258 0.1164764
#> [18,] 0.2072258 0.1164764
#> [19,] 0.2072258 0.1164764
#> [20,] 0.2072258 0.1164764