Run SMC^2

Run a SMC^2. This is experimental and subject to change. Use at your own risk.

smc2(pars, filter, control)

Arguments

pars

A smc2_parameters object containing information about parameters (ranges, priors, proposal kernel, translation functions for use with the particle filter).

filter

A particle_filter object

control

A smc2_control object to control the behaviour of the algorithm

Value

A smc2_result object, with elements

pars: a matrix of sampled parameters (n_parameter_set long)

probabilities: a matrix of probabilities (log_prior, log_likelihood, log_posterior and weight). The latter is the log posterior normalised over all samples

statistics: interesting or useful statistics about your sample, including the ess (effective sample size, over time), acceptance_rate (where a regeneration step was done, the acceptance rate), n_particles, n_parameter_sets and n_steps (inputs to the simulation). The effort field is a rough calculation of the number of particle-filter runs that this run was worth.

Examples

# We use an example from dust which implements an epidemiological SIR
# (Susceptible-Infected-Recovered) model; see vignette("sir_models")
# for more background, as this example follows from the pMCMC example
# there

# The key tuning here is the number of particles per filter and number
# of parameter sets to consider simultaneously. Ordinarily these would
# be set (much) higher with an increase in computing time
n_particles <- 42
n_parameter_sets <- 20

# Basic epidemiological (Susceptible-Infected-Recovered) model
sir <- dust::dust_example("sir")

# Pre-computed incidence data
incidence <- read.csv(system.file("sir_incidence.csv", package = "mcstate"))

# Annotate the data so that it is suitable for the particle filter to use
dat <- mcstate::particle_filter_data(incidence, "day", 4)

# Subset the output during run
index <- function(info) {
  list(run = 5L)
}

# The comparison function, used to compare simulated data with observe
# data, given the above subset
compare <- function(state, observed, pars) {
  exp_noise <- 1e6
  incidence_modelled <- state[1L, , drop = TRUE]
  incidence_observed <- observed$cases
  lambda <- incidence_modelled +
    rexp(n = length(incidence_modelled), rate = exp_noise)
  dpois(x = incidence_observed, lambda = lambda, log = TRUE)
}

# Finally, construct the particle filter:
filter <- mcstate::particle_filter$new(dat, sir, n_particles, compare,
                                       index = index)

# To control the smc2 we need to specify the parameters to consider
pars <- mcstate::smc2_parameters$new(
  list(
    mcstate::smc2_parameter("beta",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1),
    mcstate::smc2_parameter("gamma",
                            function(n) runif(n, 0, 1),
                            function(x) dunif(x, 0, 1, log = TRUE),
                            min = 0, max = 1)))
control <- mcstate::smc2_control(n_parameter_sets, progress = TRUE)

# Then we run the particle filter
res <- mcstate::smc2(pars, filter, control)
#> Finished 100 steps in 6 secs

# This returns quite a lot of information about the fit, and this will
# change in future versions
res
#> $pars
#>            beta     gamma
#>  [1,] 0.6931267 0.5704946
#>  [2,] 0.2198975 0.1315568
#>  [3,] 0.5236914 0.6567060
#>  [4,] 0.4571365 0.3783204
#>  [5,] 0.8768191 0.7969839
#>  [6,] 0.7342485 0.6364596
#>  [7,] 0.6511265 0.8232231
#>  [8,] 0.8624099 0.8303250
#>  [9,] 0.4804286 0.8635805
#> [10,] 0.5246060 0.9815987
#> [11,] 0.7168326 0.8774747
#> [12,] 0.6936957 0.4764088
#> [13,] 0.9163727 0.9691254
#> [14,] 0.5894765 0.3485129
#> [15,] 0.6620121 0.4862691
#> [16,] 0.6972085 0.4494833
#> [17,] 0.3913429 0.2804310
#> [18,] 0.3913429 0.2804310
#> [19,] 0.6037925 0.9518132
#> [20,] 0.3370872 0.3405058
#> 
#> $probabilities
#>       log_prior log_likelihood log_posterior       weight
#>  [1,]         0      -271.3089     -271.3089 2.313147e-10
#>  [2,]         0      -249.1217     -249.1217 9.999978e-01
#>  [3,]         0      -309.0265     -309.0265 9.630801e-27
#>  [4,]         0      -269.0505     -269.0505 2.213240e-09
#>  [5,]         0      -286.2250     -286.2250 7.695044e-17
#>  [6,]         0      -278.2222     -278.2222 2.300465e-13
#>  [7,]         0      -301.1047     -301.1047 2.654973e-23
#>  [8,]         0      -279.0090     -279.0090 1.047374e-13
#>  [9,]         0      -360.3242     -360.3242 5.074066e-49
#> [10,]         0      -382.4801     -382.4801 1.211037e-58
#> [11,]         0      -297.1445     -297.1445 1.392945e-21
#> [12,]         0      -276.2803     -276.2803 1.603874e-12
#> [13,]         0      -283.5020     -283.5020 1.171735e-15
#> [14,]         0      -271.1447     -271.1447 2.726041e-10
#> [15,]         0      -271.7913     -271.7913 1.427953e-10
#> [16,]         0      -267.9832     -267.9832 6.434852e-09
#> [17,]         0      -262.6220     -262.6220 1.370550e-06
#> [18,]         0      -263.1404     -263.1404 8.161089e-07
#> [19,]         0      -318.5776     -318.5776 6.849685e-31
#> [20,]         0      -285.5084     -285.5084 1.575675e-16
#> 
#> $statistics
#> $statistics$ess
#>   [1] 16.765724 14.133636 11.746830 10.595935  9.055625 16.964882 14.252128
#>   [8] 11.124537  9.620274 16.295578 12.031284 12.039873 10.808515  9.790912
#>  [15] 15.177793 12.300562 10.588963  9.676151 11.085723  9.933443 11.886175
#>  [22]  8.996733  9.348349  7.734801  6.577175 11.263083  8.101607 14.831254
#>  [29]  9.676660  8.163666 13.221561 10.224001  7.452365 17.452512  6.948197
#>  [36]  9.620476  5.398468  8.067991  8.936486  6.999567  5.354560  5.203487
#>  [43] 13.835584 13.867740 10.248207  8.957670  6.832325 11.965144  7.759196
#>  [50]  7.454657 18.797915 11.476423 11.555340  3.140203 19.857757 18.526234
#>  [57] 11.288730  8.219812 19.741176 19.456236 15.381404 12.292145 10.812639
#>  [64]  9.692572 19.363512 15.908146 11.450993 11.857034  9.558422 19.094665
#>  [71] 16.176102 15.033925 14.356996 14.780169 10.858542  9.323412 16.552127
#>  [78] 15.539790 12.617792 10.558029 11.928301 13.001380  5.562736 18.471089
#>  [85] 17.689417 18.597595 18.861023 12.100133 11.573039  8.855808 11.906240
#>  [92]  8.425464 19.219710 19.198449 19.037104 16.119297 15.107413 15.969718
#>  [99]  8.661499 18.912584
#> 
#> $statistics$acceptance_rate
#>   [1]   NA   NA   NA   NA 0.65   NA   NA   NA 0.55   NA   NA   NA   NA 0.45   NA
#>  [16]   NA   NA 0.35   NA 0.40   NA 0.40 0.50 0.55 0.45   NA 0.30   NA 0.45 0.50
#>  [31]   NA   NA 0.35   NA 0.35 0.60 0.35 0.55 0.55 0.30 0.50 0.50   NA   NA   NA
#>  [46] 0.40 0.65   NA 0.40 0.50   NA   NA   NA 0.65   NA   NA   NA 0.50   NA   NA
#>  [61]   NA   NA   NA 0.45   NA   NA   NA   NA 0.55   NA   NA   NA   NA   NA   NA
#>  [76] 0.65   NA   NA   NA   NA   NA   NA 0.55   NA   NA   NA   NA   NA   NA 0.60
#>  [91]   NA 0.45   NA   NA   NA   NA   NA   NA 0.45   NA
#> 
#> $statistics$n_particles
#> [1] 42
#> 
#> $statistics$n_parameter_sets
#> [1] 20
#> 
#> $statistics$n_steps
#> [1] 100
#> 
#> $statistics$effort
#> [1] 312.8
#> 
#> 
#> attr(,"class")
#> [1] "smc2_result"

# Most useful is likely the predict method:
predict(res)
#>            beta     gamma
#>  [1,] 0.2198975 0.1315568
#>  [2,] 0.2198975 0.1315568
#>  [3,] 0.2198975 0.1315568
#>  [4,] 0.2198975 0.1315568
#>  [5,] 0.2198975 0.1315568
#>  [6,] 0.2198975 0.1315568
#>  [7,] 0.2198975 0.1315568
#>  [8,] 0.2198975 0.1315568
#>  [9,] 0.2198975 0.1315568
#> [10,] 0.2198975 0.1315568
#> [11,] 0.2198975 0.1315568
#> [12,] 0.2198975 0.1315568
#> [13,] 0.2198975 0.1315568
#> [14,] 0.2198975 0.1315568
#> [15,] 0.2198975 0.1315568
#> [16,] 0.2198975 0.1315568
#> [17,] 0.2198975 0.1315568
#> [18,] 0.2198975 0.1315568
#> [19,] 0.2198975 0.1315568
#> [20,] 0.2198975 0.1315568